% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{aggregate_encoding}
\alias{aggregate_encoding}
\alias{vl_aggregate_color}
\alias{vl_aggregate_detail}
\alias{vl_aggregate_fill}
\alias{vl_aggregate_fillOpacity}
\alias{vl_aggregate_href}
\alias{vl_aggregate_key}
\alias{vl_aggregate_latitude}
\alias{vl_aggregate_latitude2}
\alias{vl_aggregate_longitude}
\alias{vl_aggregate_longitude2}
\alias{vl_aggregate_opacity}
\alias{vl_aggregate_order}
\alias{vl_aggregate_shape}
\alias{vl_aggregate_size}
\alias{vl_aggregate_stroke}
\alias{vl_aggregate_strokeOpacity}
\alias{vl_aggregate_strokeWidth}
\alias{vl_aggregate_text}
\alias{vl_aggregate_tooltip}
\alias{vl_aggregate_x}
\alias{vl_aggregate_x2}
\alias{vl_aggregate_xError}
\alias{vl_aggregate_xError2}
\alias{vl_aggregate_y}
\alias{vl_aggregate_y2}
\alias{vl_aggregate_yError}
\alias{vl_aggregate_yError2}
\title{Add aggregate transform to encoding}
\usage{
vl_aggregate_color(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_detail(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_fill(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_fillOpacity(spec, op = c("argmax", "argmin", "average",
  "count", "distinct", "max", "mean", "median", "min", "missing", "q1",
  "q3", "ci0", "ci1", "stderr", "stdev", "stdevp", "sum", "valid",
  "values", "variance", "variancep", NA))

vl_aggregate_href(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_key(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_latitude(spec, op = c("argmax", "argmin", "average",
  "count", "distinct", "max", "mean", "median", "min", "missing", "q1",
  "q3", "ci0", "ci1", "stderr", "stdev", "stdevp", "sum", "valid",
  "values", "variance", "variancep", NA))

vl_aggregate_latitude2(spec, op = c("argmax", "argmin", "average",
  "count", "distinct", "max", "mean", "median", "min", "missing", "q1",
  "q3", "ci0", "ci1", "stderr", "stdev", "stdevp", "sum", "valid",
  "values", "variance", "variancep", NA))

vl_aggregate_longitude(spec, op = c("argmax", "argmin", "average",
  "count", "distinct", "max", "mean", "median", "min", "missing", "q1",
  "q3", "ci0", "ci1", "stderr", "stdev", "stdevp", "sum", "valid",
  "values", "variance", "variancep", NA))

vl_aggregate_longitude2(spec, op = c("argmax", "argmin", "average",
  "count", "distinct", "max", "mean", "median", "min", "missing", "q1",
  "q3", "ci0", "ci1", "stderr", "stdev", "stdevp", "sum", "valid",
  "values", "variance", "variancep", NA))

vl_aggregate_opacity(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_order(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_shape(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_size(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_stroke(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_strokeOpacity(spec, op = c("argmax", "argmin", "average",
  "count", "distinct", "max", "mean", "median", "min", "missing", "q1",
  "q3", "ci0", "ci1", "stderr", "stdev", "stdevp", "sum", "valid",
  "values", "variance", "variancep", NA))

vl_aggregate_strokeWidth(spec, op = c("argmax", "argmin", "average",
  "count", "distinct", "max", "mean", "median", "min", "missing", "q1",
  "q3", "ci0", "ci1", "stderr", "stdev", "stdevp", "sum", "valid",
  "values", "variance", "variancep", NA))

vl_aggregate_text(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_tooltip(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_x(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_x2(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_xError(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_xError2(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_y(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_y2(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_yError(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))

vl_aggregate_yError2(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{op}{Aggregation op, one of 'argmax', 'argmin', 'average', 'count', 'distinct', 'max', 'mean', 'median', 'min', 'missing', 'q1', 'q3', 'ci0', 'ci1', 'stderr', 'stdev', 'stdevp', 'sum', 'valid', 'values', 'variance', 'variancep', NA}
}
\value{
A modified spec
}
\description{
Add aggregate parameters to an encoding
}
\seealso{
\code{\link[=vl_aggregate]{vl_aggregate()}} to add a top level aggregate transform
}
