% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\alias{vl_hconcat}
\alias{vl_vconcat}
\alias{vl_concat}
\title{Concatenation functions}
\usage{
vl_hconcat(...)

vl_vconcat(...)

vl_concat(..., columns = 2)
}
\arguments{
\item{...}{specs}

\item{columns}{Number of columns to use}
}
\value{
A composite spec
}
\description{
Concatenation functions
}
