% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\name{knit_print.vegaspec}
\alias{knit_print.vegaspec}
\title{Knit-print method}
\description{
See \code{vegawidget::\link[vegawidget]{knit_print.vegaspec}} for details,
particularly on additional packages that may have to be installed.
}
