% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{sort_encoding}
\alias{sort_encoding}
\alias{vl_sort_color}
\alias{vl_sort_fill}
\alias{vl_sort_fillOpacity}
\alias{vl_sort_opacity}
\alias{vl_sort_order}
\alias{vl_sort_shape}
\alias{vl_sort_size}
\alias{vl_sort_stroke}
\alias{vl_sort_strokeOpacity}
\alias{vl_sort_strokeWidth}
\alias{vl_sort_x}
\alias{vl_sort_y}
\title{#' Add sorting to an encoding}
\usage{
vl_sort_color(spec, value)

vl_sort_fill(spec, value)

vl_sort_fillOpacity(spec, value)

vl_sort_opacity(spec, value)

vl_sort_order(spec, value)

vl_sort_shape(spec, value)

vl_sort_size(spec, value)

vl_sort_stroke(spec, value)

vl_sort_strokeOpacity(spec, value)

vl_sort_strokeWidth(spec, value)

vl_sort_x(spec, value)

vl_sort_y(spec, value)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{value}{One of 'ascending', 'descending', a list with a custom ordering, or NA to specify no sorting}
}
\value{
A modified Vega-Lite Spec
}
\description{
#' Sort an encoding in 'ascending' or 'descending' order, or by given array
}
