% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{sort_encoding_by_encoding}
\alias{sort_encoding_by_encoding}
\alias{vl_sort_color_by_encoding}
\alias{vl_sort_fill_by_encoding}
\alias{vl_sort_fillOpacity_by_encoding}
\alias{vl_sort_opacity_by_encoding}
\alias{vl_sort_order_by_encoding}
\alias{vl_sort_shape_by_encoding}
\alias{vl_sort_size_by_encoding}
\alias{vl_sort_stroke_by_encoding}
\alias{vl_sort_strokeOpacity_by_encoding}
\alias{vl_sort_strokeWidth_by_encoding}
\alias{vl_sort_x_by_encoding}
\alias{vl_sort_y_by_encoding}
\title{Add sort transform by encoding to encoding}
\usage{
vl_sort_color_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_fill_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_fillOpacity_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_opacity_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_order_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_shape_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_size_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_stroke_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_strokeOpacity_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_strokeWidth_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_x_by_encoding(spec, encoding = NULL, order = NULL)

vl_sort_y_by_encoding(spec, encoding = NULL, order = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{encoding}{(\emph{SortByEncoding}) The \href{https://vega.github.io/vega-lite/docs/encoding.html#channels}{encoding channel} to sort by (e.g., \code{"x"}, \code{"y"})}

\item{order}{(\emph{SortByEncoding}) The sort order. One of \code{"ascending"} (default), \code{"descending"}, or \code{null} (no not sort).}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add sort by encoding parameters to an encoding
}
