% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{stack_encoding #' Add stack transform to encoding

Add stack parameters to and encoding}
\alias{stack_encoding #' Add stack transform to encoding

Add stack parameters to and encoding}
\alias{vl_stack_x}
\title{Add stack transform to encoding}
\usage{
vl_stack_x(spec = NULL, stack = c("zero", "center", "normalize", NA))
}
\arguments{
\item{spec}{A vega-lite spec}

\item{stack}{one of 'zero', 'center', 'normalize', NA}

\item{spec}{A vega-lite spec}

\item{stack}{one of 'zero', 'center', 'normalize', NA}
}
\value{
A modified spec

A modified spec
}
\description{
Add stack parameters to and encoding
}
\seealso{
\code{\link[=vl_stack]{vl_stack()}} to add a top level stack transform #' @name stack_encoding
}
