% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{stack_encoding}
\alias{stack_encoding}
\alias{vl_stack_x}
\alias{vl_stack_y}
\title{Add stack transform to encoding}
\usage{
vl_stack_x(spec, stack = c("zero", "center", "normalize", NA))

vl_stack_y(spec, stack = c("zero", "center", "normalize", NA))
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{stack}{One of 'zero', 'center', 'normalize', NA}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add stack parameters to an encoding
}
