% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{vegabrite_operators}
\alias{vegabrite_operators}
\alias{|.vegaspec_vega_lite}
\alias{+.vegaspec_vega_lite}
\alias{&.vegaspec_vega_lite}
\title{Concatenation operators}
\usage{
\method{|}{vegaspec_vega_lite}(e1, e2)

\method{+}{vegaspec_vega_lite}(e1, e2)

\method{&}{vegaspec_vega_lite}(e1, e2)
}
\arguments{
\item{e1}{vegalite spec}

\item{e2}{vegalite spec}
}
\value{
A new vegalite spec
}
\description{
Use \code{+} for layering, \code{|} for horizontal concatenation, \code{&} for vertical concatenation.
}
\seealso{
\code{\link[=vl_layer]{vl_layer()}}, \code{\link[=vl_hconcat]{vl_hconcat()}}, \code{\link[=vl_vconcat]{vl_vconcat()}}
}
