% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$AggregateTransform}
\alias{vl$AggregateTransform}
\title{AggregateTransform}
\arguments{
\item{aggregate}{Array of objects that define fields to aggregate.}

\item{groupby}{The data fields to group by. If not specified, a single group containing all data objects will be used.}
}
\value{
A component of a Vega-Lite spec, corresponding to AggregateTransform definition.
}
\description{
Create object for AggregateTransform
}
