% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BindRadioSelect}
\alias{vl$BindRadioSelect}
\title{BindRadioSelect}
\arguments{
\item{debounce}{number}

\item{element}{Element}

\item{input}{radio, select}

\item{labels}{array}

\item{name}{string}

\item{options}{array}

\item{type}{string}
}
\value{
A component of a Vega-Lite spec, corresponding to BindRadioSelect definition.
}
\description{
Create object for BindRadioSelect
}
