% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BindRange}
\alias{vl$BindRange}
\title{BindRange}
\arguments{
\item{debounce}{}

\item{element}{}

\item{input}{}

\item{max}{}

\item{min}{}

\item{name}{}

\item{step}{}

\item{type}{}
}
\value{
A component of a Vega-Lite spec, corresponding to BindRange definition.
}
\description{
Create object for BindRange
}
