% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BinParams}
\alias{vl$BinParams}
\title{BinParams}
\arguments{
\item{anchor}{(\emph{BinParams}) A value in the binned domain at which to anchor the bins, shifting the bin boundaries if necessary to ensure that a boundary aligns with the anchor value.

\strong{Default value:} the minimum bin extent value}

\item{base}{(\emph{BinParams}) The number base to use for automatic bin determination (default is base 10).

\strong{Default value:} \code{10}}

\item{binned}{(\emph{BinParams}) When set to \code{true}, Vega-Lite treats the input data as already binned.}

\item{divide}{(\emph{BinParams}) Scale factors indicating allowable subdivisions. The default value is [5, 2], which indicates that for base 10 numbers (the default base), the method may consider dividing bin sizes by 5 and/or 2. For example, for an initial step size of 10, the method can check if bin sizes of 2 (= 10/5), 5 (= 10/2), or 1 (= 10/(5*2)) might also satisfy the given constraints.

\strong{Default value:} \verb{[5, 2]}}

\item{extent}{(\emph{BinParams}) A two-element (\verb{[min, max]}) array indicating the range of desired bin values.}

\item{maxbins}{(\emph{BinParams}) Maximum number of bins.

\strong{Default value:} \code{6} for \code{row}, \code{column} and \code{shape} channels; \code{10} for other channels}

\item{minstep}{(\emph{BinParams}) A minimum allowable step size (particularly useful for integer values).}

\item{nice}{(\emph{BinParams}) If true, attempts to make the bin boundaries use human-friendly boundaries, such as multiples of ten.

\strong{Default value:} \code{true}}

\item{step}{(\emph{BinParams}) An exact step size to use between bins.

\strong{Note:} If provided, options such as maxbins will be ignored.}

\item{steps}{(\emph{BinParams}) An array of allowable step sizes to choose from.}
}
\value{
A component of a Vega-Lite spec, corresponding to BinParams definition.
}
\description{
Create object for BinParams
}
