% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BrushConfig}
\alias{vl$BrushConfig}
\title{BrushConfig}
\arguments{
\item{cursor}{The mouse cursor used over the interval mark. Any valid \href{https://developer.mozilla.org/en-US/docs/Web/CSS/cursor#Values}{CSS cursor type} can be used.}

\item{fill}{The fill color of the interval mark. #' #' \strong{Default value:} \code{"#333333"}}

\item{fillOpacity}{The fill opacity of the interval mark (a value between \code{0} and \code{1}). #' #' \strong{Default value:} \code{0.125}}

\item{stroke}{The stroke color of the interval mark. #' #' \strong{Default value:} \code{"#ffffff"}}

\item{strokeDash}{An array of alternating stroke and space lengths, for creating dashed or dotted lines.}

\item{strokeDashOffset}{The offset (in pixels) with which to begin drawing the stroke dash array.}

\item{strokeOpacity}{The stroke opacity of the interval mark (a value between \code{0} and \code{1}).}

\item{strokeWidth}{The stroke width of the interval mark.}
}
\value{
A component of a Vega-Lite spec, corresponding to BrushConfig definition.
}
\description{
Create object for BrushConfig
}
