% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$FilterTransform}
\alias{vl$FilterTransform}
\title{FilterTransform}
\arguments{
\item{filter}{(\emph{FilterTransform}) The \code{filter} property must be one of the predicate definitions:
\enumerate{
\item an \href{https://vega.github.io/vega-lite/docs/types.html#expression}{expression} string,
where \code{datum} can be used to refer to the current data object
\item one of the field predicates: \href{https://vega.github.io/vega-lite/docs/filter.html#equal-predicate}{\code{equal}},
\href{https://vega.github.io/vega-lite/docs/filter.html#lt-predicate}{\code{lt}},
\href{https://vega.github.io/vega-lite/docs/filter.html#lte-predicate}{\code{lte}},
\href{https://vega.github.io/vega-lite/docs/filter.html#gt-predicate}{\code{gt}},
\href{https://vega.github.io/vega-lite/docs/filter.html#gte-predicate}{\code{gte}},
\href{https://vega.github.io/vega-lite/docs/filter.html#range-predicate}{\code{range}},
\href{https://vega.github.io/vega-lite/docs/filter.html#one-of-predicate}{\code{oneOf}},
or \href{https://vega.github.io/vega-lite/docs/filter.html#valid-predicate}{\code{valid}},
\item a \href{https://vega.github.io/vega-lite/docs/filter.html#selection-predicate}{selection predicate}
\item a logical operand that combines (1), (2), or (3).
}}
}
\value{
A component of a Vega-Lite spec, corresponding to FilterTransform definition.
}
\description{
Create object for FilterTransform
}
