% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$FlattenTransform}
\alias{vl$FlattenTransform}
\title{FlattenTransform}
\arguments{
\item{as}{The output field names for extracted array values. #' #' \strong{Default value:} The field name of the corresponding array
field}

\item{flatten}{An array of one or more data fields containing arrays to flatten. If multiple fields are specified, their array values
should have a parallel structure, ideally with the same length. If the lengths of parallel arrays do not match, the
longest array will be used with \code{null} values added for missing entries.}
}
\value{
A component of a Vega-Lite spec, corresponding to FlattenTransform definition.
}
\description{
Create object for FlattenTransform
}
