% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$FoldTransform}
\alias{vl$FoldTransform}
\title{FoldTransform}
\arguments{
\item{as}{The output field names for the key and value properties produced by the fold transform. \strong{Default value:} \verb{["key", "value"]}}

\item{fold}{An array of data fields indicating the properties to fold.}
}
\value{
A component of a Vega-Lite spec, corresponding to FoldTransform definition.
}
\description{
Create object for FoldTransform
}
