% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$Href}
\alias{vl$Href}
\title{Href}
\arguments{
\item{aggregate}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) Aggregation function for the field
(e.g., \code{"mean"}, \code{"sum"}, \code{"median"}, \code{"min"}, \code{"max"}, \code{"count"}).

\strong{Default value:} \code{undefined} (None)

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/aggregate.html}{\code{aggregate}} documentation.}

\item{bin}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) A flag for binning a \code{quantitative} field, \href{https://vega.github.io/vega-lite/docs/bin.html#params}{an object defining binning parameters}, or indicating that the data for \code{x} or \code{y} channel are binned before they are imported into Vega-Lite (\code{"binned"}).
\itemize{
\item If \code{true}, default \href{https://vega.github.io/vega-lite/docs/bin.html}{binning parameters} will be applied.
\item If \code{"binned"}, this indicates that the data for the \code{x} (or \code{y}) channel are already binned. You can map the bin-start field to \code{x} (or \code{y}) and the bin-end field to \code{x2} (or \code{y2}). The scale and axis will be formatted similar to binning in Vega-Lite.  To adjust the axis ticks based on the bin step, you can also set the axis's \href{https://vega.github.io/vega-lite/docs/axis.html#ticks}{\code{tickMinStep}} property.
}

\strong{Default value:} \code{false}

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/bin.html}{\code{bin}} documentation.}

\item{condition}{(\emph{ValueDefWithCondition<MarkPropFieldDef,(string|null)>}) A field definition or one or more value definition(s) with a selection predicate.

(\emph{FieldDefWithCondition<StringFieldDef,string>}) One or more value definition(s) with \href{https://vega.github.io/vega-lite/docs/condition.html}{a selection or a test predicate}.

\strong{Note:} A field definition's \code{condition} property can only contain \href{https://vega.github.io/vega-lite/docs/condition.html#value}{conditional value definitions}
since Vega-Lite only allows at most one encoded field per encoding channel.}

\item{field}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) \strong{Required.} A string defining the name of the field from which to pull a data value
or an object defining iterated values from the \href{https://vega.github.io/vega-lite/docs/repeat.html}{\code{repeat}} operator.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/field.html}{\code{field}} documentation.

\strong{Notes:}
\enumerate{
\item Dots (\code{.}) and brackets (\code{[} and \verb{]}) can be used to access nested objects (e.g., \code{"field": "foo.bar"} and \verb{"field": "foo\['bar'\]"}).
If field names contain dots or brackets but are not nested, you can use \verb{\\\\} to escape dots and brackets (e.g., \code{"a\\\\.b"} and \verb{"a\\\\\[0\\\\\]"}).
See more details about escaping in the \href{https://vega.github.io/vega-lite/docs/field.html}{field documentation}.
\item \code{field} is not required if \code{aggregate} is \code{count}.
}}

\item{format}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) The text formatting pattern for labels of guides (axes, legends, headers) and text marks.
\itemize{
\item If the format type is \code{"number"} (e.g., for quantitative fields), this is D3's \href{https://github.com/d3/d3-format#locale_format}{number format pattern}.
\item If the format type is \code{"time"} (e.g., for temporal fields), this is D3's \href{https://github.com/d3/d3-time-format#locale_format}{time format pattern}.
}

See the \href{https://vega.github.io/vega-lite/docs/format.html}{format documentation} for more examples.

\strong{Default value:}  Derived from \href{https://vega.github.io/vega-lite/docs/config.html#format}{numberFormat} config for number format and from \href{https://vega.github.io/vega-lite/docs/config.html#format}{timeFormat} config for time format.}

\item{formatType}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) The format type for labels (\code{"number"} or \code{"time"}).

\strong{Default value:}
\itemize{
\item \code{"time"} for temporal fields and ordinal and nomimal fields with \code{timeUnit}.
\item \code{"number"} for quantitative fields as well as ordinal and nomimal fields without \code{timeUnit}.
}}

\item{labelExpr}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) \href{https://vega.github.io/vega/docs/expressions/}{Vega expression} for customizing labels text.

\strong{Note:} The label text and value can be assessed via the \code{label} and \code{value} properties of the axis's backing \code{datum} object.}

\item{timeUnit}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) Time unit (e.g., \code{year}, \code{yearmonth}, \code{month}, \code{hours}) for a temporal field.
or \href{https://vega.github.io/vega-lite/docs/type.html#cast}{a temporal field that gets casted as ordinal}.

\strong{Default value:} \code{undefined} (None)

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/timeunit.html}{\code{timeUnit}} documentation.}

\item{title}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) A title for the field. If \code{null}, the title will be removed.

\strong{Default value:}  derived from the field's name and transformation function (\code{aggregate}, \code{bin} and \code{timeUnit}). If the field has an aggregate function, the function is displayed as part of the title (e.g., \code{"Sum of Profit"}). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., \code{"Profit (binned)"}, \code{"Transaction Date (year-month)"}). Otherwise, the title is simply the field name.

\strong{Notes}:
\enumerate{
\item You can customize the default field title format by providing the \href{https://vega.github.io/vega-lite/docs/config.html#top-level-config}{\code{fieldTitle}} property in the \href{https://vega.github.io/vega-lite/docs/config.html}{config} or \href{https://vega.github.io/vega-lite/docs/compile.html#field-title}{\code{fieldTitle} function via the \code{compile} function's options}.
\item If both field definition's \code{title} and axis, header, or legend \code{title} are defined, axis/header/legend title will be used.
}}

\item{type}{(\emph{FieldDefWithCondition<StringFieldDef,string>}) The encoded field's type of measurement (\code{"quantitative"}, \code{"temporal"}, \code{"ordinal"}, or \code{"nominal"}).
It can also be a \code{"geojson"} type for encoding \href{https://vega.github.io/vega-lite/docs/geoshape.html}{'geoshape'}.

\strong{Note:}
\itemize{
\item Data values for a temporal field can be either a date-time string (e.g., \code{"2015-03-07 12:32:17"}, \code{"17:01"}, \code{"2015-03-16"}. \code{"2015"}) or a timestamp number (e.g., \code{1552199579097}).
\item Data \code{type} describes the semantics of the data rather than the primitive data types (number, string, etc.). The same primitive data type can have different types of measurement. For example, numeric data can represent quantitative, ordinal, or nominal data.
\item When using with \href{https://vega.github.io/vega-lite/docs/bin.html}{\code{bin}}, the \code{type} property can be either \code{"quantitative"} (for using a linear bin scale) or \href{https://vega.github.io/vega-lite/docs/type.html#cast-bin}{\code{"ordinal"} (for using an ordinal bin scale)}.
\item When using with \href{https://vega.github.io/vega-lite/docs/timeunit.html}{\code{timeUnit}}, the \code{type} property can be either \code{"temporal"} (for using a temporal scale) or \href{https://vega.github.io/vega-lite/docs/type.html#cast-bin}{\code{"ordinal"} (for using an ordinal scale)}.
\item When using with \href{https://vega.github.io/vega-lite/docs/aggregate.html}{\code{aggregate}}, the \code{type} property refers to the post-aggregation data type. For example, we can calculate count \code{distinct} of a categorical field \code{"cat"} using \verb{\{"aggregate": "distinct", "field": "cat", "type": "quantitative"\}}. The \code{"type"} of the aggregate output is \code{"quantitative"}.
\item Secondary channels (e.g., \code{x2}, \code{y2}, \code{xError}, \code{yError}) do not have \code{type} as they have exactly the same type as their primary channels (e.g., \code{x}, \code{y}).
}

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/type.html}{\code{type}} documentation.}

\item{value}{(\emph{ValueDefWithCondition<MarkPropFieldDef,(string|null)>}) A constant value in visual domain (e.g., \code{"red"} / \code{"#0099ff"} / \href{https://vega.github.io/vega-lite/docs/types.html#gradient}{gradient definition} for color, values between \code{0} to \code{1} for opacity).}
}
\value{
A component of a Vega-Lite spec, corresponding to Href definition.
}
\description{
Create object for Href
}
