% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$IntervalSelectionConfig}
\alias{vl$IntervalSelectionConfig}
\title{IntervalSelectionConfig}
\arguments{
\item{bind}{Establishes a two-way binding between the interval selection and the scales used
within the same view. This allows a user to interactively pan and zoom the view.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/bind.html}{\code{bind}}
documentation.}

\item{clear}{Clears the selection, emptying it of all values. Can be a \href{https://vega.github.io/vega/docs/event-streams/}{Event Stream} or \code{false} to disable.

\strong{Default value:} \code{dblclick}.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/clear.html}{\code{clear}}
documentation.}

\item{empty}{By default, \code{all} data values are considered to lie within an empty selection.
When set to \code{none}, empty selections contain no data values.}

\item{encodings}{An array of encoding channels. The corresponding data field values must match
for a data tuple to fall within the selection.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/project.html}{\code{encodings}}
documentation.}

\item{fields}{An array of field names whose values must match for a data tuple to fall within
the selection.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/project.html}{\code{fields}}
documentation.}

\item{init}{Initialize the selection with a mapping between \href{https://vega.github.io/vega-lite/docs/project.html}{projected channels or field names} and arrays of initial
values.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/init.html}{\code{init}}
documentation.}

\item{mark}{An interval selection also adds a rectangle mark to depict the extents of the
interval. The \code{mark} property can be used to customize the appearance of the
mark.

\strong{See also:}
\href{https://vega.github.io/vega-lite/docs/selection-mark.html}{\code{mark}}
documentation.}

\item{on}{A \href{https://vega.github.io/vega/docs/event-streams/}{Vega event stream}
(object or selector) that triggers the selection. For interval
selections, the event stream must specify a \href{https://vega.github.io/vega/docs/event-streams/#between-filters}{start and end}.}

\item{resolve}{With layered and multi-view displays, a strategy that determines how selections'
data queries are resolved when applied in a filter transform, conditional
encoding rule, or scale domain.

\strong{See also:}
\href{https://vega.github.io/vega-lite/docs/selection-resolve.html}{\code{resolve}}
documentation.}

\item{translate}{When truthy, allows a user to interactively move an interval selection
back-and-forth. Can be \code{true}, \code{false} (to disable panning), or a \href{https://vega.github.io/vega/docs/event-streams/}{Vega event stream definition} which must
include a start and end event to trigger continuous panning.

\strong{Default value:} \code{true}, which corresponds to \verb{[mousedown, window:mouseup] > window:mousemove!} which corresponds to clicks and dragging within an interval
selection to reposition it.

\strong{See also:}
\href{https://vega.github.io/vega-lite/docs/translate.html}{\code{translate}}
documentation.}

\item{zoom}{When truthy, allows a user to interactively resize an interval selection.
Can be \code{true}, \code{false} (to disable zooming), or a \href{https://vega.github.io/vega/docs/event-streams/}{Vega event stream definition}. Currently, only
\code{wheel} events are supported.

\strong{Default value:} \code{true}, which corresponds to \verb{wheel!}.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/zoom.html}{\code{zoom}}
documentation.}
}
\value{
A component of a Vega-Lite spec, corresponding to IntervalSelectionConfig definition.
}
\description{
Create object for IntervalSelectionConfig
}
