% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$JoinAggregateTransform}
\alias{vl$JoinAggregateTransform}
\title{JoinAggregateTransform}
\arguments{
\item{groupby}{The data fields for partitioning the data objects into separate groups. If unspecified, all data points will be in a
single group.}

\item{joinaggregate}{The definition of the fields in the join aggregate, and what calculations to use.}
}
\value{
A component of a Vega-Lite spec, corresponding to JoinAggregateTransform definition.
}
\description{
Create object for JoinAggregateTransform
}
