% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$Legend}
\alias{vl$Legend}
\title{Legend}
\arguments{
\item{aria}{(\emph{Legend})}

\item{clipHeight}{(\emph{Legend})}

\item{columnPadding}{(\emph{Legend})}

\item{columns}{(\emph{Legend})}

\item{cornerRadius}{(\emph{Legend})}

\item{description}{(\emph{Legend})}

\item{direction}{(\emph{Legend}) The direction of the legend, one of \code{"vertical"} or \code{"horizontal"}.

\strong{Default value:} - For top-/bottom-\code{orient}ed legends, \code{"horizontal"} - For left-/right-\code{orient}ed legends, \code{"vertical"} - For top/bottom-left/right-\code{orient}ed legends, \code{"horizontal"} for gradient legends and \code{"vertical"} for symbol legends.}

\item{fillColor}{(\emph{Legend})}

\item{format}{(\emph{Legend}) When used with the default \code{"number"} and \code{"time"} format type, the text formatting pattern for labels of guides (axes, legends, headers) and text marks.
\itemize{
\item If the format type is \code{"number"} (e.g., for quantitative fields), this is D3's \href{https://github.com/d3/d3-format#locale_format}{number format pattern}. - If the format type is \code{"time"} (e.g., for temporal fields), this is D3's \href{https://github.com/d3/d3-time-format#locale_format}{time format pattern}.
}

See the \href{https://vega.github.io/vega-lite/docs/format.html}{format documentation} for more examples.

When used with a \href{https://vega.github.io/vega-lite/docs/config.html#custom-format-type}{custom \code{formatType}}, this value will be passed as \code{format} alongside \code{datum.value} to the registered function.

\strong{Default value:}  Derived from \href{https://vega.github.io/vega-lite/docs/config.html#format}{numberFormat} config for number format and from \href{https://vega.github.io/vega-lite/docs/config.html#format}{timeFormat} config for time format.}

\item{formatType}{(\emph{Legend}) The format type for labels. One of \code{"number"}, \code{"time"}, or a \href{https://vega.github.io/vega-lite/docs/config.html#custom-format-type}{registered custom format type}.

\strong{Default value:} - \code{"time"} for temporal fields and ordinal and nominal fields with \code{timeUnit}. - \code{"number"} for quantitative fields as well as ordinal and nominal fields without \code{timeUnit}.}

\item{gradientLength}{(\emph{Legend})}

\item{gradientOpacity}{(\emph{Legend})}

\item{gradientStrokeColor}{(\emph{Legend})}

\item{gradientStrokeWidth}{(\emph{Legend})}

\item{gradientThickness}{(\emph{Legend})}

\item{gridAlign}{(\emph{Legend})}

\item{labelAlign}{(\emph{Legend})}

\item{labelBaseline}{(\emph{Legend})}

\item{labelColor}{(\emph{Legend})}

\item{labelExpr}{(\emph{Legend}) \href{https://vega.github.io/vega/docs/expressions/}{Vega expression} for customizing labels.

\strong{Note:} The label text and value can be assessed via the \code{label} and \code{value} properties of the legend's backing \code{datum} object.}

\item{labelFont}{(\emph{Legend})}

\item{labelFontSize}{(\emph{Legend})}

\item{labelFontStyle}{(\emph{Legend})}

\item{labelFontWeight}{(\emph{Legend})}

\item{labelLimit}{(\emph{Legend})}

\item{labelOffset}{(\emph{Legend})}

\item{labelOpacity}{(\emph{Legend})}

\item{labelOverlap}{(\emph{Legend})}

\item{labelPadding}{(\emph{Legend})}

\item{labelSeparation}{(\emph{Legend})}

\item{legendX}{(\emph{Legend})}

\item{legendY}{(\emph{Legend})}

\item{offset}{(\emph{Legend})}

\item{orient}{(\emph{Legend}) The orientation of the legend, which determines how the legend is positioned within the scene. One of \code{"left"}, \code{"right"}, \code{"top"}, \code{"bottom"}, \code{"top-left"}, \code{"top-right"}, \code{"bottom-left"}, \code{"bottom-right"}, \code{"none"}.

\strong{Default value:} \code{"right"}}

\item{padding}{(\emph{Legend})}

\item{rowPadding}{(\emph{Legend})}

\item{strokeColor}{(\emph{Legend})}

\item{symbolDash}{(\emph{Legend})}

\item{symbolDashOffset}{(\emph{Legend})}

\item{symbolFillColor}{(\emph{Legend})}

\item{symbolLimit}{(\emph{Legend})}

\item{symbolOffset}{(\emph{Legend})}

\item{symbolOpacity}{(\emph{Legend})}

\item{symbolSize}{(\emph{Legend})}

\item{symbolStrokeColor}{(\emph{Legend})}

\item{symbolStrokeWidth}{(\emph{Legend})}

\item{symbolType}{(\emph{Legend})}

\item{tickCount}{(\emph{Legend})}

\item{tickMinStep}{(\emph{Legend}) The minimum desired step between legend ticks, in terms of scale domain values. For example, a value of \code{1} indicates that ticks should not be less than 1 unit apart. If \code{tickMinStep} is specified, the \code{tickCount} value will be adjusted, if necessary, to enforce the minimum step value.

\strong{Default value}: \code{undefined}}

\item{title}{(\emph{Legend}) A title for the field. If \code{null}, the title will be removed.

\strong{Default value:}  derived from the field's name and transformation function (\code{aggregate}, \code{bin} and \code{timeUnit}). If the field has an aggregate function, the function is displayed as part of the title (e.g., \code{"Sum of Profit"}). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., \code{"Profit (binned)"}, \code{"Transaction Date (year-month)"}). Otherwise, the title is simply the field name.

\strong{Notes}:
\enumerate{
\item You can customize the default field title format by providing the \href{https://vega.github.io/vega-lite/docs/config.html#top-level-config}{\code{fieldTitle}} property in the \href{https://vega.github.io/vega-lite/docs/config.html}{config} or \href{https://vega.github.io/vega-lite/docs/compile.html#field-title}{\code{fieldTitle} function via the \code{compile} function's options}.
\item If both field definition's \code{title} and axis, header, or legend \code{title} are defined, axis/header/legend title will be used.
}}

\item{titleAlign}{(\emph{Legend})}

\item{titleAnchor}{(\emph{Legend})}

\item{titleBaseline}{(\emph{Legend})}

\item{titleColor}{(\emph{Legend})}

\item{titleFont}{(\emph{Legend})}

\item{titleFontSize}{(\emph{Legend})}

\item{titleFontStyle}{(\emph{Legend})}

\item{titleFontWeight}{(\emph{Legend})}

\item{titleLimit}{(\emph{Legend})}

\item{titleLineHeight}{(\emph{Legend})}

\item{titleOpacity}{(\emph{Legend})}

\item{titleOrient}{(\emph{Legend})}

\item{titlePadding}{(\emph{Legend})}

\item{type}{(\emph{Legend}) The type of the legend. Use \code{"symbol"} to create a discrete legend and \code{"gradient"} for a continuous color gradient.

\strong{Default value:} \code{"gradient"} for non-binned quantitative fields and temporal fields; \code{"symbol"} otherwise.}

\item{values}{(\emph{Legend}) Explicitly set the visible legend values.}

\item{zindex}{(\emph{Legend}) A non-negative integer indicating the z-index of the legend. If zindex is 0, legend should be drawn behind all chart elements. To put them in front, use zindex = 1.}
}
\value{
A component of a Vega-Lite spec, corresponding to Legend definition.
}
\description{
Create object for Legend
}
