% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$LegendConfig}
\alias{vl$LegendConfig}
\title{LegendConfig}
\arguments{
\item{aria}{(\emph{LegendConfig})}

\item{clipHeight}{(\emph{LegendConfig})}

\item{columnPadding}{(\emph{LegendConfig})}

\item{columns}{(\emph{LegendConfig})}

\item{cornerRadius}{(\emph{LegendConfig})}

\item{description}{(\emph{LegendConfig})}

\item{direction}{(\emph{LegendConfig}) The direction of the legend, one of \code{"vertical"} or \code{"horizontal"}.

\strong{Default value:} - For top-/bottom-\code{orient}ed legends, \code{"horizontal"} - For left-/right-\code{orient}ed legends, \code{"vertical"} - For top/bottom-left/right-\code{orient}ed legends, \code{"horizontal"} for gradient legends and \code{"vertical"} for symbol legends.}

\item{disable}{(\emph{LegendConfig}) Disable legend by default}

\item{fillColor}{(\emph{LegendConfig})}

\item{gradientDirection}{(\emph{LegendConfig})}

\item{gradientHorizontalMaxLength}{(\emph{LegendConfig}) Max legend length for a horizontal gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{200}}

\item{gradientHorizontalMinLength}{(\emph{LegendConfig}) Min legend length for a horizontal gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{100}}

\item{gradientLabelLimit}{(\emph{LegendConfig})}

\item{gradientLabelOffset}{(\emph{LegendConfig})}

\item{gradientLength}{(\emph{LegendConfig})}

\item{gradientOpacity}{(\emph{LegendConfig})}

\item{gradientStrokeColor}{(\emph{LegendConfig})}

\item{gradientStrokeWidth}{(\emph{LegendConfig})}

\item{gradientThickness}{(\emph{LegendConfig})}

\item{gradientVerticalMaxLength}{(\emph{LegendConfig}) Max legend length for a vertical gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{200}}

\item{gradientVerticalMinLength}{(\emph{LegendConfig}) Min legend length for a vertical gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{100}}

\item{gridAlign}{(\emph{LegendConfig})}

\item{labelAlign}{(\emph{LegendConfig})}

\item{labelBaseline}{(\emph{LegendConfig})}

\item{labelColor}{(\emph{LegendConfig})}

\item{labelFont}{(\emph{LegendConfig})}

\item{labelFontSize}{(\emph{LegendConfig})}

\item{labelFontStyle}{(\emph{LegendConfig})}

\item{labelFontWeight}{(\emph{LegendConfig})}

\item{labelLimit}{(\emph{LegendConfig})}

\item{labelOffset}{(\emph{LegendConfig})}

\item{labelOpacity}{(\emph{LegendConfig})}

\item{labelOverlap}{(\emph{LegendConfig}) The strategy to use for resolving overlap of labels in gradient legends. If \code{false}, no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a strategy of removing every other label is used. If set to \code{"greedy"}, a linear scan of the labels is performed, removing any label that overlaps with the last visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{"greedy"} for \verb{log scales otherwise }true`.}

\item{labelPadding}{(\emph{LegendConfig})}

\item{labelSeparation}{(\emph{LegendConfig})}

\item{layout}{(\emph{LegendConfig})}

\item{legendX}{(\emph{LegendConfig})}

\item{legendY}{(\emph{LegendConfig})}

\item{offset}{(\emph{LegendConfig})}

\item{orient}{(\emph{LegendConfig}) The orientation of the legend, which determines how the legend is positioned within the scene. One of \code{"left"}, \code{"right"}, \code{"top"}, \code{"bottom"}, \code{"top-left"}, \code{"top-right"}, \code{"bottom-left"}, \code{"bottom-right"}, \code{"none"}.

\strong{Default value:} \code{"right"}}

\item{padding}{(\emph{LegendConfig})}

\item{rowPadding}{(\emph{LegendConfig})}

\item{strokeColor}{(\emph{LegendConfig})}

\item{strokeDash}{(\emph{LegendConfig})}

\item{strokeWidth}{(\emph{LegendConfig})}

\item{symbolBaseFillColor}{(\emph{LegendConfig})}

\item{symbolBaseStrokeColor}{(\emph{LegendConfig})}

\item{symbolDash}{(\emph{LegendConfig})}

\item{symbolDashOffset}{(\emph{LegendConfig})}

\item{symbolDirection}{(\emph{LegendConfig})}

\item{symbolFillColor}{(\emph{LegendConfig})}

\item{symbolLimit}{(\emph{LegendConfig})}

\item{symbolOffset}{(\emph{LegendConfig})}

\item{symbolOpacity}{(\emph{LegendConfig})}

\item{symbolSize}{(\emph{LegendConfig})}

\item{symbolStrokeColor}{(\emph{LegendConfig})}

\item{symbolStrokeWidth}{(\emph{LegendConfig})}

\item{symbolType}{(\emph{LegendConfig})}

\item{tickCount}{(\emph{LegendConfig})}

\item{title}{(\emph{LegendConfig}) Set to null to disable title for the axis, legend, or header.}

\item{titleAlign}{(\emph{LegendConfig})}

\item{titleAnchor}{(\emph{LegendConfig})}

\item{titleBaseline}{(\emph{LegendConfig})}

\item{titleColor}{(\emph{LegendConfig})}

\item{titleFont}{(\emph{LegendConfig})}

\item{titleFontSize}{(\emph{LegendConfig})}

\item{titleFontStyle}{(\emph{LegendConfig})}

\item{titleFontWeight}{(\emph{LegendConfig})}

\item{titleLimit}{(\emph{LegendConfig})}

\item{titleLineHeight}{(\emph{LegendConfig})}

\item{titleOpacity}{(\emph{LegendConfig})}

\item{titleOrient}{(\emph{LegendConfig})}

\item{titlePadding}{(\emph{LegendConfig})}

\item{unselectedOpacity}{(\emph{LegendConfig}) The opacity of unselected legend entries.

\strong{Default value:} 0.35.}

\item{zindex}{(\emph{LegendConfig})}
}
\value{
A component of a Vega-Lite spec, corresponding to LegendConfig definition.
}
\description{
Create object for LegendConfig
}
