% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$LoessTransform}
\alias{vl$LoessTransform}
\title{LoessTransform}
\arguments{
\item{as}{The output field names for the smoothed points generated by the loess transform. #' #' \strong{Default value:} The field
names of the input x and y values.}

\item{bandwidth}{A bandwidth parameter in the range \verb{[0, 1]} that determines the amount of smoothing. #' #' \strong{Default value:} \code{0.3}}

\item{groupby}{The data fields to group by. If not specified, a single group containing all data objects will be used.}

\item{loess}{The data field of the dependent variable to smooth.}

\item{on}{The data field of the independent variable to use a predictor.}
}
\value{
A component of a Vega-Lite spec, corresponding to LoessTransform definition.
}
\description{
Create object for LoessTransform
}
