% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$PivotTransform}
\alias{vl$PivotTransform}
\title{PivotTransform}
\arguments{
\item{groupby}{The optional data fields to group by. If not specified, a single group containing all data objects will be used.}

\item{limit}{An optional parameter indicating the maximum number of pivoted fields to generate. The default (\code{0}) applies no limit.
The pivoted \code{pivot} names are sorted in ascending order prior to enforcing the limit. \strong{Default value:} \code{0}}

\item{op}{The aggregation operation to apply to grouped \code{value} field values. \strong{Default value:} \code{sum}}

\item{pivot}{The data field to pivot on. The unique values of this field become new field names in the output stream.}

\item{value}{The data field to populate pivoted fields. The aggregate values of this field become the values of the new pivoted
fields.}
}
\value{
A component of a Vega-Lite spec, corresponding to PivotTransform definition.
}
\description{
Create object for PivotTransform
}
