% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$ProjectionConfig}
\alias{vl$ProjectionConfig}
\title{ProjectionConfig}
\arguments{
\item{center}{The projection's center, a two-element array of longitude and latitude in degrees. #' #' \strong{Default value:} \verb{[0, 0]}}

\item{clipAngle}{The projection's clipping circle radius to the specified angle in degrees. If \code{null}, switches to
\href{http://bl.ocks.org/mbostock/3788999}{antimeridian} cutting rather than small-circle clipping.}

\item{clipExtent}{The projection's viewport clip extent to the specified bounds in pixels. The extent bounds are specified as an array
\verb{[\[x0, y0\], \[x1, y1\]]}, where \code{x0} is the left-side of the viewport, \code{y0} is the top, \code{x1} is the right and \code{y1} is
the bottom. If \code{null}, no viewport clipping is performed.}

\item{coefficient}{}

\item{distance}{}

\item{extent}{}

\item{fit}{}

\item{fraction}{}

\item{lobes}{}

\item{parallel}{}

\item{parallels}{For conic projections, the \href{https://en.wikipedia.org/wiki/Map_projection#Conic}{two standard parallels} that define the
map layout. The default depends on the specific conic projection used.}

\item{pointRadius}{The default radius (in pixels) to use when drawing GeoJSON \code{Point} and \code{MultiPoint} geometries. This parameter sets a
constant default value. To modify the point radius in response to data, see the corresponding parameter of the
GeoPath and GeoShape transforms. #' #' \strong{Default value:} \code{4.5}}

\item{precision}{The threshold for the projection's \href{http://bl.ocks.org/mbostock/3795544}{adaptive resampling} to the specified value in
pixels. This value corresponds to the \href{http://en.wikipedia.org/wiki/Ramer-Douglas-Peucker_algorithm}{Douglas–Peucker distance}. If precision is not
specified, returns the projection's current resampling precision which defaults to \verb{√0.5 ≅ 0.70710…}.}

\item{radius}{}

\item{ratio}{}

\item{reflectX}{}

\item{reflectY}{}

\item{rotate}{The projection's three-axis rotation to the specified angles, which must be a two- or three-element array of numbers
[\code{lambda}, \code{phi}, \code{gamma}] specifying the rotation angles in degrees about each spherical axis. (These correspond to
yaw, pitch and roll.) #' #' \strong{Default value:} \verb{[0, 0, 0]}}

\item{scale}{The projection’s scale (zoom) factor, overriding automatic fitting. The default scale is projection-specific. The scale
factor corresponds linearly to the distance between projected points; however, scale factor values are not
equivalent across projections.}

\item{size}{}

\item{spacing}{}

\item{tilt}{}

\item{translate}{The projection’s translation offset as a two-element array \verb{[tx, ty]}.}

\item{type}{The cartographic projection to use. This value is case-insensitive, for example \code{"albers"} and \code{"Albers"} indicate the
same projection type. You can find all valid projection types \href{https://vega.github.io/vega-lite/docs/projection.html#projection-types}{in the documentation}. #' #' \strong{Default value:}
\code{mercator}}
}
\value{
A component of a Vega-Lite spec, corresponding to ProjectionConfig definition.
}
\description{
Create object for ProjectionConfig
}
