% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$RangeConfig}
\alias{vl$RangeConfig}
\title{RangeConfig}
\arguments{
\item{category}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for categorical data.}

\item{diverging}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for diverging quantitative ramps.}

\item{heatmap}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for quantitative heatmaps.}

\item{ordinal}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for rank-ordered data.}

\item{ramp}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for sequential quantitative ramps.}

\item{symbol}{(\emph{RangeConfig}) Array of \href{https://vega.github.io/vega/docs/marks/symbol/}{symbol} names or paths for the default shape palette.}
}
\value{
A component of a Vega-Lite spec, corresponding to RangeConfig definition.
}
\description{
Create object for RangeConfig
}
