% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$SampleTransform}
\alias{vl$SampleTransform}
\title{SampleTransform}
\arguments{
\item{sample}{The maximum number of data objects to include in the sample. #' #' \strong{Default value:} \code{1000}}
}
\value{
A component of a Vega-Lite spec, corresponding to SampleTransform definition.
}
\description{
Create object for SampleTransform
}
