% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$SelectionConfig}
\alias{vl$SelectionConfig}
\title{SelectionConfig}
\arguments{
\item{interval}{(\emph{SelectionConfig}) The default definition for an \href{https://vega.github.io/vega-lite/docs/selection.html#type}{\code{interval}} selection. All properties and transformations
for an interval selection definition (except \code{type}) may be specified here.

For instance, setting \code{interval} to \code{{"translate": false}} disables the ability to move
interval selections by default.}

\item{multi}{(\emph{SelectionConfig}) The default definition for a \href{https://vega.github.io/vega-lite/docs/selection.html#type}{\code{multi}} selection. All properties and transformations
for a multi selection definition (except \code{type}) may be specified here.

For instance, setting \code{multi} to \code{{"toggle": "event.altKey"}} adds additional values to
multi selections when clicking with the alt-key pressed by default.}

\item{single}{(\emph{SelectionConfig}) The default definition for a \href{https://vega.github.io/vega-lite/docs/selection.html#type}{\code{single}} selection. All properties and transformations
for a single selection definition (except \code{type}) may be specified here.

For instance, setting \code{single} to \code{{"on": "dblclick"}} populates single selections on double-click by default.}
}
\value{
A component of a Vega-Lite spec, corresponding to SelectionConfig definition.
}
\description{
Create object for SelectionConfig
}
