% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$SortField}
\alias{vl$SortField}
\title{SortField}
\arguments{
\item{field}{The name of the field to sort.}

\item{order}{Whether to sort the field in ascending or descending order. One of \code{"ascending"}
(default), \code{"descending"}, or \code{null} (no not sort).}
}
\value{
A component of a Vega-Lite spec, corresponding to SortField definition.
}
\description{
Create object for SortField
}
