% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$StackTransform}
\alias{vl$StackTransform}
\title{StackTransform}
\arguments{
\item{as}{(\emph{StackTransform}) Output field names. This can be either a string or an array of strings with two elements denoting the name for the fields for stack start and stack end respectively.
If a single string(e.g., \code{"val"}) is provided, the end field will be \code{"val_end"}.}

\item{groupby}{(\emph{StackTransform}) The data fields to group by.}

\item{offset}{(\emph{StackTransform}) Mode for stacking marks. One of \code{"zero"} (default), \code{"center"}, or \code{"normalize"}.
The \code{"zero"} offset will stack starting at \code{0}. The \code{"center"} offset will center the stacks. The \code{"normalize"} offset will compute percentage values for each stack point, with output values in the range \code{[0,1]}.

\strong{Default value:} \code{"zero"}}

\item{sort}{(\emph{StackTransform}) Field that determines the order of leaves in the stacked charts.}

\item{stack}{(\emph{StackTransform}) The field which is stacked.}
}
\value{
A component of a Vega-Lite spec, corresponding to StackTransform definition.
}
\description{
Create object for StackTransform
}
