% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$TimeUnitTransform}
\alias{vl$TimeUnitTransform}
\title{TimeUnitTransform}
\arguments{
\item{as}{The output field to write the timeUnit value.}

\item{field}{The data field to apply time unit.}

\item{timeUnit}{The timeUnit.}
}
\value{
A component of a Vega-Lite spec, corresponding to TimeUnitTransform definition.
}
\description{
Create object for TimeUnitTransform
}
