% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$TitleConfig}
\alias{vl$TitleConfig}
\title{TitleConfig}
\arguments{
\item{align}{(\emph{BaseTitleNoValueRefs}) Horizontal text alignment for title text. One of \code{"left"}, \code{"center"}, or \code{"right"}.}

\item{anchor}{(\emph{BaseTitleNoValueRefs})}

\item{angle}{(\emph{BaseTitleNoValueRefs})}

\item{aria}{(\emph{BaseTitleNoValueRefs})}

\item{baseline}{(\emph{BaseTitleNoValueRefs}) Vertical text baseline for title and subtitle text. One of \code{"alphabetic"} (default), \code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, or \code{"line-bottom"}. The \code{"line-top"} and \code{"line-bottom"} values operate similarly to \code{"top"} and \code{"bottom"}, but are calculated relative to the \emph{lineHeight} rather than \emph{fontSize} alone.}

\item{color}{(\emph{BaseTitleNoValueRefs})}

\item{dx}{(\emph{BaseTitleNoValueRefs})}

\item{dy}{(\emph{BaseTitleNoValueRefs})}

\item{font}{(\emph{BaseTitleNoValueRefs})}

\item{fontSize}{(\emph{BaseTitleNoValueRefs})}

\item{fontStyle}{(\emph{BaseTitleNoValueRefs})}

\item{fontWeight}{(\emph{BaseTitleNoValueRefs})}

\item{frame}{(\emph{BaseTitleNoValueRefs})}

\item{limit}{(\emph{BaseTitleNoValueRefs})}

\item{lineHeight}{(\emph{BaseTitleNoValueRefs})}

\item{offset}{(\emph{BaseTitleNoValueRefs})}

\item{orient}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleColor}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFont}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFontSize}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFontStyle}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFontWeight}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleLineHeight}{(\emph{BaseTitleNoValueRefs})}

\item{subtitlePadding}{(\emph{BaseTitleNoValueRefs})}

\item{zindex}{(\emph{BaseTitleNoValueRefs})}
}
\value{
A component of a Vega-Lite spec, corresponding to TitleConfig definition.
}
\description{
Create object for TitleConfig
}
