% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$ViewConfig}
\alias{vl$ViewConfig}
\title{ViewConfig}
\arguments{
\item{clip}{Whether the view should be clipped.}

\item{continuousHeight}{The default height when the plot has a continuous y-field for x or latitude, or has arc marks. #' #' \strong{Default value:}
\code{200}}

\item{continuousWidth}{The default width when the plot has a continuous field for x or longitude, or has arc marks. #' #' \strong{Default value:}
\code{200}}

\item{cornerRadius}{}

\item{cursor}{The mouse cursor used over the view. Any valid \href{https://developer.mozilla.org/en-US/docs/Web/CSS/cursor#Values}{CSS cursor type} can be used.}

\item{discreteHeight}{The default height when the plot has non arc marks and either a discrete y-field or no y-field. The height can be
either a number indicating a fixed height or an object in the form of \code{{step: number}} defining the height per
discrete step. #' #' \strong{Default value:} a step size based on \code{config.view.step}.}

\item{discreteWidth}{The default width when the plot has non-arc marks and either a discrete x-field or no x-field. The width can be either
a number indicating a fixed width or an object in the form of \code{{step: number}} defining the width per discrete step.
#' #' \strong{Default value:} a step size based on \code{config.view.step}.}

\item{fill}{The fill color. #' #' \strong{Default value:} \code{undefined}}

\item{fillOpacity}{}

\item{height}{Default height #' #' \strong{Deprecated:} Since Vega-Lite 4.0. Please use continuousHeight and discreteHeight instead.}

\item{opacity}{The overall opacity (value between [0,1]). #' #' \strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick},
\code{circle}, or \code{square} marks or layered \code{bar} charts and \code{1} otherwise.}

\item{step}{Default step size for x-/y- discrete fields.}

\item{stroke}{The stroke color. #' #' \strong{Default value:} \code{"#ddd"}}

\item{strokeCap}{}

\item{strokeDash}{}

\item{strokeDashOffset}{}

\item{strokeJoin}{}

\item{strokeMiterLimit}{}

\item{strokeOpacity}{}

\item{strokeWidth}{}

\item{width}{Default width #' #' \strong{Deprecated:} Since Vega-Lite 4.0. Please use continuousWidth and discreteWidth instead.}
}
\value{
A component of a Vega-Lite spec, corresponding to ViewConfig definition.
}
\description{
Create object for ViewConfig
}
