% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_config}
\alias{vl_add_config}
\title{vl_add_config}
\usage{
vl_add_config(spec, area = NULL, autosize = NULL, axis = NULL,
  axisBand = NULL, axisBottom = NULL, axisLeft = NULL,
  axisRight = NULL, axisTop = NULL, axisX = NULL, axisY = NULL,
  background = NULL, bar = NULL, boxplot = NULL, circle = NULL,
  concat = NULL, countTitle = NULL, errorband = NULL,
  errorbar = NULL, facet = NULL, fieldTitle = NULL,
  geoshape = NULL, header = NULL, headerColumn = NULL,
  headerFacet = NULL, headerRow = NULL, invalidValues = NULL,
  legend = NULL, line = NULL, mark = NULL, numberFormat = NULL,
  padding = NULL, point = NULL, projection = NULL, range = NULL,
  rect = NULL, `repeat` = NULL, rule = NULL, scale = NULL,
  selection = NULL, square = NULL, stack = NULL, style = NULL,
  text = NULL, tick = NULL, timeFormat = NULL, title = NULL,
  trail = NULL, view = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{area}{Area-Specific Config (type = AreaConfig)}

\item{autosize}{Sets how the visualization size should be determined. If a string, should be one of \code{"pad"}, \code{"fit"} or \code{"none"}.
Object values can additionally specify parameters for content sizing and automatic resizing.
\code{"fit"} is only supported for single and layered views that don't use \code{rangeStep}.

\strong{Default value}: \code{pad} (type = Varies)}

\item{axis}{Axis configuration, which determines default properties for all \code{x} and \code{y} \href{https://vega.github.io/vega-lite/docs/axis.html}{axes}. For a full list of axis configuration options, please see the \href{https://vega.github.io/vega-lite/docs/axis.html#config}{corresponding section of the axis documentation}. (type = AxisConfig)}

\item{axisBand}{Specific axis config for axes with "band" scales. (type = AxisConfig)}

\item{axisBottom}{Specific axis config for x-axis along the bottom edge of the chart. (type = AxisConfig)}

\item{axisLeft}{Specific axis config for y-axis along the left edge of the chart. (type = AxisConfig)}

\item{axisRight}{Specific axis config for y-axis along the right edge of the chart. (type = AxisConfig)}

\item{axisTop}{Specific axis config for x-axis along the top edge of the chart. (type = AxisConfig)}

\item{axisX}{X-axis specific config. (type = AxisConfig)}

\item{axisY}{Y-axis specific config. (type = AxisConfig)}

\item{background}{CSS color property to use as the background of the whole Vega-Lite view

\strong{Default value:} none (transparent) (type = string)}

\item{bar}{Bar-Specific Config (type = BarConfig)}

\item{boxplot}{Box Config (type = BoxPlotConfig)}

\item{circle}{Circle-Specific Config (type = MarkConfig)}

\item{concat}{Default configuration for all concatenation view composition operators (\code{concat}, \code{hconcat}, and \code{vconcat}) (type = CompositionConfig)}

\item{countTitle}{Default axis and legend title for count fields.

\strong{Default value:} \code{'Count of Records}. (type = string)}

\item{errorband}{ErrorBand Config (type = ErrorBandConfig)}

\item{errorbar}{ErrorBar Config (type = ErrorBarConfig)}

\item{facet}{Default configuration for the \code{facet} view composition operator (type = CompositionConfig)}

\item{fieldTitle}{Defines how Vega-Lite generates title for fields.  There are three possible styles:
\itemize{
\item \code{"verbal"} (Default) - displays function in a verbal style (e.g., "Sum of field", "Year-month of date", "field (binned)").
\item \code{"function"} - displays function using parentheses and capitalized texts (e.g., "SUM(field)", "YEARMONTH(date)", "BIN(field)").
\item \code{"plain"} - displays only the field name without functions (e.g., "field", "date", "field"). (type = string)
}}

\item{geoshape}{Geoshape-Specific Config (type = MarkConfig)}

\item{header}{Header configuration, which determines default properties for all \href{https://vega.github.io/vega-lite/docs/header.html}{headers}.

For a full list of header configuration options, please see the \href{https://vega.github.io/vega-lite/docs/header.html#config}{corresponding section of in the header documentation}. (type = HeaderConfig)}

\item{headerColumn}{Header configuration, which determines default properties for column \href{https://vega.github.io/vega-lite/docs/header.html}{headers}.

For a full list of header configuration options, please see the \href{https://vega.github.io/vega-lite/docs/header.html#config}{corresponding section of in the header documentation}. (type = HeaderConfig)}

\item{headerFacet}{Header configuration, which determines default properties for non-row/column facet \href{https://vega.github.io/vega-lite/docs/header.html}{headers}.

For a full list of header configuration options, please see the \href{https://vega.github.io/vega-lite/docs/header.html#config}{corresponding section of in the header documentation}. (type = HeaderConfig)}

\item{headerRow}{Header configuration, which determines default properties for row \href{https://vega.github.io/vega-lite/docs/header.html}{headers}.

For a full list of header configuration options, please see the \href{https://vega.github.io/vega-lite/docs/header.html#config}{corresponding section of in the header documentation}. (type = HeaderConfig)}

\item{invalidValues}{Defines how Vega-Lite should handle invalid values (\code{null} and \code{NaN}).
\itemize{
\item If set to \code{"filter"} (default), all data items with null values will be skipped (for line, trail, and area marks) or filtered (for other marks).
\item If \code{null}, all data items are included. In this case, invalid values will be interpreted as zeroes. (type = string OR null)
}}

\item{legend}{Legend configuration, which determines default properties for all \href{https://vega.github.io/vega-lite/docs/legend.html}{legends}. For a full list of legend configuration options, please see the \href{https://vega.github.io/vega-lite/docs/legend.html#config}{corresponding section of in the legend documentation}. (type = LegendConfig)}

\item{line}{Line-Specific Config (type = LineConfig)}

\item{mark}{Mark Config (type = MarkConfig)}

\item{numberFormat}{D3 Number format for guide labels and text marks. For example "s" for SI units. Use \href{https://github.com/d3/d3-format#locale_format}{D3's number format pattern}. (type = string)}

\item{padding}{The default visualization padding, in pixels, from the edge of the visualization canvas to the data rectangle.  If a number, specifies padding for all sides.
If an object, the value should have the format \code{{"left": 5, "top": 5, "right": 5, "bottom": 5}} to specify padding for each side of the visualization.

\strong{Default value}: \code{5} (type = Padding)}

\item{point}{Point-Specific Config (type = MarkConfig)}

\item{projection}{Projection configuration, which determines default properties for all \href{https://vega.github.io/vega-lite/docs/projection.html}{projections}. For a full list of projection configuration options, please see the \href{https://vega.github.io/vega-lite/docs/projection.html#config}{corresponding section of the projection documentation}. (type = ProjectionConfig)}

\item{range}{An object hash that defines default range arrays or schemes for using with scales.
For a full list of scale range configuration options, please see the \href{https://vega.github.io/vega-lite/docs/scale.html#config}{corresponding section of the scale documentation}. (type = RangeConfig)}

\item{rect}{Rect-Specific Config (type = MarkConfig)}

\item{repeat}{Default configuration for the \code{repeat} view composition operator (type = CompositionConfig)}

\item{rule}{Rule-Specific Config (type = MarkConfig)}

\item{scale}{Scale configuration determines default properties for all \href{https://vega.github.io/vega-lite/docs/scale.html}{scales}. For a full list of scale configuration options, please see the \href{https://vega.github.io/vega-lite/docs/scale.html#config}{corresponding section of the scale documentation}. (type = ScaleConfig)}

\item{selection}{An object hash for defining default properties for each type of selections. (type = SelectionConfig)}

\item{square}{Square-Specific Config (type = MarkConfig)}

\item{stack}{Default stack offset for stackable mark. (type = StackOffset)}

\item{style}{An object hash that defines key-value mappings to determine default properties for marks with a given \href{https://vega.github.io/vega-lite/docs/mark.html#mark-def}{style}.  The keys represent styles names; the values have to be valid \href{https://vega.github.io/vega-lite/docs/mark.html#config}{mark configuration objects}. (type = StyleConfigIndex)}

\item{text}{Text-Specific Config (type = TextConfig)}

\item{tick}{Tick-Specific Config (type = TickConfig)}

\item{timeFormat}{Default time format for raw time values (without time units) in text marks, legend labels and header labels.

\strong{Default value:} \code{"\%b \%d, \%Y"}
\strong{Note:} Axes automatically determine format each label automatically so this config would not affect axes. (type = string)}

\item{title}{Title configuration, which determines default properties for all \href{https://vega.github.io/vega-lite/docs/title.html}{titles}. For a full list of title configuration options, please see the \href{https://vega.github.io/vega-lite/docs/title.html#config}{corresponding section of the title documentation}. (type = TitleConfig)}

\item{trail}{Trail-Specific Config (type = LineConfig)}

\item{view}{Default properties for \href{https://vega.github.io/vega-lite/docs/spec.html#single}{single view plots}. (type = ViewConfig)}
}
\value{
A modified spec
}
\description{
Add top-level config to a vega-lite spec.
}
