% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{vl_add_data_frame}
\alias{vl_add_data_frame}
\title{vl_add_data_frame}
\usage{
vl_add_data_frame(spec, values, ...)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{values}{Data frame with data to add to chart}

\item{...}{Additional arguments to pass to \code{\link[=vl_add_data]{vl_add_data()}}, such as \code{name} or \code{format}}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add data in the format of an R data frame
}
\examples{

vl_chart() \%>\%
  vl_add_data_frame(mtcars) \%>\%
  vl_mark_point() \%>\%
  vl_encode_x("wt") \%>\%
  vl_encode_y("mpg") 
}
\seealso{
\code{\link[=vl_add_data]{vl_add_data()}}
}
