% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_errorbar_config}
\alias{vl_add_errorbar_config}
\title{vl_add_errorbar_config}
\usage{
vl_add_errorbar_config(spec, extent = NULL, rule = NULL,
  ticks = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{extent}{(\emph{ErrorBarConfig}) The extent of the rule. Available options include:
\itemize{
\item \code{"ci"}: Extend the rule to the confidence interval of the mean.
\item \code{"stderr"}: The size of rule are set to the value of standard error, extending from the mean.
\item \code{"stdev"}: The size of rule are set to the value of standard deviation, extending from the mean.
\item \code{"iqr"}: Extend the rule to the q1 and q3.
}

\strong{Default value:} \code{"stderr"}.}

\item{rule}{(\emph{ErrorBarConfig})}

\item{ticks}{(\emph{ErrorBarConfig})}
}
\value{
A modified spec
}
\description{
Add errorbar config (ErrorBarConfig) to a vega-lite spec.
}
