% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_headerRow_config}
\alias{vl_add_headerRow_config}
\title{vl_add_headerRow_config}
\usage{
vl_add_headerRow_config(spec, format = NULL, formatType = NULL,
  labelAlign = NULL, labelAnchor = NULL, labelAngle = NULL,
  labelColor = NULL, labelFont = NULL, labelFontSize = NULL,
  labelLimit = NULL, labelOrient = NULL, labelPadding = NULL,
  shortTimeLabels = NULL, title = NULL, titleAlign = NULL,
  titleAnchor = NULL, titleAngle = NULL, titleBaseline = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontWeight = NULL, titleLimit = NULL, titleOrient = NULL,
  titlePadding = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{format}{(\emph{HeaderConfig}) The text formatting pattern for labels of guides (axes, legends, headers) and text marks.
\itemize{
\item If the format type is \code{"number"} (e.g., for quantitative fields), this is D3's \href{https://github.com/d3/d3-format#locale_format}{number format pattern}.
\item If the format type is \code{"time"} (e.g., for temporal fields), this is D3's \href{https://github.com/d3/d3-time-format#locale_format}{time format pattern}.
}

See the \href{https://vega.github.io/vega-lite/docs/format.html}{format documentation} for more examples.

\strong{Default value:}  Derived from \href{https://vega.github.io/vega-lite/docs/config.html#format}{numberFormat} config for number format and from \href{https://vega.github.io/vega-lite/docs/config.html#format}{timeFormat} config for time format.}

\item{formatType}{(\emph{HeaderConfig}) The format type for labels (\code{"number"} or \code{"time"}).

\strong{Default value:}
\itemize{
\item \code{"time"} for temporal fields and ordinal and nomimal fields with \code{timeUnit}.
\item \code{"number"} for quantitative fields as well as ordinal and nomimal fields without \code{timeUnit}.
}}

\item{labelAlign}{(\emph{HeaderConfig}) Horizontal text alignment of header labels.}

\item{labelAnchor}{(\emph{HeaderConfig}) The anchor position for placing the labels. One of \code{"start"}, \code{"middle"}, or \code{"end"}. For example, with a label orientation of top these anchor positions map to a left-, center-, or right-aligned label.}

\item{labelAngle}{(\emph{HeaderConfig}) The rotation angle of the header labels.

\strong{Default value:} \code{0} for column header, \code{-90} for row header.}

\item{labelColor}{(\emph{HeaderConfig}) The color of the header label, can be in hex color code or regular color name.}

\item{labelFont}{(\emph{HeaderConfig}) The font of the header label.}

\item{labelFontSize}{(\emph{HeaderConfig}) The font size of the header label, in pixels.}

\item{labelLimit}{(\emph{HeaderConfig}) The maximum length of the header label in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

\strong{Default value:} \code{0}, indicating no limit}

\item{labelOrient}{(\emph{HeaderConfig}) The orientation of the header label. One of \code{"top"}, \code{"bottom"}, \code{"left"} or \code{"right"}.}

\item{labelPadding}{(\emph{HeaderConfig}) The padding, in pixel, between facet header's label and the plot.

\strong{Default value:} \code{10}}

\item{shortTimeLabels}{(\emph{HeaderConfig}) Whether month names and weekday names should be abbreviated.

\strong{Default value:}  \code{false}}

\item{title}{(\emph{HeaderConfig}) Set to null to disable title for the axis, legend, or header.}

\item{titleAlign}{(\emph{HeaderConfig}) Horizontal text alignment (to the anchor) of header titles.}

\item{titleAnchor}{(\emph{HeaderConfig}) The anchor position for placing the title. One of \code{"start"}, \code{"middle"}, or \code{"end"}. For example, with an orientation of top these anchor positions map to a left-, center-, or right-aligned title.}

\item{titleAngle}{(\emph{HeaderConfig}) The rotation angle of the header title.

\strong{Default value:} \code{0}.}

\item{titleBaseline}{(\emph{HeaderConfig}) Vertical text baseline for the header title. One of \code{"top"}, \code{"bottom"}, \code{"middle"}.

\strong{Default value:} \code{"middle"}}

\item{titleColor}{(\emph{HeaderConfig}) Color of the header title, can be in hex color code or regular color name.}

\item{titleFont}{(\emph{HeaderConfig}) Font of the header title. (e.g., \code{"Helvetica Neue"}).}

\item{titleFontSize}{(\emph{HeaderConfig}) Font size of the header title.}

\item{titleFontWeight}{(\emph{HeaderConfig}) Font weight of the header title.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{titleLimit}{(\emph{HeaderConfig}) The maximum length of the header title in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

\strong{Default value:} \code{0}, indicating no limit}

\item{titleOrient}{(\emph{HeaderConfig}) The orientation of the header title. One of \code{"top"}, \code{"bottom"}, \code{"left"} or \code{"right"}.}

\item{titlePadding}{(\emph{HeaderConfig}) The padding, in pixel, between facet header's title and the label.

\strong{Default value:} \code{10}}
}
\value{
A modified spec
}
\description{
Add headerRow config (HeaderConfig) to a vega-lite spec.
}
