% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_legend_config}
\alias{vl_add_legend_config}
\title{vl_add_legend_config}
\usage{
vl_add_legend_config(spec, clipHeight = NULL, columnPadding = NULL,
  columns = NULL, cornerRadius = NULL, fillColor = NULL,
  gradientDirection = NULL, gradientHorizontalMaxLength = NULL,
  gradientHorizontalMinLength = NULL, gradientLabelLimit = NULL,
  gradientLabelOffset = NULL, gradientLength = NULL,
  gradientOpacity = NULL, gradientStrokeColor = NULL,
  gradientStrokeWidth = NULL, gradientThickness = NULL,
  gradientVerticalMaxLength = NULL, gradientVerticalMinLength = NULL,
  gridAlign = NULL, labelAlign = NULL, labelBaseline = NULL,
  labelColor = NULL, labelFont = NULL, labelFontSize = NULL,
  labelFontStyle = NULL, labelFontWeight = NULL, labelLimit = NULL,
  labelOffset = NULL, labelOpacity = NULL, labelOverlap = NULL,
  labelPadding = NULL, labelSeparation = NULL, layout = NULL,
  offset = NULL, orient = NULL, padding = NULL, rowPadding = NULL,
  shortTimeLabels = NULL, strokeColor = NULL, strokeDash = NULL,
  strokeWidth = NULL, symbolBaseFillColor = NULL,
  symbolBaseStrokeColor = NULL, symbolDash = NULL,
  symbolDashOffset = NULL, symbolDirection = NULL,
  symbolFillColor = NULL, symbolOffset = NULL, symbolOpacity = NULL,
  symbolSize = NULL, symbolStrokeColor = NULL,
  symbolStrokeWidth = NULL, symbolType = NULL, title = NULL,
  titleAlign = NULL, titleAnchor = NULL, titleBaseline = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontStyle = NULL, titleFontWeight = NULL, titleLimit = NULL,
  titleOpacity = NULL, titleOrient = NULL, titlePadding = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{clipHeight}{The height in pixels to clip symbol legend entries and limit their size. (type = number)}

\item{columnPadding}{The horizontal padding in pixels between symbol legend entries.

\strong{Default value:} \code{10}. (type = number)}

\item{columns}{The number of columns in which to arrange symbol legend entries. A value of \code{0} or lower indicates a single row with one column per entry. (type = number)}

\item{cornerRadius}{Corner radius for the full legend. (type = number)}

\item{fillColor}{Background fill color for the full legend. (type = Color)}

\item{gradientDirection}{The default direction (\code{"horizontal"} or \code{"vertical"}) for gradient legends.

\strong{Default value:} \code{"vertical"}. (type = Orientation)}

\item{gradientHorizontalMaxLength}{Max legend length for a horizontal gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{200} (type = number)}

\item{gradientHorizontalMinLength}{Min legend length for a horizontal gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{100} (type = number)}

\item{gradientLabelLimit}{The maximum allowed length in pixels of color ramp gradient labels. (type = number)}

\item{gradientLabelOffset}{Vertical offset in pixels for color ramp gradient labels.

\strong{Default value:} \code{2}. (type = number)}

\item{gradientLength}{The length in pixels of the primary axis of a color gradient. This value corresponds to the height of a vertical gradient or the width of a horizontal gradient.

\strong{Default value:} \code{200}. (type = number)}

\item{gradientOpacity}{Opacity of the color gradient. (type = number)}

\item{gradientStrokeColor}{The color of the gradient stroke, can be in hex color code or regular color name.

\strong{Default value:} \code{"lightGray"}. (type = Color)}

\item{gradientStrokeWidth}{The width of the gradient stroke, in pixels.

\strong{Default value:} \code{0}. (type = number)}

\item{gradientThickness}{The thickness in pixels of the color gradient. This value corresponds to the width of a vertical gradient or the height of a horizontal gradient.

\strong{Default value:} \code{16}. (type = number)}

\item{gradientVerticalMaxLength}{Max legend length for a vertical gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{200} (type = number)}

\item{gradientVerticalMinLength}{Min legend length for a vertical gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{100} (type = number)}

\item{gridAlign}{The alignment to apply to symbol legends rows and columns. The supported string values are \code{"all"}, \code{"each"} (the default), and \code{none}. For more information, see the \href{https://vega.github.io/vega/docs/layout}{grid layout documentation}.

\strong{Default value:} \code{"each"}. (type = LayoutAlign)}

\item{labelAlign}{The alignment of the legend label, can be left, center, or right. (type = Align)}

\item{labelBaseline}{The position of the baseline of legend label, can be \code{"top"}, \code{"middle"}, \code{"bottom"}, or \code{"alphabetic"}.

\strong{Default value:} \code{"middle"}. (type = TextBaseline)}

\item{labelColor}{The color of the legend label, can be in hex color code or regular color name. (type = Color)}

\item{labelFont}{The font of the legend label. (type = string)}

\item{labelFontSize}{The font size of legend label.

\strong{Default value:} \code{10}. (type = number)}

\item{labelFontStyle}{The font style of legend label. (type = FontStyle)}

\item{labelFontWeight}{The font weight of legend label. (type = FontWeight)}

\item{labelLimit}{Maximum allowed pixel width of axis tick labels.

\strong{Default value:} \code{160}. (type = number)}

\item{labelOffset}{The offset of the legend label. (type = number)}

\item{labelOpacity}{Opacity of labels. (type = number)}

\item{labelOverlap}{The strategy to use for resolving overlap of labels in gradient legends. If \code{false}, no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a strategy of removing every other label is used. If set to \code{"greedy"}, a linear scan of the labels is performed, removing any label that overlaps with the last visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{"greedy"} for \code{log scales otherwise}true`.
\itemize{
\item (type = LabelOverlap)
}}

\item{labelPadding}{Padding in pixels between the legend and legend labels. (type = number)}

\item{labelSeparation}{The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default \code{0}). This property is ignored if \emph{labelOverlap} resolution is not enabled. (type = number)}

\item{layout}{Legend orient group layout parameters. (type = LegendLayout)}

\item{offset}{The offset in pixels by which to displace the legend from the data rectangle and axes.

\strong{Default value:} \code{18}. (type = number)}

\item{orient}{The orientation of the legend, which determines how the legend is positioned within the scene. One of "left", "right", "top-left", "top-right", "bottom-left", "bottom-right", "none".

\strong{Default value:} \code{"right"} (type = LegendOrient)}

\item{padding}{The padding between the border and content of the legend group.

\strong{Default value:} \code{0}. (type = number)}

\item{rowPadding}{The vertical padding in pixels between symbol legend entries.

\strong{Default value:} \code{2}. (type = number)}

\item{shortTimeLabels}{Whether month names and weekday names should be abbreviated.

\strong{Default value:}  \code{false} (type = boolean)}

\item{strokeColor}{Border stroke color for the full legend. (type = Color)}

\item{strokeDash}{Border stroke dash pattern for the full legend. (type = array)}

\item{strokeWidth}{Border stroke width for the full legend. (type = number)}

\item{symbolBaseFillColor}{Default fill color for legend symbols. Only applied if there is no \code{"fill"} scale color encoding for the legend.

\strong{Default value:} \code{"transparent"}. (type = Color)}

\item{symbolBaseStrokeColor}{Default stroke color for legend symbols. Only applied if there is no \code{"fill"} scale color encoding for the legend.

\strong{Default value:} \code{"gray"}. (type = Color)}

\item{symbolDash}{An array of alternating \link{stroke, space} lengths for dashed symbol strokes. (type = array)}

\item{symbolDashOffset}{The pixel offset at which to start drawing with the symbol stroke dash array. (type = number)}

\item{symbolDirection}{The default direction (\code{"horizontal"} or \code{"vertical"}) for symbol legends.

\strong{Default value:} \code{"vertical"}. (type = Orientation)}

\item{symbolFillColor}{The color of the legend symbol, (type = Color)}

\item{symbolOffset}{Horizontal pixel offset for legend symbols.

\strong{Default value:} \code{0}. (type = number)}

\item{symbolOpacity}{Opacity of the legend symbols. (type = number)}

\item{symbolSize}{The size of the legend symbol, in pixels.

\strong{Default value:} \code{100}. (type = number)}

\item{symbolStrokeColor}{Stroke color for legend symbols. (type = Color)}

\item{symbolStrokeWidth}{The width of the symbol's stroke.

\strong{Default value:} \code{1.5}. (type = number)}

\item{symbolType}{Default shape type (such as "circle") for legend symbols.
Can be one of ``"circle"\code{,}"square"`, `"cross"`, `"diamond"`, `"triangle-up"`, `"triangle-down"`, `"triangle-right"`, or `"triangle-left"`.
\itemize{
\item In addition to a set of built-in shapes, custom shapes can be defined using \href{https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths}{SVG path strings}.
\item 
\item \strong{Default value:} \code{"circle"}.
\item (type = SymbolShape)
}}

\item{title}{Set to null to disable title for the axis, legend, or header. (type = null)}

\item{titleAlign}{Horizontal text alignment for legend titles.

\strong{Default value:} \code{"left"}. (type = Align)}

\item{titleAnchor}{Text anchor position for placing legend titles. (type = TitleAnchor)}

\item{titleBaseline}{Vertical text baseline for legend titles.

\strong{Default value:} \code{"top"}. (type = TextBaseline)}

\item{titleColor}{The color of the legend title, can be in hex color code or regular color name. (type = Color)}

\item{titleFont}{The font of the legend title. (type = string)}

\item{titleFontSize}{The font size of the legend title. (type = number)}

\item{titleFontStyle}{The font style of the legend title. (type = FontStyle)}

\item{titleFontWeight}{The font weight of the legend title.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}). (type = FontWeight)}

\item{titleLimit}{Maximum allowed pixel width of axis titles.

\strong{Default value:} \code{180}. (type = number)}

\item{titleOpacity}{Opacity of the legend title. (type = number)}

\item{titleOrient}{Orientation of the legend title. (type = Orient)}

\item{titlePadding}{The padding, in pixels, between title and legend.

\strong{Default value:} \code{5}. (type = number)}
}
\value{
A modified spec
}
\description{
Add legend config (LegendConfig) to a vega-lite spec.
}
