% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_range_config}
\alias{vl_add_range_config}
\title{vl_add_range_config}
\usage{
vl_add_range_config(spec, category = NULL, diverging = NULL,
  heatmap = NULL, ordinal = NULL, ramp = NULL, symbol = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{category}{Default range for \emph{nominal} (categorical) fields. (type = Varies)}

\item{diverging}{Default range for diverging \emph{quantitative} fields. (type = Varies)}

\item{heatmap}{Default range for \emph{quantitative} heatmaps. (type = Varies)}

\item{ordinal}{Default range for \emph{ordinal} fields. (type = Varies)}

\item{ramp}{Default range for \emph{quantitative} and \emph{temporal} fields. (type = Varies)}

\item{symbol}{Default range palette for the \code{shape} channel. (type = array)}
}
\value{
A modified spec
}
\description{
Add range config (RangeConfig) to a vega-lite spec.
}
