% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_repeat_config}
\alias{vl_add_repeat_config}
\title{vl_add_repeat_config}
\usage{
vl_add_repeat_config(spec, columns = NULL, spacing = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{columns}{The number of columns to include in the view composition layout.

\strong{Default value}: \code{undefined} -- An infinite number of columns (a single row) will be assumed. This is equivalent to
\code{hconcat} (for \code{concat}) and to using the \code{column} channel (for \code{facet} and \code{repeat}).

\strong{Note}:
\enumerate{
\item This property is only for:
}
\itemize{
\item the general (wrappable) \code{concat} operator (not \code{hconcat}/\code{vconcat})
\item the \code{facet} and \code{repeat} operator with one field/repetition definition (without row/column nesting)
}
\enumerate{
\item Setting the \code{columns} to \code{1} is equivalent to \code{vconcat} (for \code{concat}) and to using the \code{row} channel (for \code{facet} and \code{repeat}). (type = number)
}}

\item{spacing}{The default spacing in pixels between composed sub-views.

\strong{Default value}: \code{20} (type = number)}
}
\value{
A modified spec
}
\description{
Add repeat config (CompositionConfig) to a vega-lite spec.
}
