% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_scale_config}
\alias{vl_add_scale_config}
\title{vl_add_scale_config}
\usage{
vl_add_scale_config(spec, bandPaddingInner = NULL,
  bandPaddingOuter = NULL, barBandPaddingInner = NULL,
  barBandPaddingOuter = NULL, clamp = NULL, continuousPadding = NULL,
  maxBandSize = NULL, maxFontSize = NULL, maxOpacity = NULL,
  maxSize = NULL, maxStrokeWidth = NULL, minBandSize = NULL,
  minFontSize = NULL, minOpacity = NULL, minSize = NULL,
  minStrokeWidth = NULL, pointPadding = NULL, quantileCount = NULL,
  quantizeCount = NULL, rangeStep = NULL,
  rectBandPaddingInner = NULL, rectBandPaddingOuter = NULL,
  round = NULL, textXRangeStep = NULL, useUnaggregatedDomain = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{bandPaddingInner}{Default inner padding for \code{x} and \code{y} band-ordinal scales.

\strong{Default value:}
\itemize{
\item \code{barBandPaddingInner} for bar marks (\code{0.1} by default)
\item \code{rectBandPaddingInner} for rect and other marks (\code{0} by default) (type = number)
}}

\item{bandPaddingOuter}{Default outer padding for \code{x} and \code{y} band-ordinal scales.

If not specified, by default, band scale's paddingOuter is paddingInner/2. (type = number)}

\item{barBandPaddingInner}{Default inner padding for \code{x} and \code{y} band-ordinal scales of \code{"bar"} marks.

\strong{Default value:} \code{0.1} (type = number)}

\item{barBandPaddingOuter}{Default outer padding for \code{x} and \code{y} band-ordinal scales of \code{"bar"} marks.
If not specified, by default, band scale's paddingOuter is paddingInner/2. (type = number)}

\item{clamp}{If true, values that exceed the data domain are clamped to either the minimum or maximum range value (type = boolean)}

\item{continuousPadding}{Default padding for continuous scales.

\strong{Default:} \code{5} for continuous x-scale of a vertical bar and continuous y-scale of a horizontal bar.; \code{0} otherwise. (type = number)}

\item{maxBandSize}{The default max value for mapping quantitative fields to bar's size/bandSize.

If undefined (default), we will use the scale's \code{rangeStep} - 1. (type = number)}

\item{maxFontSize}{The default max value for mapping quantitative fields to text's size/fontSize.

\strong{Default value:} \code{40} (type = number)}

\item{maxOpacity}{Default max opacity for mapping a field to opacity.

\strong{Default value:} \code{0.8} (type = number)}

\item{maxSize}{Default max value for point size scale. (type = number)}

\item{maxStrokeWidth}{Default max strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks.

\strong{Default value:} \code{4} (type = number)}

\item{minBandSize}{The default min value for mapping quantitative fields to bar and tick's size/bandSize scale with zero=false.

\strong{Default value:} \code{2} (type = number)}

\item{minFontSize}{The default min value for mapping quantitative fields to tick's size/fontSize scale with zero=false

\strong{Default value:} \code{8} (type = number)}

\item{minOpacity}{Default minimum opacity for mapping a field to opacity.

\strong{Default value:} \code{0.3} (type = number)}

\item{minSize}{Default minimum value for point size scale with zero=false.

\strong{Default value:} \code{9} (type = number)}

\item{minStrokeWidth}{Default minimum strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks with zero=false.

\strong{Default value:} \code{1} (type = number)}

\item{pointPadding}{Default outer padding for \code{x} and \code{y} point-ordinal scales.

\strong{Default value:} \code{0.5} (type = number)}

\item{quantileCount}{Default range cardinality for \href{https://vega.github.io/vega-lite/docs/scale.html#quantile}{quantile} scale.

\strong{Default value:} \code{4} (type = number)}

\item{quantizeCount}{Default range cardinality for \href{https://vega.github.io/vega-lite/docs/scale.html#quantize}{quantize} scale.

\strong{Default value:} \code{4} (type = number)}

\item{rangeStep}{Default range step for band and point scales of (1) the \code{y} channel
and (2) the \code{x} channel when the mark is not \code{text}.

\strong{Default value:} \code{20} (type = number OR null)}

\item{rectBandPaddingInner}{Default inner padding for \code{x} and \code{y} band-ordinal scales of \code{"rect"} marks.

\strong{Default value:} \code{0} (type = number)}

\item{rectBandPaddingOuter}{Default outer padding for \code{x} and \code{y} band-ordinal scales of \code{"rect"} marks.
If not specified, by default, band scale's paddingOuter is paddingInner/2. (type = number)}

\item{round}{If true, rounds numeric output values to integers.
This can be helpful for snapping to the pixel grid.
(Only available for \code{x}, \code{y}, and \code{size} scales.) (type = boolean)}

\item{textXRangeStep}{Default range step for \code{x} band and point scales of text marks.

\strong{Default value:} \code{90} (type = number)}

\item{useUnaggregatedDomain}{Use the source data range before aggregation as scale domain instead of aggregated data for aggregate axis.

This is equivalent to setting \code{domain} to \code{"unaggregate"} for aggregated \emph{quantitative} fields by default.

This property only works with aggregate functions that produce values within the raw data domain (\code{"mean"}, \code{"average"}, \code{"median"}, \code{"q1"}, \code{"q3"}, \code{"min"}, \code{"max"}). For other aggregations that produce values outside of the raw data domain (e.g. \code{"count"}, \code{"sum"}), this property is ignored.

\strong{Default value:} \code{false} (type = boolean)}
}
\value{
A modified spec
}
\description{
Add scale config (ScaleConfig) to a vega-lite spec.
}
