% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_selection_config}
\alias{vl_add_selection_config}
\title{vl_add_selection_config}
\usage{
vl_add_selection_config(spec, interval = NULL, multi = NULL,
  single = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{interval}{The default definition for an \href{https://vega.github.io/vega-lite/docs/selection.html#type}{interval} selection. All properties and transformations
for an interval selection definition (except \code{type}) may be specified here.

For instance, setting \code{interval} to \code{{"translate": false}} disables the ability to move
interval selections by default. (type = IntervalSelectionConfig)}

\item{multi}{The default definition for a \href{https://vega.github.io/vega-lite/docs/selection.html#type}{multi} selection. All properties and transformations
for a multi selection definition (except \code{type}) may be specified here.

For instance, setting \code{multi} to \code{{"toggle": "event.altKey"}} adds additional values to
multi selections when clicking with the alt-key pressed by default. (type = MultiSelectionConfig)}

\item{single}{The default definition for a \href{https://vega.github.io/vega-lite/docs/selection.html#type}{single} selection. All properties and transformations
for a single selection definition (except \code{type}) may be specified here.

For instance, setting \code{single} to \code{{"on": "dblclick"}} populates single selections on double-click by default. (type = SingleSelectionConfig)}
}
\value{
A modified spec
}
\description{
Add selection config (SelectionConfig) to a vega-lite spec.
}
