% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_selection_parameter}
\alias{vl_add_selection_parameter}
\title{vl_add_selection_parameter}
\usage{
vl_add_selection_parameter(
  spec,
  .object = NULL,
  name = NULL,
  select = NULL,
  bind = NULL,
  value = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: SelectionParameter}

\item{name}{Required. A unique name for the selection parameter. Selection names should be
valid JavaScript identifiers: they should contain only alphanumeric characters
(or "$", or "_") and may not start with a digit. Reserved keywords that may not
be used as parameter names are "datum", "event", "item", and "parent".}

\item{select}{Determines the default event processing and data query for the selection.
Vega-Lite currently supports two selection types:
\itemize{
\item \code{"point"} -- to select multiple discrete data values; the first value is
selected on \code{click} and additional values toggled on shift-click.
\item \code{"interval"} -- to select a continuous range of data values on \code{drag}.
}}

\item{bind}{When set, a selection is populated by input elements (also
known as dynamic query widgets) or by interacting with the
corresponding legend. Direct manipulation interaction is
disabled by default; to re-enable it, set the selection's
\href{https://vega.github.io/vega-lite/docs/selection.html#common-selection-properties}{\code{on}}
property.

Legend bindings are restricted to selections that only specify a single field or
encoding.

Query widget binding takes the form of Vega's \href{https://vega.github.io/vega/docs/signals/#bind}{input element binding definition} or can be a mapping
between projected field/encodings and binding definitions.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/bind.html}{\code{bind}}
documentation.}

\item{value}{Initialize the selection with a mapping between \href{https://vega.github.io/vega-lite/docs/selection.html#project}{projected channels or field names} and initial
values.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/value.html}{\code{init}}
documentation.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add a paramter to a spec
}
