% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_single_selection}
\alias{vl_add_single_selection}
\title{vl_add_single_selection}
\usage{
vl_add_single_selection(spec, selection_name, bind = NULL,
  clear = NULL, empty = NULL, encodings = NULL, fields = NULL,
  init = NULL, nearest = NULL, on = NULL, resolve = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{selection_name}{Name of selection}

\item{bind}{Establish a two-way binding between a single selection and input elements
(also known as dynamic query widgets). A binding takes the form of
Vega's \href{https://vega.github.io/vega/docs/signals/#bind}{input element binding definition}
or can be a mapping between projected field/encodings and binding definitions.

See the \href{https://vega.github.io/vega-lite/docs/bind.html}{bind transform} documentation for more information. (type = Varies)}

\item{clear}{Clears the selection, emptying it of all values. Can be an
\href{https://vega.github.io/vega/docs/event-streams/}{EventStream} or \code{false} to disable.

\strong{Default value:} \code{dblclick}.

See the \href{https://vega.github.io/vega-lite/docs/clear.html}{clear} documentation for more information. (type = Varies)}

\item{empty}{By default, \code{all} data values are considered to lie within an empty selection.
When set to \code{none}, empty selections contain no data values. (type = string)}

\item{encodings}{An array of encoding channels. The corresponding data field values
must match for a data tuple to fall within the selection. (type = array)}

\item{fields}{An array of field names whose values must match for a data tuple to
fall within the selection. (type = array)}

\item{init}{Initialize the selection with a mapping between \href{https://vega.github.io/vega-lite/docs/project.html}{projected channels or field names} and initial values. (type = SelectionInitMapping)}

\item{nearest}{When true, an invisible voronoi diagram is computed to accelerate discrete
selection. The data value \emph{nearest} the mouse cursor is added to the selection.

See the \href{https://vega.github.io/vega-lite/docs/nearest.html}{nearest transform} documentation for more information. (type = boolean)}

\item{on}{A \href{https://vega.github.io/vega/docs/event-streams/}{Vega event stream} (object or selector) that triggers the selection.
For interval selections, the event stream must specify a \href{https://vega.github.io/vega/docs/event-streams/#between-filters}{start and end}. (type = EventStream)}

\item{resolve}{With layered and multi-view displays, a strategy that determines how
selections' data queries are resolved when applied in a filter transform,
conditional encoding rule, or scale domain. (type = SelectionResolution)}
}
\value{
A modified spec
}
\description{
Add SingleSelection to a vega-lite spec.
}
