% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_title_config}
\alias{vl_add_title_config}
\title{vl_add_title_config}
\usage{
vl_add_title_config(spec, align = NULL, anchor = NULL, angle = NULL,
  baseline = NULL, color = NULL, dx = NULL, dy = NULL,
  font = NULL, fontSize = NULL, fontStyle = NULL,
  fontWeight = NULL, frame = NULL, limit = NULL, offset = NULL,
  orient = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{align}{(type = Align)}

\item{anchor}{The anchor position for placing the title. One of \code{"start"}, \code{"middle"}, or \code{"end"}. For example, with an orientation of top these anchor positions map to a left-, center-, or right-aligned title. (type = TitleAnchor)}

\item{angle}{Angle in degrees of title text. (type = number)}

\item{baseline}{Vertical text baseline for title text. One of \code{"top"}, \code{"middle"}, \code{"bottom"}, or \code{"alphabetic"}. (type = TextBaseline)}

\item{color}{Text color for title text. (type = Color)}

\item{dx}{Delta offset for title text x-coordinate. (type = number)}

\item{dy}{Delta offset for title text y-coordinate. (type = number)}

\item{font}{Font name for title text. (type = string)}

\item{fontSize}{Font size in pixels for title text.

\strong{Default value:} \code{10}. (type = number)}

\item{fontStyle}{Font style for title text. (type = FontStyle)}

\item{fontWeight}{Font weight for title text.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}). (type = FontWeight)}

\item{frame}{The reference frame for the anchor position, one of \code{"bounds"} (to anchor relative to the full bounding box) or \code{"group"} (to anchor relative to the group width or height). (type = TitleFrame)}

\item{limit}{The maximum allowed length in pixels of legend labels. (type = number)}

\item{offset}{The orthogonal offset in pixels by which to displace the title from its position along the edge of the chart. (type = number)}

\item{orient}{Default title orientation (\code{"top"}, \code{"bottom"}, \code{"left"}, or \code{"right"}) (type = TitleOrient)}
}
\value{
A modified spec
}
\description{
Add title config (TitleConfig) to a vega-lite spec.
}
