% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_variable_parameter}
\alias{vl_add_variable_parameter}
\title{vl_add_variable_parameter}
\usage{
vl_add_variable_parameter(
  spec,
  .object = NULL,
  name = NULL,
  bind = NULL,
  expr = NULL,
  value = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: VariableParameter}

\item{name}{A unique name for the variable parameter. Parameter names should be valid
JavaScript identifiers: they should contain only alphanumeric characters (or
"$", or "_") and may not start with a digit. Reserved keywords that may not be
used as parameter names are "datum", "event", "item", and "parent".}

\item{bind}{Binds the parameter to an external input element such as a slider, selection
list or radio button group.}

\item{expr}{An expression for the value of the parameter. This expression may include other
parameters, in which case the parameter will automatically update in response to
upstream parameter changes.}

\item{value}{The \href{http://vega.github.io/vega-lite/docs/value.html}{initial value} of the
parameter.

\strong{Default value:} \code{undefined}}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add a paramter to a spec
}
