% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_view_config}
\alias{vl_add_view_config}
\title{vl_add_view_config}
\usage{
vl_add_view_config(spec, clip = NULL, cornerRadius = NULL,
  fill = NULL, fillOpacity = NULL, height = NULL, opacity = NULL,
  stroke = NULL, strokeCap = NULL, strokeDash = NULL,
  strokeDashOffset = NULL, strokeJoin = NULL,
  strokeMiterLimit = NULL, strokeOpacity = NULL, strokeWidth = NULL,
  width = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{clip}{Whether the view should be clipped. (type = boolean)}

\item{cornerRadius}{The radius in pixels of rounded rectangle corners.

\strong{Default value:} \code{0} (type = number)}

\item{fill}{The fill color.

\strong{Default value:} \code{undefined} (type = string OR null)}

\item{fillOpacity}{The fill opacity (value between \link{0,1}).

\strong{Default value:} \code{1} (type = number)}

\item{height}{The default height of the single plot or each plot in a trellis plot when the visualization has a continuous (non-ordinal) y-scale with \code{rangeStep} = \code{null}.

\strong{Default value:} \code{200} (type = number)}

\item{opacity}{The overall opacity (value between \link{0,1}).

\strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick}, \code{circle}, or \code{square} marks or layered \code{bar} charts and \code{1} otherwise. (type = number)}

\item{stroke}{The stroke color.

\strong{Default value:} \code{"#ddd"} (type = string OR null)}

\item{strokeCap}{The stroke cap for line ending style. One of \code{"butt"}, \code{"round"}, or \code{"square"}.

\strong{Default value:} \code{"square"} (type = StrokeCap)}

\item{strokeDash}{An array of alternating stroke, space lengths for creating dashed or dotted lines. (type = array)}

\item{strokeDashOffset}{The offset (in pixels) into which to begin drawing with the stroke dash array. (type = number)}

\item{strokeJoin}{The stroke line join method. One of \code{"miter"}, \code{"round"} or \code{"bevel"}.

\strong{Default value:} \code{"miter"} (type = StrokeJoin)}

\item{strokeMiterLimit}{The miter limit at which to bevel a line join. (type = number)}

\item{strokeOpacity}{The stroke opacity (value between \link{0,1}).

\strong{Default value:} \code{1} (type = number)}

\item{strokeWidth}{The stroke width, in pixels. (type = number)}

\item{width}{The default width of the single plot or each plot in a trellis plot when the visualization has a continuous (non-ordinal) x-scale or ordinal x-scale with \code{rangeStep} = \code{null}.

\strong{Default value:} \code{200} (type = number)}
}
\value{
A modified spec
}
\description{
Add view config (ViewConfig) to a vega-lite spec.
}
