% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_aggregate_shape}
\alias{vl_aggregate_shape}
\title{vl_aggregate_shape}
\usage{
vl_aggregate_shape(spec, op = c("argmax", "argmin", "average", "count",
  "distinct", "max", "mean", "median", "min", "missing", "q1", "q3", "ci0",
  "ci1", "stderr", "stdev", "stdevp", "sum", "valid", "values", "variance",
  "variancep", NA))
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{op}{Aggregation op, one of 'argmax', 'argmin', 'average', 'count', 'distinct', 'max', 'mean', 'median', 'min', 'missing', 'q1', 'q3', 'ci0', 'ci1', 'stderr', 'stdev', 'stdevp', 'sum', 'valid', 'values', 'variance', 'variancep', NA}
}
\value{
A modified spec
}
\description{
Add aggregate to encoding for shape to a vega-lite spec.
}
\seealso{
\code{\link[=vl_aggregate]{vl_aggregate()}} to add a top level aggregate transform
}
