% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_AggregateTransform}
\alias{vl_AggregateTransform}
\title{vl_AggregateTransform}
\usage{
vl_AggregateTransform(aggregate = NULL, groupby = NULL)
}
\arguments{
\item{aggregate}{Array of objects that define fields to aggregate. (type = array)}

\item{groupby}{The data fields to group by. If not specified, a single group containing all data objects will be used. (type = array)}
}
\value{
A modified spec
}
\description{
Create spec for AggregateTransform.
}
