% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_axis_y}
\alias{vl_axis_y}
\title{vl_axis_y}
\usage{
vl_axis_y(spec, bandPosition = NULL, domain = NULL,
  domainColor = NULL, domainDash = NULL, domainDashOffset = NULL,
  domainOpacity = NULL, domainWidth = NULL, format = NULL,
  formatType = NULL, grid = NULL, gridColor = NULL,
  gridDash = NULL, gridDashOffset = NULL, gridOpacity = NULL,
  gridWidth = NULL, labelAlign = NULL, labelAngle = NULL,
  labelBaseline = NULL, labelBound = NULL, labelColor = NULL,
  labelFlush = NULL, labelFlushOffset = NULL, labelFont = NULL,
  labelFontSize = NULL, labelFontStyle = NULL,
  labelFontWeight = NULL, labelLimit = NULL, labelOpacity = NULL,
  labelOverlap = NULL, labelPadding = NULL, labelSeparation = NULL,
  labels = NULL, maxExtent = NULL, minExtent = NULL, offset = NULL,
  orient = NULL, position = NULL, tickColor = NULL,
  tickCount = NULL, tickDash = NULL, tickDashOffset = NULL,
  tickExtra = NULL, tickMinStep = NULL, tickOffset = NULL,
  tickOpacity = NULL, tickRound = NULL, tickSize = NULL,
  tickWidth = NULL, ticks = NULL, title = NULL, titleAlign = NULL,
  titleAnchor = NULL, titleAngle = NULL, titleBaseline = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontStyle = NULL, titleFontWeight = NULL, titleLimit = NULL,
  titleOpacity = NULL, titlePadding = NULL, titleX = NULL,
  titleY = NULL, values = NULL, zindex = NULL, remove = FALSE)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{bandPosition}{An interpolation fraction indicating where, for \code{band} scales, axis ticks should be positioned. A value of \code{0} places ticks at the left edge of their bands. A value of \code{0.5} places ticks in the middle of their bands.

\strong{Default value:} \code{0.5} (type = number)}

\item{domain}{A boolean flag indicating if the domain (the axis baseline) should be included as part of the axis.

\strong{Default value:} \code{true} (type = boolean)}

\item{domainColor}{Color of axis domain line.

\strong{Default value:} \code{"gray"}. (type = Color)}

\item{domainDash}{An array of alternating \link{stroke, space} lengths for dashed domain lines. (type = array)}

\item{domainDashOffset}{The pixel offset at which to start drawing with the domain dash array. (type = number)}

\item{domainOpacity}{Opacity of the axis domain line. (type = number)}

\item{domainWidth}{Stroke width of axis domain line

\strong{Default value:} \code{1} (type = number)}

\item{format}{The text formatting pattern for labels of guides (axes, legends, headers) and text marks.
\itemize{
\item If the format type is \code{"number"} (e.g., for quantitative fields), this is D3's \href{https://github.com/d3/d3-format#locale_format}{number format pattern}.
\item If the format type is \code{"time"} (e.g., for temporal fields), this is D3's \href{https://github.com/d3/d3-time-format#locale_format}{time format pattern}.
}

See the \href{https://vega.github.io/vega-lite/docs/format.html}{format documentation} for more examples.

\strong{Default value:}  Derived from \href{https://vega.github.io/vega-lite/docs/config.html#format}{numberFormat} config for number format and from \href{https://vega.github.io/vega-lite/docs/config.html#format}{timeFormat} config for time format. (type = string)}

\item{formatType}{The format type for labels (\code{"number"} or \code{"time"}).

\strong{Default value:}
\itemize{
\item \code{"time"} for temporal fields and ordinal and nomimal fields with \code{timeUnit}.
\item \code{"number"} for quantitative fields as well as ordinal and nomimal fields without \code{timeUnit}. (type = string)
}}

\item{grid}{A boolean flag indicating if grid lines should be included as part of the axis

\strong{Default value:} \code{true} for \href{https://vega.github.io/vega-lite/docs/scale.html#continuous}{continuous scales} that are not binned; otherwise, \code{false}. (type = boolean)}

\item{gridColor}{Color of gridlines.

\strong{Default value:} \code{"lightGray"}. (type = Color)}

\item{gridDash}{An array of alternating \link{stroke, space} lengths for dashed grid lines. (type = array)}

\item{gridDashOffset}{The pixel offset at which to start drawing with the grid dash array. (type = number)}

\item{gridOpacity}{The stroke opacity of grid (value between \link{0,1})

\strong{Default value:} \code{1} (type = number)}

\item{gridWidth}{The grid width, in pixels.

\strong{Default value:} \code{1} (type = number)}

\item{labelAlign}{Horizontal text alignment of axis tick labels, overriding the default setting for the current axis orientation. (type = Align)}

\item{labelAngle}{The rotation angle of the axis labels.

\strong{Default value:} \code{-90} for nominal and ordinal fields; \code{0} otherwise. (type = number)}

\item{labelBaseline}{Vertical text baseline of axis tick labels, overriding the default setting for the current axis orientation. Can be \code{"top"}, \code{"middle"}, \code{"bottom"}, or \code{"alphabetic"}. (type = TextBaseline)}

\item{labelBound}{Indicates if labels should be hidden if they exceed the axis range. If \code{false} (the default) no bounds overlap analysis is performed. If \code{true}, labels will be hidden if they exceed the axis range by more than 1 pixel. If this property is a number, it specifies the pixel tolerance: the maximum amount by which a label bounding box may exceed the axis range.

\strong{Default value:} \code{false}. (type = number OR boolean)}

\item{labelColor}{The color of the tick label, can be in hex color code or regular color name. (type = Color)}

\item{labelFlush}{Indicates if the first and last axis labels should be aligned flush with the scale range. Flush alignment for a horizontal axis will left-align the first label and right-align the last label. For vertical axes, bottom and top text baselines are applied instead. If this property is a number, it also indicates the number of pixels by which to offset the first and last labels; for example, a value of 2 will flush-align the first and last labels and also push them 2 pixels outward from the center of the axis. The additional adjustment can sometimes help the labels better visually group with corresponding axis ticks.

\strong{Default value:} \code{true} for axis of a continuous x-scale. Otherwise, \code{false}. (type = boolean OR number)}

\item{labelFlushOffset}{Indicates the number of pixels by which to offset flush-adjusted labels. For example, a value of \code{2} will push flush-adjusted labels 2 pixels outward from the center of the axis. Offsets can help the labels better visually group with corresponding axis ticks.

\strong{Default value:} \code{0}. (type = number)}

\item{labelFont}{The font of the tick label. (type = string)}

\item{labelFontSize}{The font size of the label, in pixels. (type = number)}

\item{labelFontStyle}{Font style of the title. (type = FontStyle)}

\item{labelFontWeight}{Font weight of axis tick labels. (type = FontWeight)}

\item{labelLimit}{Maximum allowed pixel width of axis tick labels.

\strong{Default value:} \code{180} (type = number)}

\item{labelOpacity}{The opacity of the labels. (type = number)}

\item{labelOverlap}{The strategy to use for resolving overlap of axis labels. If \code{false} (the default), no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a strategy of removing every other label is used (this works well for standard linear axes). If set to \code{"greedy"}, a linear scan of the labels is performed, removing any labels that overlaps with the last visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{true} for non-nominal fields with non-log scales; \code{"greedy"} for log scales; otherwise \code{false}. (type = LabelOverlap)}

\item{labelPadding}{The padding, in pixels, between axis and text labels.

\strong{Default value:} \code{2} (type = number)}

\item{labelSeparation}{The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default \code{0}). This property is ignored if \emph{labelOverlap} resolution is not enabled. (type = number)}

\item{labels}{A boolean flag indicating if labels should be included as part of the axis.

\strong{Default value:} \code{true}. (type = boolean)}

\item{maxExtent}{The maximum extent in pixels that axis ticks and labels should use. This determines a maximum offset value for axis titles.

\strong{Default value:} \code{undefined}. (type = number)}

\item{minExtent}{The minimum extent in pixels that axis ticks and labels should use. This determines a minimum offset value for axis titles.

\strong{Default value:} \code{30} for y-axis; \code{undefined} for x-axis. (type = number)}

\item{offset}{The offset, in pixels, by which to displace the axis from the edge of the enclosing group or data rectangle.

\strong{Default value:} derived from the \href{https://vega.github.io/vega-lite/docs/config.html#facet-scale-config}{axis config}'s \code{offset} (\code{0} by default) (type = number)}

\item{orient}{The orientation of the axis. One of \code{"top"}, \code{"bottom"}, \code{"left"} or \code{"right"}. The orientation can be used to further specialize the axis type (e.g., a y-axis oriented towards the right edge of the chart).

\strong{Default value:} \code{"bottom"} for x-axes and \code{"left"} for y-axes. (type = AxisOrient)}

\item{position}{The anchor position of the axis in pixels. For x-axes with top or bottom orientation, this sets the axis group x coordinate. For y-axes with left or right orientation, this sets the axis group y coordinate.

\strong{Default value}: \code{0} (type = number)}

\item{tickColor}{The color of the axis's tick.

\strong{Default value:} \code{"gray"} (type = Color)}

\item{tickCount}{A desired number of ticks, for axes visualizing quantitative scales. The resulting number may be different so that values are "nice" (multiples of 2, 5, 10) and lie within the underlying scale's range. (type = number)}

\item{tickDash}{An array of alternating \link{stroke, space} lengths for dashed tick mark lines. (type = array)}

\item{tickDashOffset}{The pixel offset at which to start drawing with the tick mark dash array. (type = number)}

\item{tickExtra}{Boolean flag indicating if an extra axis tick should be added for the initial position of the axis. This flag is useful for styling axes for \code{band} scales such that ticks are placed on band boundaries rather in the middle of a band. Use in conjunction with \code{"bandPosition": 1} and an axis \code{"padding"} value of \code{0}. (type = boolean)}

\item{tickMinStep}{The minimum desired step between axis ticks, in terms of scale domain values. For example, a value of \code{1} indicates that ticks should not be less than 1 unit apart. If \code{tickMinStep} is specified, the \code{tickCount} value will be adjusted, if necessary, to enforce the minimum step value.

\strong{Default value}: \code{undefined} (type = number)}

\item{tickOffset}{Position offset in pixels to apply to ticks, labels, and gridlines. (type = number)}

\item{tickOpacity}{Opacity of the ticks. (type = number)}

\item{tickRound}{Boolean flag indicating if pixel position values should be rounded to the nearest integer.

\strong{Default value:} \code{true} (type = boolean)}

\item{tickSize}{The size in pixels of axis ticks.

\strong{Default value:} \code{5} (type = number)}

\item{tickWidth}{The width, in pixels, of ticks.

\strong{Default value:} \code{1} (type = number)}

\item{ticks}{Boolean value that determines whether the axis should include ticks.

\strong{Default value:} \code{true} (type = boolean)}

\item{title}{A title for the field. If \code{null}, the title will be removed.

\strong{Default value:}  derived from the field's name and transformation function (\code{aggregate}, \code{bin} and \code{timeUnit}).  If the field has an aggregate function, the function is displayed as part of the title (e.g., \code{"Sum of Profit"}). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., \code{"Profit (binned)"}, \code{"Transaction Date (year-month)"}).  Otherwise, the title is simply the field name.

\strong{Notes}:
\enumerate{
\item You can customize the default field title format by providing the \href{https://vega.github.io/vega-lite/docs/config.html#top-level-config}{fieldTitle} property in the \href{https://vega.github.io/vega-lite/docs/config.html}{config} or \href{https://vega.github.io/vega-lite/docs/compile.html#field-title}{fieldTitle function via the compile function's options}.
\item If both field definition's \code{title} and axis, header, or legend \code{title} are defined, axis/header/legend title will be used. (type = string OR null)
}}

\item{titleAlign}{Horizontal text alignment of axis titles. (type = Align)}

\item{titleAnchor}{Text anchor position for placing axis titles. (type = TitleAnchor)}

\item{titleAngle}{Angle in degrees of axis titles. (type = number)}

\item{titleBaseline}{Vertical text baseline for axis titles. (type = TextBaseline)}

\item{titleColor}{Color of the title, can be in hex color code or regular color name. (type = Color)}

\item{titleFont}{Font of the title. (e.g., \code{"Helvetica Neue"}). (type = string)}

\item{titleFontSize}{Font size of the title. (type = number)}

\item{titleFontStyle}{Font style of the title. (type = FontStyle)}

\item{titleFontWeight}{Font weight of the title.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}). (type = FontWeight)}

\item{titleLimit}{Maximum allowed pixel width of axis titles. (type = number)}

\item{titleOpacity}{Opacity of the axis title. (type = number)}

\item{titlePadding}{The padding, in pixels, between title and axis. (type = number)}

\item{titleX}{X-coordinate of the axis title relative to the axis group. (type = number)}

\item{titleY}{Y-coordinate of the axis title relative to the axis group. (type = number)}

\item{values}{Explicitly set the visible axis tick values. (type = Varies)}

\item{zindex}{A non-positive integer indicating z-index of the axis.
If zindex is 0, axes should be drawn behind all chart elements.
To put them in front, use \code{"zindex = 1"}.

\strong{Default value:} \code{1} (in front of the marks) for actual axis and \code{0} (behind the marks) for grids. (type = number)}

\item{remove}{Remove the axis?}
}
\value{
A modified spec
}
\description{
Add axis to encoding for y in a vega-lite spec.
}
\seealso{
\code{\link[=vl_y]{vl_y()}}
}
