% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bin}
\alias{vl_bin}
\title{vl_bin}
\usage{
vl_bin(spec, as = NULL, bin = NULL, field = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{as}{The output fields at which to write the start and end bin values. (type = Varies)}

\item{bin}{An object indicating bin properties, or simply \code{true} for using default bin parameters. (type = Varies)}

\item{field}{The data field to bin. (type = string)}
}
\value{
A modified spec
}
\description{
Add BinTransform to a vega-lite spec.
}
