% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_direct_input}
\alias{vl_bind_direct_input}
\title{vl_bind_direct_input}
\usage{
vl_bind_direct_input(
  spec,
  parameter_name,
  projection_name = NULL,
  debounce = NULL,
  element = NULL,
  event = NULL,
  .object = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{parameter_name}{Name of parameter to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{debounce}{If defined, delays event handling until the specified milliseconds have elapsed
since the last event was fired.}

\item{element}{An input element that exposes a \emph{value} property and supports the
\href{https://developer.mozilla.org/en-US/docs/Web/API/EventTarget}{EventTarget}
interface, or a CSS selector string to such an element. When the element updates
and dispatches an event, the \emph{value} property will be used as the new, bound
signal value. When the signal updates independent of the element, the \emph{value}
property will be set to the signal value and a new event will be dispatched on
the element.}

\item{event}{The event (default \code{"input"}) to listen for to track changes on the external
element.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than \code{spec}. Objects can be of type: BindDirect}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add direct binding to a vega-lite spec.
}
