% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_range_input}
\alias{vl_bind_range_input}
\title{vl_bind_range_input}
\usage{
vl_bind_range_input(spec, selection_name, projection_name = NULL,
  debounce = NULL, element = NULL, max = NULL, min = NULL,
  name = NULL, step = NULL, type = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{selection_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{debounce}{(\emph{BindRange})}

\item{element}{(\emph{BindRange})}

\item{max}{(\emph{BindRange})}

\item{min}{(\emph{BindRange})}

\item{name}{(\emph{BindRange})}

\item{step}{(\emph{BindRange})}

\item{type}{(\emph{BindRange})}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add range binding to a vega-lite spec.
}
