% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_BinTransform}
\alias{vl_BinTransform}
\title{vl_BinTransform}
\usage{
vl_BinTransform(as = NULL, bin = NULL, field = NULL)
}
\arguments{
\item{as}{The output fields at which to write the start and end bin values. (type = Varies)}

\item{bin}{An object indicating bin properties, or simply \code{true} for using default bin parameters. (type = Varies)}

\item{field}{The data field to bin. (type = string)}
}
\value{
A modified spec
}
\description{
Create spec for BinTransform.
}
