% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{vl_chart}
\alias{vl_chart}
\title{vl_chart}
\usage{
vl_chart(description = NULL)
}
\arguments{
\item{description}{Description field for chart}
}
\value{
A vega-lite spec, as an S3 object of class vegaspec using \code{\link[vegawidget:as_vegaspec]{vegawidget::as_vegaspec()}}
}
\description{
vl_chart
}
