% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_axisBand}
\alias{vl_config_axisBand}
\title{vl_config_axisBand}
\usage{
vl_config_axisBand(
  spec,
  .object = NULL,
  bandPosition = NULL,
  domain = NULL,
  domainColor = NULL,
  domainDash = NULL,
  domainDashOffset = NULL,
  domainOpacity = NULL,
  domainWidth = NULL,
  grid = NULL,
  gridColor = NULL,
  gridDash = NULL,
  gridDashOffset = NULL,
  gridOpacity = NULL,
  gridWidth = NULL,
  labelAlign = NULL,
  labelAngle = NULL,
  labelBaseline = NULL,
  labelBound = NULL,
  labelColor = NULL,
  labelFlush = NULL,
  labelFlushOffset = NULL,
  labelFont = NULL,
  labelFontSize = NULL,
  labelFontStyle = NULL,
  labelFontWeight = NULL,
  labelLimit = NULL,
  labelOpacity = NULL,
  labelOverlap = NULL,
  labelPadding = NULL,
  labelSeparation = NULL,
  labels = NULL,
  maxExtent = NULL,
  minExtent = NULL,
  orient = NULL,
  tickBand = NULL,
  tickColor = NULL,
  tickDash = NULL,
  tickDashOffset = NULL,
  tickExtra = NULL,
  tickOffset = NULL,
  tickOpacity = NULL,
  tickRound = NULL,
  tickSize = NULL,
  tickWidth = NULL,
  ticks = NULL,
  title = NULL,
  titleAlign = NULL,
  titleAnchor = NULL,
  titleAngle = NULL,
  titleBaseline = NULL,
  titleColor = NULL,
  titleFont = NULL,
  titleFontSize = NULL,
  titleFontStyle = NULL,
  titleFontWeight = NULL,
  titleLimit = NULL,
  titleLineHeight = NULL,
  titleOpacity = NULL,
  titlePadding = NULL,
  titleX = NULL,
  titleY = NULL,
  translate = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: AxisConfig}

\item{bandPosition}{(\emph{AxisConfig}) An interpolation fraction indicating where, for \code{band} scales, axis ticks should be positioned. A value of \code{0} places ticks at the left edge of their bands. A value of \code{0.5} places ticks in the middle of their bands.

\strong{Default value:} \code{0.5}}

\item{domain}{(\emph{AxisConfig}) A boolean flag indicating if the domain (the axis baseline) should be included as part of the axis.

\strong{Default value:} \code{true}}

\item{domainColor}{(\emph{AxisConfig}) Color of axis domain line.

\strong{Default value:} \code{"gray"}.}

\item{domainDash}{(\emph{AxisConfig}) An array of alternating [stroke, space] lengths for dashed domain lines.}

\item{domainDashOffset}{(\emph{AxisConfig}) The pixel offset at which to start drawing with the domain dash array.}

\item{domainOpacity}{(\emph{AxisConfig}) Opacity of the axis domain line.}

\item{domainWidth}{(\emph{AxisConfig}) Stroke width of axis domain line

\strong{Default value:} \code{1}}

\item{grid}{(\emph{AxisConfig}) A boolean flag indicating if grid lines should be included as part of the axis

\strong{Default value:} \code{true} for \href{https://vega.github.io/vega-lite/docs/scale.html#continuous}{continuous scales} that are not binned; otherwise, \code{false}.}

\item{gridColor}{(\emph{AxisConfig})}

\item{gridDash}{(\emph{AxisConfig})}

\item{gridDashOffset}{(\emph{AxisConfig})}

\item{gridOpacity}{(\emph{AxisConfig})}

\item{gridWidth}{(\emph{AxisConfig})}

\item{labelAlign}{(\emph{AxisConfig})}

\item{labelAngle}{(\emph{AxisConfig}) The rotation angle of the axis labels.

\strong{Default value:} \code{-90} for nominal and ordinal fields; \code{0} otherwise.}

\item{labelBaseline}{(\emph{AxisConfig})}

\item{labelBound}{(\emph{AxisConfig}) Indicates if labels should be hidden if they exceed the axis range. If \code{false} (the default) no bounds overlap analysis is performed. If \code{true}, labels will be hidden if they exceed the axis range by more than 1 pixel. If this property is a number, it specifies the pixel tolerance: the maximum amount by which a label bounding box may exceed the axis range.

\strong{Default value:} \code{false}.}

\item{labelColor}{(\emph{AxisConfig})}

\item{labelFlush}{(\emph{AxisConfig}) Indicates if the first and last axis labels should be aligned flush with the scale range. Flush alignment for a horizontal axis will left-align the first label and right-align the last label. For vertical axes, bottom and top text baselines are applied instead. If this property is a number, it also indicates the number of pixels by which to offset the first and last labels; for example, a value of 2 will flush-align the first and last labels and also push them 2 pixels outward from the center of the axis. The additional adjustment can sometimes help the labels better visually group with corresponding axis ticks.

\strong{Default value:} \code{true} for axis of a continuous x-scale. Otherwise, \code{false}.}

\item{labelFlushOffset}{(\emph{AxisConfig}) Indicates the number of pixels by which to offset flush-adjusted labels. For example, a value of \code{2} will push flush-adjusted labels 2 pixels outward from the center of the axis. Offsets can help the labels better visually group with corresponding axis ticks.

\strong{Default value:} \code{0}.}

\item{labelFont}{(\emph{AxisConfig})}

\item{labelFontSize}{(\emph{AxisConfig})}

\item{labelFontStyle}{(\emph{AxisConfig})}

\item{labelFontWeight}{(\emph{AxisConfig})}

\item{labelLimit}{(\emph{AxisConfig}) Maximum allowed pixel width of axis tick labels.

\strong{Default value:} \code{180}}

\item{labelOpacity}{(\emph{AxisConfig})}

\item{labelOverlap}{(\emph{AxisConfig}) The strategy to use for resolving overlap of axis labels. If \code{false} (the default), no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a strategy of removing every other label is used (this works well for standard linear axes). If set to \code{"greedy"}, a linear scan of the labels is performed, removing any labels that overlaps with the last visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{true} for non-nominal fields with non-log scales; \code{"greedy"} for log scales; otherwise \code{false}.}

\item{labelPadding}{(\emph{AxisConfig}) The padding, in pixels, between axis and text labels.

\strong{Default value:} \code{2}}

\item{labelSeparation}{(\emph{AxisConfig}) The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default \code{0}). This property is ignored if \emph{labelOverlap} resolution is not enabled.}

\item{labels}{(\emph{AxisConfig}) A boolean flag indicating if labels should be included as part of the axis.

\strong{Default value:} \code{true}.}

\item{maxExtent}{(\emph{AxisConfig}) The maximum extent in pixels that axis ticks and labels should use. This determines a maximum offset value for axis titles.

\strong{Default value:} \code{undefined}.}

\item{minExtent}{(\emph{AxisConfig}) The minimum extent in pixels that axis ticks and labels should use. This determines a minimum offset value for axis titles.

\strong{Default value:} \code{30} for y-axis; \code{undefined} for x-axis.}

\item{orient}{(\emph{AxisConfig}) The orientation of the axis. One of \code{"top"}, \code{"bottom"}, \code{"left"} or \code{"right"}. The orientation can be used to further specialize the axis type (e.g., a y-axis oriented towards the right edge of the chart).

\strong{Default value:} \code{"bottom"} for x-axes and \code{"left"} for y-axes.}

\item{tickBand}{(\emph{AxisConfig}) For band scales, indicates if ticks and grid lines should be placed at the center of a band (default) or at the band extents to indicate intervals.}

\item{tickColor}{(\emph{AxisConfig})}

\item{tickDash}{(\emph{AxisConfig})}

\item{tickDashOffset}{(\emph{AxisConfig})}

\item{tickExtra}{(\emph{AxisConfig}) Boolean flag indicating if an extra axis tick should be added for the initial position of the axis. This flag is useful for styling axes for \code{band} scales such that ticks are placed on band boundaries rather in the middle of a band. Use in conjunction with \code{"bandPosition": 1} and an axis \code{"padding"} value of \code{0}.}

\item{tickOffset}{(\emph{AxisConfig}) Position offset in pixels to apply to ticks, labels, and gridlines.}

\item{tickOpacity}{(\emph{AxisConfig})}

\item{tickRound}{(\emph{AxisConfig}) Boolean flag indicating if pixel position values should be rounded to the nearest integer.

\strong{Default value:} \code{true}}

\item{tickSize}{(\emph{AxisConfig}) The size in pixels of axis ticks.

\strong{Default value:} \code{5}}

\item{tickWidth}{(\emph{AxisConfig})}

\item{ticks}{(\emph{AxisConfig}) Boolean value that determines whether the axis should include ticks.

\strong{Default value:} \code{true}}

\item{title}{(\emph{AxisConfig}) Set to null to disable title for the axis, legend, or header.}

\item{titleAlign}{(\emph{AxisConfig}) Horizontal text alignment of axis titles.}

\item{titleAnchor}{(\emph{AxisConfig}) Text anchor position for placing axis titles.}

\item{titleAngle}{(\emph{AxisConfig}) Angle in degrees of axis titles.}

\item{titleBaseline}{(\emph{AxisConfig}) Vertical text baseline for axis titles.}

\item{titleColor}{(\emph{AxisConfig}) Color of the title, can be in hex color code or regular color name.}

\item{titleFont}{(\emph{AxisConfig}) Font of the title. (e.g., \code{"Helvetica Neue"}).}

\item{titleFontSize}{(\emph{AxisConfig}) Font size of the title.}

\item{titleFontStyle}{(\emph{AxisConfig}) Font style of the title.}

\item{titleFontWeight}{(\emph{AxisConfig}) Font weight of the title.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{titleLimit}{(\emph{AxisConfig}) Maximum allowed pixel width of axis titles.}

\item{titleLineHeight}{(\emph{AxisConfig}) Line height in pixels for multi-line title text.}

\item{titleOpacity}{(\emph{AxisConfig}) Opacity of the axis title.}

\item{titlePadding}{(\emph{AxisConfig}) The padding, in pixels, between title and axis.}

\item{titleX}{(\emph{AxisConfig}) X-coordinate of the axis title relative to the axis group.}

\item{titleY}{(\emph{AxisConfig}) Y-coordinate of the axis title relative to the axis group.}

\item{translate}{(\emph{AxisConfig}) Translation offset in pixels applied to the axis group mark x and y. If specified, overrides the default behavior of a 0.5 offset to pixel-align stroked lines.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add axisBand config (AxisConfig) to a vega-lite spec.
}
