% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_axisYTemporal}
\alias{vl_config_axisYTemporal}
\title{vl_config_axisYTemporal}
\usage{
vl_config_axisYTemporal(
  spec,
  aria = NULL,
  bandPosition = NULL,
  description = NULL,
  disable = NULL,
  domain = NULL,
  domainCap = NULL,
  domainColor = NULL,
  domainDash = NULL,
  domainDashOffset = NULL,
  domainOpacity = NULL,
  domainWidth = NULL,
  format = NULL,
  formatType = NULL,
  grid = NULL,
  gridCap = NULL,
  gridColor = NULL,
  gridDash = NULL,
  gridDashOffset = NULL,
  gridOpacity = NULL,
  gridWidth = NULL,
  labelAlign = NULL,
  labelAngle = NULL,
  labelBaseline = NULL,
  labelBound = NULL,
  labelColor = NULL,
  labelExpr = NULL,
  labelFlush = NULL,
  labelFlushOffset = NULL,
  labelFont = NULL,
  labelFontSize = NULL,
  labelFontStyle = NULL,
  labelFontWeight = NULL,
  labelLimit = NULL,
  labelLineHeight = NULL,
  labelOffset = NULL,
  labelOpacity = NULL,
  labelOverlap = NULL,
  labelPadding = NULL,
  labelSeparation = NULL,
  labels = NULL,
  maxExtent = NULL,
  minExtent = NULL,
  offset = NULL,
  orient = NULL,
  position = NULL,
  style = NULL,
  tickBand = NULL,
  tickCap = NULL,
  tickColor = NULL,
  tickCount = NULL,
  tickDash = NULL,
  tickDashOffset = NULL,
  tickExtra = NULL,
  tickMinStep = NULL,
  tickOffset = NULL,
  tickOpacity = NULL,
  tickRound = NULL,
  tickSize = NULL,
  tickWidth = NULL,
  ticks = NULL,
  title = NULL,
  titleAlign = NULL,
  titleAnchor = NULL,
  titleAngle = NULL,
  titleBaseline = NULL,
  titleColor = NULL,
  titleFont = NULL,
  titleFontSize = NULL,
  titleFontStyle = NULL,
  titleFontWeight = NULL,
  titleLimit = NULL,
  titleLineHeight = NULL,
  titleOpacity = NULL,
  titlePadding = NULL,
  titleX = NULL,
  titleY = NULL,
  translate = NULL,
  values = NULL,
  zindex = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{aria}{A boolean flag indicating if \href{https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA}{ARIA attributes}
should be included (SVG output only). If \code{false}, the "aria-hidden" attribute
will be set on the output SVG group, removing the axis from the ARIA
accessibility tree.

\strong{Default value:} \code{true}

Or: ExprRef}

\item{bandPosition}{An interpolation fraction indicating where, for \code{band} scales, axis ticks should
be positioned. A value of \code{0} places ticks at the left edge of their bands. A
value of \code{0.5} places ticks in the middle of their bands.

\strong{Default value:} \code{0.5}

Or: ExprRef}

\item{description}{A text description of this axis for \href{https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA}{ARIA accessibility}
(SVG output only). If the \code{aria} property
is true, for SVG output the \href{https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_aria-label_attribute}{"aria-label" attribute}
will be set to this description. If the description is unspecified it will be
automatically generated.

Or: ExprRef}

\item{disable}{Disable axis by default.}

\item{domain}{A boolean flag indicating if the domain (the axis baseline) should be included
as part of the axis.

\strong{Default value:} \code{true}}

\item{domainCap}{The stroke cap for the domain line's ending style. One of \code{"butt"}, \code{"round"} or
\code{"square"}.

\strong{Default value:} \code{"butt"}

Or: ExprRef}

\item{domainColor}{Color of axis domain line.

\strong{Default value:} \code{"gray"}.

Or: ExprRef}

\item{domainDash}{An array of alternating [stroke, space] lengths for dashed domain lines.

Or: ExprRef}

\item{domainDashOffset}{The pixel offset at which to start drawing with the domain dash array.

Or: ExprRef}

\item{domainOpacity}{Opacity of the axis domain line.

Or: ExprRef}

\item{domainWidth}{Stroke width of axis domain line

\strong{Default value:} \code{1}

Or: ExprRef}

\item{format}{When used with the default \code{"number"} and \code{"time"} format type, the text
formatting pattern for labels of guides (axes, legends, headers) and text marks.
\itemize{
\item If the format type is \code{"number"} (e.g., for quantitative fields), this is D3's
\href{https://github.com/d3/d3-format#locale_format}{number format pattern}.
\item If the format type is \code{"time"} (e.g., for temporal fields), this is D3's \href{https://github.com/d3/d3-time-format#locale_format}{time format pattern}.
}

See the \href{https://vega.github.io/vega-lite/docs/format.html}{format documentation} for more
examples.

When used with a \href{https://vega.github.io/vega-lite/docs/config.html#custom-format-type}{custom \code{formatType}},
this value will be passed as \code{format} alongside \code{datum.value} to the registered
function.

\strong{Default value:} Derived from
\href{https://vega.github.io/vega-lite/docs/config.html#format}{numberFormat}
config for number format and from
\href{https://vega.github.io/vega-lite/docs/config.html#format}{timeFormat} config
for time format.}

\item{formatType}{The format type for labels. One of \code{"number"},
\code{"time"}, or a \href{https://vega.github.io/vega-lite/docs/config.html#custom-format-type}{registered custom format type}.

\strong{Default value:}
\itemize{
\item \code{"time"} for temporal fields and ordinal and nominal fields with \code{timeUnit}.
\item \code{"number"} for quantitative fields as well as ordinal and nominal fields
without \code{timeUnit}.
}}

\item{grid}{A boolean flag indicating if grid lines should be included as part of the axis

\strong{Default value:} \code{true} for \href{https://vega.github.io/vega-lite/docs/scale.html#continuous}{continuous scales} that are
not binned; otherwise, \code{false}.}

\item{gridCap}{The stroke cap for grid lines' ending style. One of \code{"butt"}, \code{"round"} or
\code{"square"}.

\strong{Default value:} \code{"butt"}

Or: ExprRef}

\item{gridColor}{Color of gridlines.

\strong{Default value:} \code{"lightGray"}.

Or: ExprRef

Or: ConditionalAxisColor}

\item{gridDash}{An array of alternating [stroke, space] lengths for dashed grid lines.

Or: ExprRef

Or: ConditionalAxisNumberArray}

\item{gridDashOffset}{The pixel offset at which to start drawing with the grid dash array.

Or: ExprRef

Or: ConditionalAxisNumber}

\item{gridOpacity}{The stroke opacity of grid (value between [0,1])

\strong{Default value:} \code{1}

Or: ExprRef

Or: ConditionalAxisNumber}

\item{gridWidth}{The grid width, in pixels.

\strong{Default value:} \code{1}

Or: ExprRef

Or: ConditionalAxisNumber}

\item{labelAlign}{Horizontal text alignment of axis tick labels, overriding the default setting
for the current axis orientation.

Or: ExprRef

Or: ConditionalAxisLabelAlign}

\item{labelAngle}{The rotation angle of the axis labels.

\strong{Default value:} \code{-90} for nominal and ordinal fields; \code{0} otherwise.

Or: ExprRef}

\item{labelBaseline}{Vertical text baseline of axis tick labels, overriding the default setting
for the current axis orientation. One of \code{"alphabetic"} (default), \code{"top"},
\code{"middle"}, \code{"bottom"}, \code{"line-top"}, or \code{"line-bottom"}. The \code{"line-top"} and
\code{"line-bottom"} values operate similarly to \code{"top"} and \code{"bottom"}, but are
calculated relative to the \emph{lineHeight} rather than \emph{fontSize} alone.

Or: ExprRef

Or: ConditionalAxisLabelBaseline}

\item{labelBound}{Indicates if labels should be hidden if they exceed the axis range. If \code{false}
(the default) no bounds overlap analysis is performed. If \code{true}, labels will be
hidden if they exceed the axis range by more than 1 pixel. If this property is
a number, it specifies the pixel tolerance: the maximum amount by which a label
bounding box may exceed the axis range.

\strong{Default value:} \code{false}.

Or: ExprRef}

\item{labelColor}{The color of the tick label, can be in hex color code or regular color name.

Or: ExprRef

Or: ConditionalAxisColor}

\item{labelExpr}{\href{https://vega.github.io/vega/docs/expressions/}{Vega expression} for customizing
labels.

\strong{Note:} The label text and value can be assessed via the \code{label} and \code{value}
properties of the axis's backing \code{datum} object.}

\item{labelFlush}{Indicates if the first and last axis labels should be aligned flush with the
scale range. Flush alignment for a horizontal axis will left-align the first
label and right-align the last label. For vertical axes, bottom and top text
baselines are applied instead. If this property is a number, it also indicates
the number of pixels by which to offset the first and last labels; for example,
a value of 2 will flush-align the first and last labels and also push them
2 pixels outward from the center of the axis. The additional adjustment can
sometimes help the labels better visually group with corresponding axis ticks.

\strong{Default value:} \code{true} for axis of a continuous x-scale. Otherwise, \code{false}.}

\item{labelFlushOffset}{Indicates the number of pixels by which to offset flush-adjusted labels. For
example, a value of \code{2} will push flush-adjusted labels 2 pixels outward from
the center of the axis. Offsets can help the labels better visually group with
corresponding axis ticks.

\strong{Default value:} \code{0}.

Or: ExprRef}

\item{labelFont}{The font of the tick label.

Or: ExprRef

Or: ConditionalAxisString}

\item{labelFontSize}{The font size of the label, in pixels.

Or: ExprRef

Or: ConditionalAxisNumber}

\item{labelFontStyle}{Font style of the title.

Or: ExprRef

Or: ConditionalAxisLabelFontStyle}

\item{labelFontWeight}{Font weight of axis tick labels.

Or: ExprRef

Or: ConditionalAxisLabelFontWeight}

\item{labelLimit}{Maximum allowed pixel width of axis tick labels.

\strong{Default value:} \code{180}

Or: ExprRef}

\item{labelLineHeight}{Line height in pixels for multi-line label text or label text with \code{"line-top"}
or \code{"line-bottom"} baseline.

Or: ExprRef}

\item{labelOffset}{Position offset in pixels to apply to labels, in addition to tickOffset.

\strong{Default value:} \code{0}

Or: ExprRef

Or: ConditionalAxisNumber}

\item{labelOpacity}{The opacity of the labels.

Or: ExprRef

Or: ConditionalAxisNumber}

\item{labelOverlap}{The strategy to use for resolving overlap of axis labels. If \code{false} (the
default), no overlap reduction is attempted. If set to \code{true} or \code{"parity"},
a strategy of removing every other label is used (this works well for standard
linear axes). If set to \code{"greedy"}, a linear scan of the labels is performed,
removing any labels that overlaps with the last visible label (this often works
better for log-scaled axes).

\strong{Default value:} \code{true} for non-nominal fields with non-log scales; \code{"greedy"}
for log scales; otherwise \code{false}.}

\item{labelPadding}{The padding in pixels between labels and ticks.

\strong{Default value:} \code{2}

Or: ExprRef

Or: ConditionalAxisNumber}

\item{labelSeparation}{The minimum separation that must be between label bounding boxes for them
to be considered non-overlapping (default \code{0}). This property is ignored if
\emph{labelOverlap} resolution is not enabled.

Or: ExprRef}

\item{labels}{A boolean flag indicating if labels should be included as part of the axis.

\strong{Default value:} \code{true}.}

\item{maxExtent}{The maximum extent in pixels that axis ticks and labels should use. This
determines a maximum offset value for axis titles.

\strong{Default value:} \code{undefined}.

Or: ExprRef}

\item{minExtent}{The minimum extent in pixels that axis ticks and labels should use. This
determines a minimum offset value for axis titles.

\strong{Default value:} \code{30} for y-axis; \code{undefined} for x-axis.

Or: ExprRef}

\item{offset}{The offset, in pixels, by which to displace the axis from the edge of the
enclosing group or data rectangle.

\strong{Default value:} derived from the \href{https://vega.github.io/vega-lite/docs/config.html#facet-scale-config}{axis config}'s
\code{offset} (\code{0} by default)}

\item{orient}{The orientation of the axis. One of \code{"top"}, \code{"bottom"}, \code{"left"} or \code{"right"}.
The orientation can be used to further specialize the axis type (e.g., a y-axis
oriented towards the right edge of the chart).

\strong{Default value:} \code{"bottom"} for x-axes and \code{"left"} for y-axes.}

\item{position}{The anchor position of the axis in pixels. For x-axes with top or bottom
orientation, this sets the axis group x coordinate. For y-axes with left or
right orientation, this sets the axis group y coordinate.

\strong{Default value}: \code{0}}

\item{style}{A string or array of strings indicating the name of
custom styles to apply to the axis. A style is a named
collection of axis property defined within the \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style configuration}. If
style is an array, later styles will override earlier styles.

\strong{Default value:} (none) \strong{Note:} Any specified style will augment the default
style. For example, an x-axis mark with \code{"style": "foo"} will use \code{config.axisX}
and \code{config.style.foo} (the specified style \code{"foo"} has higher precedence).}

\item{tickBand}{For band scales, indicates if ticks and grid lines should be placed at the
\code{"center"} of a band (default) or at the band \code{"extent"}s to indicate intervals

Or: ExprRef}

\item{tickCap}{The stroke cap for the tick lines' ending style. One of \code{"butt"}, \code{"round"} or
\code{"square"}.

\strong{Default value:} \code{"butt"}

Or: ExprRef}

\item{tickColor}{The color of the axis's tick.

\strong{Default value:} \code{"gray"}

Or: ExprRef

Or: ConditionalAxisColor}

\item{tickCount}{A desired number of ticks, for axes visualizing quantitative scales. The
resulting number may be different so that values are "nice" (multiples of 2, 5,
10) and lie within the underlying scale's range.

For scales of type \code{"time"} or \code{"utc"}, the tick count can instead be a time
interval specifier. Legal string values are \code{"millisecond"}, \code{"second"},
\code{"minute"}, \code{"hour"}, \code{"day"}, \code{"week"}, \code{"month"}, and \code{"year"}. Alternatively,
an object-valued interval specifier of the form \verb{\{"interval": "month", "step": 3\}} includes a desired number of interval steps. Here, ticks are generated for
each quarter (Jan, Apr, Jul, Oct) boundary.

\strong{Default value}: Determine using a formula \code{ceil(width/40)} for x and
\code{ceil(height/40)} for y.}

\item{tickDash}{An array of alternating [stroke, space] lengths for dashed tick mark lines.

Or: ExprRef

Or: ConditionalAxisNumberArray}

\item{tickDashOffset}{The pixel offset at which to start drawing with the tick mark dash array.

Or: ExprRef

Or: ConditionalAxisNumber}

\item{tickExtra}{Boolean flag indicating if an extra axis tick should be added for the initial
position of the axis. This flag is useful for styling axes for \code{band} scales
such that ticks are placed on band boundaries rather in the middle of a band.
Use in conjunction with \code{"bandPosition": 1} and an axis \code{"padding"} value of
\code{0}.}

\item{tickMinStep}{The minimum desired step between axis ticks, in terms of scale domain values.
For example, a value of \code{1} indicates that ticks should not be less than 1 unit
apart. If \code{tickMinStep} is specified, the \code{tickCount} value will be adjusted, if
necessary, to enforce the minimum step value.}

\item{tickOffset}{Position offset in pixels to apply to ticks, labels, and gridlines.

Or: ExprRef}

\item{tickOpacity}{Opacity of the ticks.

Or: ExprRef

Or: ConditionalAxisNumber}

\item{tickRound}{Boolean flag indicating if pixel position values should be rounded to the
nearest integer.

\strong{Default value:} \code{true}}

\item{tickSize}{The size in pixels of axis ticks.

\strong{Default value:} \code{5}

Or: ExprRef

Or: ConditionalAxisNumber}

\item{tickWidth}{The width, in pixels, of ticks.

\strong{Default value:} \code{1}

Or: ExprRef

Or: ConditionalAxisNumber}

\item{ticks}{Boolean value that determines whether the axis should include ticks.

\strong{Default value:} \code{true}}

\item{title}{A title for the field. If \code{null}, the title will be removed.

\strong{Default value:} derived from the field's name and transformation function
(\code{aggregate}, \code{bin} and \code{timeUnit}). If the field has an aggregate function,
the function is displayed as part of the title (e.g., \code{"Sum of Profit"}). If
the field is binned or has a time unit applied, the applied function is shown
in parentheses (e.g., \code{"Profit (binned)"}, \code{"Transaction Date (year-month)"}).
Otherwise, the title is simply the field name.

\strong{Notes}:
\enumerate{
\item You can customize the default field title format by providing the
\href{https://vega.github.io/vega-lite/docs/config.html#top-level-config}{\code{fieldTitle}}
property in the \href{https://vega.github.io/vega-lite/docs/config.html}{config}
or \href{https://vega.github.io/vega-lite/usage/compile.html#field-title}{\code{fieldTitle} function via the \code{compile} function's options}.
\item If both field definition's \code{title} and axis, header, or legend \code{title} are
defined, axis/header/legend title will be used.
}}

\item{titleAlign}{Horizontal text alignment of axis titles.

Or: ExprRef}

\item{titleAnchor}{Text anchor position for placing axis titles.

Or: ExprRef}

\item{titleAngle}{Angle in degrees of axis titles.

Or: ExprRef}

\item{titleBaseline}{Vertical text baseline for axis titles. One of \code{"alphabetic"} (default),
\code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, or \code{"line-bottom"}. The
\code{"line-top"} and \code{"line-bottom"} values operate similarly to \code{"top"} and
\code{"bottom"}, but are calculated relative to the \emph{lineHeight} rather than
\emph{fontSize} alone.

Or: ExprRef}

\item{titleColor}{Color of the title, can be in hex color code or regular color name.

Or: ExprRef}

\item{titleFont}{Font of the title. (e.g., \code{"Helvetica Neue"}).

Or: ExprRef}

\item{titleFontSize}{Font size of the title.

Or: ExprRef}

\item{titleFontStyle}{Font style of the title.

Or: ExprRef}

\item{titleFontWeight}{Font weight of the title. This can be either a string (e.g \code{"bold"}, \code{"normal"})
or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and
\code{"bold"} = \code{700}).

Or: ExprRef}

\item{titleLimit}{Maximum allowed pixel width of axis titles.

Or: ExprRef}

\item{titleLineHeight}{Line height in pixels for multi-line title text or title text with \code{"line-top"}
or \code{"line-bottom"} baseline.

Or: ExprRef}

\item{titleOpacity}{Opacity of the axis title.

Or: ExprRef}

\item{titlePadding}{The padding, in pixels, between title and axis.

Or: ExprRef}

\item{titleX}{X-coordinate of the axis title relative to the axis group.

Or: ExprRef}

\item{titleY}{Y-coordinate of the axis title relative to the axis group.

Or: ExprRef}

\item{translate}{Coordinate space translation offset for axis layout. By default, axes are
translated by a 0.5 pixel offset for both the x and y coordinates in order to
align stroked lines with the pixel grid. However, for vector graphics output
these pixel-specific adjustments may be undesirable, in which case translate can
be changed (for example, to zero).

\strong{Default value:} \code{0.5}

Or: ExprRef}

\item{values}{Explicitly set the visible axis tick values.}

\item{zindex}{A non-negative integer indicating the z-index of the axis. If zindex is 0, axes
should be drawn behind all chart elements. To put them in front, set \code{zindex} to
\code{1} or more.

\strong{Default value:} \code{0} (behind the marks).}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add axisYTemporal config (AxisConfig) to a vega-lite spec.
}
