% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_concat}
\alias{vl_config_concat}
\title{vl_config_concat}
\usage{
vl_config_concat(spec, .object = NULL, columns = NULL, spacing = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: CompositionConfig}

\item{columns}{(\emph{CompositionConfig}) The number of columns to include in the view composition layout.

\strong{Default value}: \code{undefined} -- An infinite number of columns (a single row) will be assumed. This is equivalent to \code{hconcat} (for \code{concat}) and to using the \code{column} channel (for \code{facet} and \code{repeat}).

\strong{Note}:
\enumerate{
\item This property is only for: - the general (wrappable) \code{concat} operator (not \code{hconcat}/\code{vconcat}) - the \code{facet} and \code{repeat} operator with one field/repetition definition (without row/column nesting)
\item Setting the \code{columns} to \code{1} is equivalent to \code{vconcat} (for \code{concat}) and to using the \code{row} channel (for \code{facet} and \code{repeat}).
}}

\item{spacing}{(\emph{CompositionConfig}) The default spacing in pixels between composed sub-views.

\strong{Default value}: \code{20}}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add concat config (CompositionConfig) to a vega-lite spec.
}
