% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_mark}
\alias{vl_config_mark}
\title{vl_config_mark}
\usage{
vl_config_mark(spec, .object = NULL, align = NULL, angle = NULL,
  aspect = NULL, baseline = NULL, color = NULL,
  cornerRadius = NULL, cornerRadiusBottomLeft = NULL,
  cornerRadiusBottomRight = NULL, cornerRadiusTopLeft = NULL,
  cornerRadiusTopRight = NULL, cursor = NULL, dir = NULL,
  dx = NULL, dy = NULL, ellipsis = NULL, fill = NULL,
  fillOpacity = NULL, filled = NULL, font = NULL, fontSize = NULL,
  fontStyle = NULL, fontWeight = NULL, height = NULL, href = NULL,
  interpolate = NULL, invalid = NULL, limit = NULL,
  lineBreak = NULL, lineHeight = NULL, opacity = NULL,
  order = NULL, orient = NULL, radius = NULL, shape = NULL,
  size = NULL, stroke = NULL, strokeCap = NULL, strokeDash = NULL,
  strokeDashOffset = NULL, strokeJoin = NULL,
  strokeMiterLimit = NULL, strokeOpacity = NULL, strokeWidth = NULL,
  tension = NULL, text = NULL, theta = NULL, timeUnitBand = NULL,
  timeUnitBandPosition = NULL, tooltip = NULL, width = NULL,
  x = NULL, x2 = NULL, y = NULL, y2 = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: MarkConfig}

\item{align}{(\emph{MarkConfig}) The horizontal alignment of the text or ranged marks (area, bar, image, rect, rule). One of \code{"left"}, \code{"right"}, \code{"center"}.}

\item{angle}{(\emph{MarkConfig}) The rotation angle of the text, in degrees.}

\item{aspect}{(\emph{MarkConfig}) Whether to keep aspect ratio of image marks.}

\item{baseline}{(\emph{MarkConfig}) The vertical alignment of the text or ranged marks (area, bar, image, rect, rule). One of \code{"top"}, \code{"middle"}, \code{"bottom"}.

\strong{Default value:} \code{"middle"}}

\item{color}{(\emph{MarkConfig}) Default color.

\strong{Default value:} <span style="color: #4682b4;">■</span> \code{"#4682b4"}

\strong{Note:}
\itemize{
\item This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}.
\item The \code{fill} and \code{stroke} properties have higher precedence than \code{color} and will override \code{color}.
}}

\item{cornerRadius}{(\emph{MarkConfig}) The radius in pixels of rounded rectangle corners.

\strong{Default value:} \code{0}}

\item{cornerRadiusBottomLeft}{(\emph{MarkConfig}) The radius in pixels of rounded rectangle bottom left corner.

\strong{Default value:} \code{0}}

\item{cornerRadiusBottomRight}{(\emph{MarkConfig}) The radius in pixels of rounded rectangle bottom right corner.

\strong{Default value:} \code{0}}

\item{cornerRadiusTopLeft}{(\emph{MarkConfig}) The radius in pixels of rounded rectangle top right corner.

\strong{Default value:} \code{0}}

\item{cornerRadiusTopRight}{(\emph{MarkConfig}) The radius in pixels of rounded rectangle top left corner.

\strong{Default value:} \code{0}}

\item{cursor}{(\emph{MarkConfig}) The mouse cursor used over the mark. Any valid \href{https://developer.mozilla.org/en-US/docs/Web/CSS/cursor#Values}{CSS cursor type} can be used.}

\item{dir}{(\emph{MarkConfig}) The direction of the text. One of \code{"ltr"} (left-to-right) or \code{"rtl"} (right-to-left). This property determines on which side is truncated in response to the limit parameter.

\strong{Default value:} \code{"ltr"}}

\item{dx}{(\emph{MarkConfig}) The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the \emph{angle} property.}

\item{dy}{(\emph{MarkConfig}) The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the \emph{angle} property.}

\item{ellipsis}{(\emph{MarkConfig}) The ellipsis string for text truncated in response to the limit parameter.

\strong{Default value:} \code{"…"}}

\item{fill}{(\emph{MarkConfig}) Default Fill Color. This has higher precedence than \code{config.color}.

\strong{Default value:} (None)}

\item{fillOpacity}{(\emph{MarkConfig}) The fill opacity (value between [0,1]).

\strong{Default value:} \code{1}}

\item{filled}{(\emph{MarkConfig}) Whether the mark's color should be used as fill color instead of stroke color.

\strong{Default value:} \code{false} for all \code{point}, \code{line}, and \code{rule} marks as well as \code{geoshape} marks for \href{https://vega.github.io/vega-lite/docs/data.html#graticule}{graticule} data sources; otherwise, \code{true}.

\strong{Note:} This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}.}

\item{font}{(\emph{MarkConfig}) The typeface to set the text in (e.g., \code{"Helvetica Neue"}).}

\item{fontSize}{(\emph{MarkConfig}) The font size, in pixels.}

\item{fontStyle}{(\emph{MarkConfig}) The font style (e.g., \code{"italic"}).}

\item{fontWeight}{(\emph{MarkConfig}) The font weight.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{height}{(\emph{MarkConfig}) Height of the marks.}

\item{href}{(\emph{MarkConfig}) A URL to load upon mouse click. If defined, the mark acts as a hyperlink.}

\item{interpolate}{(\emph{MarkConfig}) The line interpolation method to use for line and area marks. One of the following:
\itemize{
\item \code{"linear"}: piecewise linear segments, as in a polyline.
\item \code{"linear-closed"}: close the linear segments to form a polygon.
\item \code{"step"}: alternate between horizontal and vertical segments, as in a step function.
\item \code{"step-before"}: alternate between vertical and horizontal segments, as in a step function.
\item \code{"step-after"}: alternate between horizontal and vertical segments, as in a step function.
\item \code{"basis"}: a B-spline, with control point duplication on the ends.
\item \code{"basis-open"}: an open B-spline; may not intersect the start or end.
\item \code{"basis-closed"}: a closed B-spline, as in a loop.
\item \code{"cardinal"}: a Cardinal spline, with control point duplication on the ends.
\item \code{"cardinal-open"}: an open Cardinal spline; may not intersect the start or end, but will intersect other control points.
\item \code{"cardinal-closed"}: a closed Cardinal spline, as in a loop.
\item \code{"bundle"}: equivalent to basis, except the tension parameter is used to straighten the spline.
\item \code{"monotone"}: cubic interpolation that preserves monotonicity in y.
}}

\item{invalid}{(\emph{MarkConfig}) Defines how Vega-Lite should handle marks for invalid values (\code{null} and \code{NaN}).
\itemize{
\item If set to \code{"filter"} (default), all data items with null values will be skipped (for line, trail, and area marks) or filtered (for other marks).
\item If \code{null}, all data items are included. In this case, invalid values will be interpreted as zeroes.
}}

\item{limit}{(\emph{MarkConfig}) The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

\strong{Default value:} \code{0}, indicating no limit}

\item{lineBreak}{(\emph{MarkConfig}) A delimiter, such as a newline character, upon which to break text strings into multiple lines. This property will be ignored if the text property is array-valued.}

\item{lineHeight}{(\emph{MarkConfig}) The height, in pixels, of each line of text in a multi-line text mark.}

\item{opacity}{(\emph{MarkConfig}) The overall opacity (value between [0,1]).

\strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick}, \code{circle}, or \code{square} marks or layered \code{bar} charts and \code{1} otherwise.}

\item{order}{(\emph{MarkConfig}) For line and trail marks, this \code{order} property can be set to \code{null} or \code{false} to make the lines use the original order in the data sources.}

\item{orient}{(\emph{MarkConfig}) The orientation of a non-stacked bar, tick, area, and line charts.
The value is either horizontal (default) or vertical.
\itemize{
\item For bar, rule and tick, this determines whether the size of the bar and tick
should be applied to x or y dimension.
\item For area, this property determines the orient property of the Vega output.
\item For line and trail marks, this property determines the sort order of the points in the line
if \code{config.sortLineBy} is not specified.
For stacked charts, this is always determined by the orientation of the stack;
therefore explicitly specified value will be ignored.
}}

\item{radius}{(\emph{MarkConfig}) Polar coordinate radial offset, in pixels, of the text label from the origin determined by the \code{x} and \code{y} properties.}

\item{shape}{(\emph{MarkConfig}) Shape of the point marks. Supported values include:
\itemize{
\item plotting shapes: \code{"circle"}, \code{"square"}, \code{"cross"}, \code{"diamond"}, \code{"triangle-up"}, \code{"triangle-down"}, \code{"triangle-right"}, or \code{"triangle-left"}.
\item the line symbol \code{"stroke"}
\item centered directional shapes \code{"arrow"}, \code{"wedge"}, or \code{"triangle"}
\item a custom \href{https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths}{SVG path string} (For correct sizing, custom shape paths should be defined within a square bounding box with coordinates ranging from -1 to 1 along both the x and y dimensions.)
}

\strong{Default value:} \code{"circle"}}

\item{size}{(\emph{MarkConfig}) Default size for marks.
\itemize{
\item For \code{point}/\code{circle}/\code{square}, this represents the pixel area of the marks. For example: in the case of circles, the radius is determined in part by the square root of the size value.
\item For \code{bar}, this represents the band size of the bar, in pixels.
\item For \code{text}, this represents the font size, in pixels.
}

\strong{Default value:}
\itemize{
\item \code{30} for point, circle, square marks; width/height's \code{step}
\item \code{2} for bar marks with discrete dimensions;
\item \code{5} for bar marks with continuous dimensions;
\item \code{11} for text marks.
}}

\item{stroke}{(\emph{MarkConfig}) Default Stroke Color. This has higher precedence than \code{config.color}.

\strong{Default value:} (None)}

\item{strokeCap}{(\emph{MarkConfig}) The stroke cap for line ending style. One of \code{"butt"}, \code{"round"}, or \code{"square"}.

\strong{Default value:} \code{"square"}}

\item{strokeDash}{(\emph{MarkConfig}) An array of alternating stroke, space lengths for creating dashed or dotted lines.}

\item{strokeDashOffset}{(\emph{MarkConfig}) The offset (in pixels) into which to begin drawing with the stroke dash array.}

\item{strokeJoin}{(\emph{MarkConfig}) The stroke line join method. One of \code{"miter"}, \code{"round"} or \code{"bevel"}.

\strong{Default value:} \code{"miter"}}

\item{strokeMiterLimit}{(\emph{MarkConfig}) The miter limit at which to bevel a line join.}

\item{strokeOpacity}{(\emph{MarkConfig}) The stroke opacity (value between [0,1]).

\strong{Default value:} \code{1}}

\item{strokeWidth}{(\emph{MarkConfig}) The stroke width, in pixels.}

\item{tension}{(\emph{MarkConfig}) Depending on the interpolation type, sets the tension parameter (for line and area marks).}

\item{text}{(\emph{MarkConfig}) Placeholder text if the \code{text} channel is not specified}

\item{theta}{(\emph{MarkConfig}) Polar coordinate angle, in radians, of the text label from the origin determined by the \code{x} and \code{y} properties. Values for \code{theta} follow the same convention of \code{arc} mark \code{startAngle} and \code{endAngle} properties: angles are measured in radians, with \code{0} indicating "north".}

\item{timeUnitBand}{(\emph{MarkConfig}) Default relative band size for a time unit. If set to \code{1}, the bandwidth of the marks will be equal to the time unit band step.
If set to \code{0.5}, bandwidth of the marks will be half of the time unit band step.}

\item{timeUnitBandPosition}{(\emph{MarkConfig}) Default relative band position for a time unit. If set to \code{0}, the marks will be positioned at the beginning of the time unit band step.
If set to \code{0.5}, the marks will be positioned in the middle of the time unit band step.}

\item{tooltip}{(\emph{MarkConfig}) The tooltip text string to show upon mouse hover or an object defining which fields should the tooltip be derived from.
\itemize{
\item If \code{tooltip} is \code{true} or \code{{"content": "encoding"}}, then all fields from \code{encoding} will be used.
\item If \code{tooltip} is \code{{"content": "data"}}, then all fields that appear in the highlighted data point will be used.
\item If set to \code{null} or \code{false}, then no tooltip will be used.
}

See the \href{https://vega.github.io/vega-lite/docs/tooltip.html}{tooltip} documentation for a detailed discussion about tooltip  in Vega-Lite.

\strong{Default value:} \code{null}}

\item{width}{(\emph{MarkConfig}) Width of the marks.}

\item{x}{(\emph{MarkConfig}) X coordinates of the marks, or width of horizontal \code{"bar"} and \code{"area"} without specified \code{x2} or \code{width}.

The \code{value} of this channel can be a number or a string \code{"width"} for the width of the plot.}

\item{x2}{(\emph{MarkConfig}) X2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and  \code{"rule"}.

The \code{value} of this channel can be a number or a string \code{"width"} for the width of the plot.}

\item{y}{(\emph{MarkConfig}) Y coordinates of the marks, or height of vertical \code{"bar"} and \code{"area"} without specified \code{y2} or \code{height}.

The \code{value} of this channel can be a number or a string \code{"height"} for the height of the plot.}

\item{y2}{(\emph{MarkConfig}) Y2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and  \code{"rule"}.

The \code{value} of this channel can be a number or a string \code{"height"} for the height of the plot.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add mark config (MarkConfig) to a vega-lite spec.
}
