% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_range}
\alias{vl_config_range}
\title{vl_config_range}
\usage{
vl_config_range(spec, .object = NULL, category = NULL,
  diverging = NULL, heatmap = NULL, ordinal = NULL, ramp = NULL,
  symbol = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: RangeConfig}

\item{category}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for categorical data.}

\item{diverging}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for diverging quantitative ramps.}

\item{heatmap}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for quantitative heatmaps.}

\item{ordinal}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for rank-ordered data.}

\item{ramp}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for sequential quantitative ramps.}

\item{symbol}{(\emph{RangeConfig}) Array of \href{https://vega.github.io/vega/docs/marks/symbol/}{symbol} names or paths for the default shape palette.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add range config (RangeConfig) to a vega-lite spec.
}
