% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_title}
\alias{vl_config_title}
\title{vl_config_title}
\usage{
vl_config_title(
  spec,
  .object = NULL,
  align = NULL,
  anchor = NULL,
  angle = NULL,
  aria = NULL,
  baseline = NULL,
  color = NULL,
  dx = NULL,
  dy = NULL,
  font = NULL,
  fontSize = NULL,
  fontStyle = NULL,
  fontWeight = NULL,
  frame = NULL,
  limit = NULL,
  lineHeight = NULL,
  offset = NULL,
  orient = NULL,
  subtitleColor = NULL,
  subtitleFont = NULL,
  subtitleFontSize = NULL,
  subtitleFontStyle = NULL,
  subtitleFontWeight = NULL,
  subtitleLineHeight = NULL,
  subtitlePadding = NULL,
  zindex = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BaseTitleNoValueRefs}

\item{align}{(\emph{BaseTitleNoValueRefs}) Horizontal text alignment for title text. One of \code{"left"}, \code{"center"}, or \code{"right"}.}

\item{anchor}{(\emph{BaseTitleNoValueRefs})}

\item{angle}{(\emph{BaseTitleNoValueRefs})}

\item{aria}{(\emph{BaseTitleNoValueRefs})}

\item{baseline}{(\emph{BaseTitleNoValueRefs}) Vertical text baseline for title and subtitle text. One of \code{"alphabetic"} (default), \code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, or \code{"line-bottom"}. The \code{"line-top"} and \code{"line-bottom"} values operate similarly to \code{"top"} and \code{"bottom"}, but are calculated relative to the \emph{lineHeight} rather than \emph{fontSize} alone.}

\item{color}{(\emph{BaseTitleNoValueRefs})}

\item{dx}{(\emph{BaseTitleNoValueRefs})}

\item{dy}{(\emph{BaseTitleNoValueRefs})}

\item{font}{(\emph{BaseTitleNoValueRefs})}

\item{fontSize}{(\emph{BaseTitleNoValueRefs})}

\item{fontStyle}{(\emph{BaseTitleNoValueRefs})}

\item{fontWeight}{(\emph{BaseTitleNoValueRefs})}

\item{frame}{(\emph{BaseTitleNoValueRefs})}

\item{limit}{(\emph{BaseTitleNoValueRefs})}

\item{lineHeight}{(\emph{BaseTitleNoValueRefs})}

\item{offset}{(\emph{BaseTitleNoValueRefs})}

\item{orient}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleColor}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFont}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFontSize}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFontStyle}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleFontWeight}{(\emph{BaseTitleNoValueRefs})}

\item{subtitleLineHeight}{(\emph{BaseTitleNoValueRefs})}

\item{subtitlePadding}{(\emph{BaseTitleNoValueRefs})}

\item{zindex}{(\emph{BaseTitleNoValueRefs})}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add title config (TitleConfig) to a vega-lite spec.
}
