% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{vl_encode}
\alias{vl_encode}
\title{vl_encoding}
\usage{
vl_encode(spec, ...)
}
\arguments{
\item{spec}{a vegalite spec}

\item{...}{encodings to add. Each encoding should be a list. See example.}
}
\value{
a modified vegalite spec
}
\description{
Add encodings to a spec. This is one way to add encodings... each encoding
can also be added using vl_encoding_<name> where <name> is the name of the
encoding.
}
\examples{

vl_chart() \%>\%
  vl_add_data(values = mtcars) \%>\%
  vl_mark_point() \%>\%
  vl_encode(
    x = vl_X(field = "wt", type = "quantitative"),
    y = vl_Y(field = "mpg", type = "quantitative")
  )
}
