% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_encode_latitude2}
\alias{vl_encode_latitude2}
\title{vl_encode_latitude2}
\usage{
vl_encode_latitude2(
  spec,
  field = NULL,
  type = NULL,
  value = NULL,
  aggregate = NULL,
  bandPosition = NULL,
  bin = NULL,
  timeUnit = NULL,
  title = NULL,
  datum = NULL,
  .object = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{field}{\strong{Required.} A string defining the name of the field from which
to pull a data value or an object defining iterated values from the
\href{https://vega.github.io/vega-lite/docs/repeat.html}{\code{repeat}} operator.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/field.html}{\code{field}}
documentation.

\strong{Notes:} 1) Dots (\code{.}) and brackets (\code{[} and \verb{]}) can be used to access nested
objects (e.g., \code{"field": "foo.bar"} and \verb{"field": "foo\['bar'\]"}). If field names
contain dots or brackets but are not nested, you can use \verb{\\\\} to escape dots and
brackets (e.g., \code{"a\\\\.b"} and \verb{"a\\\\\[0\\\\\]"}). See more details about escaping in
the \href{https://vega.github.io/vega-lite/docs/field.html}{field documentation}. 2)
\code{field} is not required if \code{aggregate} is \code{count}.}

\item{type}{The type of measurement (\code{"quantitative"}, \code{"temporal"},
\code{"ordinal"}, or \code{"nominal"}) for the encoded field or constant
value (\code{datum}). It can also be a \code{"geojson"} type for encoding
\href{https://vega.github.io/vega-lite/docs/geoshape.html}{'geoshape'}.

Vega-Lite automatically infers data types in many cases as discussed below.
However, type is required for a field if: (1) the field is not nominal and the
field encoding has no specified \code{aggregate} (except \code{argmin} and \code{argmax}),
\code{bin}, scale type, custom \code{sort} order, nor \code{timeUnit} or (2) if you wish to use
an ordinal scale for a field with \code{bin} or \code{timeUnit}.

\strong{Default value:}
\enumerate{
\item For a data \code{field}, \code{"nominal"} is the default data type unless the field
encoding has \code{aggregate}, \code{channel}, \code{bin}, scale type, \code{sort}, or \code{timeUnit}
that satisfies the following criteria:
}
\itemize{
\item \code{"quantitative"} is the default type if (1) the encoded field contains \code{bin}
or \code{aggregate} except \code{"argmin"} and \code{"argmax"}, (2) the encoding channel is
\code{latitude} or \code{longitude} channel or (3) if the specified scale type is \href{https://vega.github.io/vega-lite/docs/scale.html#type}{a quantitative scale}.
\item \code{"temporal"} is the default type if (1) the encoded field contains \code{timeUnit}
or (2) the specified scale type is a time or utc scale
\item \verb{ordinal""} is the default type if (1)
the encoded field contains a \href{https://vega.github.io/vega-lite/docs/sort.html#specifying-custom-sort-order}{custom \code{sort} order},
(2) the specified scale type is an ordinal/point/band scale, or (3) the encoding
channel is \code{order}.
}
\enumerate{
\item For a constant value in data domain (\code{datum}):
}
\itemize{
\item \code{"quantitative"} if the datum is a number
\item \code{"nominal"} if the datum is a string
\item \code{"temporal"} if the datum is \href{https://vega.github.io/vega-lite/docs/datetime.html}{a date time object}
}

\strong{Note:}
\itemize{
\item Data \code{type} describes the semantics of the data rather than the primitive data
types (number, string, etc.). The same primitive data type can have different
types of measurement. For example, numeric data can represent quantitative,
ordinal, or nominal data.
\item Data values for a temporal field can be either a date-time string (e.g.,
\code{"2015-03-07 12:32:17"}, \code{"17:01"}, \code{"2015-03-16"}. \code{"2015"}) or a timestamp
number (e.g., \code{1552199579097}).
\item When using with \href{https://vega.github.io/vega-lite/docs/bin.html}{\code{bin}},
the \code{type} property can be either \code{"quantitative"} (for using
a linear bin scale) or \href{https://vega.github.io/vega-lite/docs/type.html#cast-bin}{\code{"ordinal"} (for using an ordinal bin scale)}.
\item When using with
\href{https://vega.github.io/vega-lite/docs/timeunit.html}{\code{timeUnit}},
the \code{type} property can be either \code{"temporal"} (default, for
using a temporal scale) or \href{https://vega.github.io/vega-lite/docs/type.html#cast-bin}{\code{"ordinal"} (for using an ordinal scale)}.
\item When using with
\href{https://vega.github.io/vega-lite/docs/aggregate.html}{\code{aggregate}}, the \code{type}
property refers to the post-aggregation data type. For example, we can calculate
count \code{distinct} of a categorical field \code{"cat"} using \verb{\{"aggregate": "distinct", "field": "cat"\}}. The \code{"type"} of the aggregate output is \code{"quantitative"}.
\item Secondary channels (e.g., \code{x2}, \code{y2}, \code{xError}, \code{yError}) do not have \code{type}
as they must have exactly the same type as their primary channels (e.g., \code{x},
\code{y}).
}

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/type.html}{\code{type}}
documentation.}

\item{value}{A constant value in visual domain (e.g., \code{"red"} / \code{"#0099ff"} / \href{https://vega.github.io/vega-lite/docs/types.html#gradient}{gradient definition} for
color, values between \code{0} to \code{1} for opacity).}

\item{aggregate}{Aggregation function for the field (e.g., \code{"mean"}, \code{"sum"}, \code{"median"},
\code{"min"}, \code{"max"}, \code{"count"}).

\strong{Default value:} \code{undefined} (None)

\strong{See also:}
\href{https://vega.github.io/vega-lite/docs/aggregate.html}{\code{aggregate}}
documentation.}

\item{bandPosition}{Relative position on a band of a stacked, binned, time unit, or band scale. For
example, the marks will be positioned at the beginning of the band if set to
\code{0}, and at the middle of the band if set to \code{0.5}.}

\item{bin}{A flag for binning a \code{quantitative} field, \href{https://vega.github.io/vega-lite/docs/bin.html#bin-parameters}{an object defining binning parameters},
or indicating that the data for \code{x} or \code{y} channel are binned before they are
imported into Vega-Lite (\code{"binned"}).
\itemize{
\item If \code{true}, default \href{https://vega.github.io/vega-lite/docs/bin.html#bin-parameters}{binning parameters} will
be applied.
\item If \code{"binned"}, this indicates that the data for the \code{x} (or \code{y})
channel are already binned. You can map the bin-start field to \code{x}
(or \code{y}) and the bin-end field to \code{x2} (or \code{y2}). The scale and
axis will be formatted similar to binning in Vega-Lite. To adjust
the axis ticks based on the bin step, you can also set the axis's
\href{https://vega.github.io/vega-lite/docs/axis.html#ticks}{\code{tickMinStep}} property.
}

\strong{Default value:} \code{false}

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/bin.html}{\code{bin}}
documentation.}

\item{timeUnit}{Time unit (e.g., \code{year}, \code{yearmonth}, \code{month}, \code{hours}) for
a temporal field. or \href{https://vega.github.io/vega-lite/docs/type.html#cast}{a temporal field that gets casted as ordinal}.

\strong{Default value:} \code{undefined} (None)

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/timeunit.html}{\code{timeUnit}}
documentation.}

\item{title}{A title for the field. If \code{null}, the title will be removed.

\strong{Default value:} derived from the field's name and transformation function
(\code{aggregate}, \code{bin} and \code{timeUnit}). If the field has an aggregate function,
the function is displayed as part of the title (e.g., \code{"Sum of Profit"}). If
the field is binned or has a time unit applied, the applied function is shown
in parentheses (e.g., \code{"Profit (binned)"}, \code{"Transaction Date (year-month)"}).
Otherwise, the title is simply the field name.

\strong{Notes}:
\enumerate{
\item You can customize the default field title format by providing the
\href{https://vega.github.io/vega-lite/docs/config.html#top-level-config}{\code{fieldTitle}}
property in the \href{https://vega.github.io/vega-lite/docs/config.html}{config}
or \href{https://vega.github.io/vega-lite/usage/compile.html#field-title}{\code{fieldTitle} function via the \code{compile} function's options}.
\item If both field definition's \code{title} and axis, header, or legend \code{title} are
defined, axis/header/legend title will be used.
}}

\item{datum}{A constant value in data domain.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than \code{spec}. Objects can be of type: SecondaryFieldDef, DatumDef or ValueDef<(number|"width"|"height"|ExprRef)>}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add encoding for latitude2 to a vega-lite spec.
}
