% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_encode_url}
\alias{vl_encode_url}
\title{vl_encode_url}
\usage{
vl_encode_url(
  spec,
  .object = NULL,
  field = NULL,
  type = NULL,
  aggregate = NULL,
  band = NULL,
  bin = NULL,
  condition = NULL,
  format = NULL,
  formatType = NULL,
  labelExpr = NULL,
  timeUnit = NULL,
  title = NULL,
  value = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: FieldOrDatumDefWithCondition<StringFieldDef,string> or ValueDefWithCondition<MarkPropFieldOrDatumDef,(string|null)>}

\item{field}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) \strong{Required.} A string defining the name of the field from which to pull a data value or an object defining iterated values from the \href{https://vega.github.io/vega-lite/docs/repeat.html}{\code{repeat}} operator.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/field.html}{\code{field}} documentation.

\strong{Notes:} 1)  Dots (\code{.}) and brackets (\code{[} and \verb{]}) can be used to access nested objects (e.g., \code{"field": "foo.bar"} and \verb{"field": "foo\['bar'\]"}). If field names contain dots or brackets but are not nested, you can use \verb{\\\\} to escape dots and brackets (e.g., \code{"a\\\\.b"} and \verb{"a\\\\\[0\\\\\]"}). See more details about escaping in the \href{https://vega.github.io/vega-lite/docs/field.html}{field documentation}. 2) \code{field} is not required if \code{aggregate} is \code{count}.}

\item{type}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) The type of measurement (\code{"quantitative"}, \code{"temporal"}, \code{"ordinal"}, or \code{"nominal"}) for the encoded field or constant value (\code{datum}). It can also be a \code{"geojson"} type for encoding \href{https://vega.github.io/vega-lite/docs/geoshape.html}{'geoshape'}.

Vega-Lite automatically infers data types in many cases as discussed below. However, type is required for a field if: (1) the field is not nominal and the field encoding has no specified \code{aggregate} (except \code{argmin} and \code{argmax}), \code{bin}, scale type, custom \code{sort} order, nor \code{timeUnit} or (2) if you wish to use an ordinal scale for a field with \code{bin} or \code{timeUnit}.

\strong{Default value:}
\enumerate{
\item For a data \code{field}, \code{"nominal"} is the default data type unless the field encoding has \code{aggregate}, \code{channel}, \code{bin}, scale type, \code{sort}, or \code{timeUnit} that satisfies the following criteria: - \code{"quantitative"} is the default type if (1) the encoded field contains \code{bin} or \code{aggregate} except \code{"argmin"} and \code{"argmax"}, (2) the encoding channel is \code{latitude} or \code{longitude} channel or (3) if the specified scale type is \href{https://vega.github.io/vega-lite/docs/scale.html#type}{a quantitative scale}. - \code{"temporal"} is the default type if (1) the encoded field contains \code{timeUnit} or (2) the specified scale type is a time or utc scale - \verb{ordinal""} is the default type if (1) the encoded field contains a \href{https://vega.github.io/vega-lite/docs/sort.html#specifying-custom-sort-order}{custom \code{sort} order}, (2) the specified scale type is an ordinal/point/band scale, or (3) the encoding channel i
\item For a constant value in data domain (\code{datum}): - \code{"quantitative"} if the datum is a number - \code{"nominal"} if the datum is a string - \code{"temporal"} if the datum is \href{https://vega.github.io/vega-lite/docs/datetime.html}{a date time object}
}

\strong{Note:} - Data \code{type} describes the semantics of the data rather than the primitive data types (number, string, etc.). The same primitive data type can have different types of measurement. For example, numeric data can represent quantitative, ordinal, or nominal data. - Data values for a temporal field can be either a date-time string (e.g., \code{"2015-03-07 12:32:17"}, \code{"17:01"}, \code{"2015-03-16"}. \code{"2015"}) or a timestamp number (e.g., \code{1552199579097}). - When using with \href{https://vega.github.io/vega-lite/docs/bin.html}{\code{bin}}, the \code{type} property can be either \code{"quantitative"} (for using a linear bin scale) or \href{https://vega.github.io/vega-lite/docs/type.html#cast-bin}{\code{"ordinal"} (for using an ordinal bin scale)}. - When using with \href{https://vega.github.io/vega-lite/docs/timeunit.html}{\code{timeUnit}}, the \code{type} property can be either \code{"temporal"} (default, for using a temporal scale) or \href{https://vega.gi

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/type.html}{\code{type}} documentation.}

\item{aggregate}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) Aggregation function for the field (e.g., \code{"mean"}, \code{"sum"}, \code{"median"}, \code{"min"}, \code{"max"}, \code{"count"}).

\strong{Default value:} \code{undefined} (None)

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/aggregate.html}{\code{aggregate}} documentation.}

\item{band}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) For rect-based marks (\code{rect}, \code{bar}, and \code{image}), mark size relative to bandwidth of \href{https://vega.github.io/vega-lite/docs/scale.html#band}{band scales}, bins or time units. If set to \code{1}, the mark size is set to the bandwidth, the bin interval, or the time unit interval. If set to \code{0.5}, the mark size is half of the bandwidth or the time unit interval.

For other marks, relative position on a band of a stacked, binned, time unit or band scale. If set to \code{0}, the marks will be positioned at the beginning of the band. If set to \code{0.5}, the marks will be positioned in the middle of the band.}

\item{bin}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) A flag for binning a \code{quantitative} field, \href{https://vega.github.io/vega-lite/docs/bin.html#params}{an object defining binning parameters}, or indicating that the data for \code{x} or \code{y} channel are binned before they are imported into Vega-Lite (\code{"binned"}).
\itemize{
\item If \code{true}, default \href{https://vega.github.io/vega-lite/docs/bin.html}{binning parameters} will be applied.
\item If \code{"binned"}, this indicates that the data for the \code{x} (or \code{y}) channel are already binned. You can map the bin-start field to \code{x} (or \code{y}) and the bin-end field to \code{x2} (or \code{y2}). The scale and axis will be formatted similar to binning in Vega-Lite.  To adjust the axis ticks based on the bin step, you can also set the axis's \href{https://vega.github.io/vega-lite/docs/axis.html#ticks}{\code{tickMinStep}} property.
}

\strong{Default value:} \code{false}

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/bin.html}{\code{bin}} documentation.}

\item{condition}{(\emph{ValueDefWithCondition<MarkPropFieldOrDatumDef,(string|null)>}) A field definition or one or more value definition(s) with a selection predicate.

(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) One or more value definition(s) with \href{https://vega.github.io/vega-lite/docs/condition.html}{a selection or a test predicate}.

\strong{Note:} A field definition's \code{condition} property can only contain \href{https://vega.github.io/vega-lite/docs/condition.html#value}{conditional value definitions} since Vega-Lite only allows at most one encoded field per encoding channel.}

\item{format}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) When used with the default \code{"number"} and \code{"time"} format type, the text formatting pattern for labels of guides (axes, legends, headers) and text marks.
\itemize{
\item If the format type is \code{"number"} (e.g., for quantitative fields), this is D3's \href{https://github.com/d3/d3-format#locale_format}{number format pattern}. - If the format type is \code{"time"} (e.g., for temporal fields), this is D3's \href{https://github.com/d3/d3-time-format#locale_format}{time format pattern}.
}

See the \href{https://vega.github.io/vega-lite/docs/format.html}{format documentation} for more examples.

When used with a \href{https://vega.github.io/vega-lite/docs/config.html#custom-format-type}{custom \code{formatType}}, this value will be passed as \code{format} alongside \code{datum.value} to the registered function.

\strong{Default value:}  Derived from \href{https://vega.github.io/vega-lite/docs/config.html#format}{numberFormat} config for number format and from \href{https://vega.github.io/vega-lite/docs/config.html#format}{timeFormat} config for time format.}

\item{formatType}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) The format type for labels. One of \code{"number"}, \code{"time"}, or a \href{https://vega.github.io/vega-lite/docs/config.html#custom-format-type}{registered custom format type}.

\strong{Default value:} - \code{"time"} for temporal fields and ordinal and nominal fields with \code{timeUnit}. - \code{"number"} for quantitative fields as well as ordinal and nominal fields without \code{timeUnit}.}

\item{labelExpr}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) \href{https://vega.github.io/vega/docs/expressions/}{Vega expression} for customizing labels text.

\strong{Note:} The label text and value can be assessed via the \code{label} and \code{value} properties of the axis's backing \code{datum} object.}

\item{timeUnit}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) Time unit (e.g., \code{year}, \code{yearmonth}, \code{month}, \code{hours}) for a temporal field. or \href{https://vega.github.io/vega-lite/docs/type.html#cast}{a temporal field that gets casted as ordinal}.

\strong{Default value:} \code{undefined} (None)

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/timeunit.html}{\code{timeUnit}} documentation.}

\item{title}{(\emph{FieldOrDatumDefWithCondition<StringFieldDef,string>}) A title for the field. If \code{null}, the title will be removed.

\strong{Default value:}  derived from the field's name and transformation function (\code{aggregate}, \code{bin} and \code{timeUnit}). If the field has an aggregate function, the function is displayed as part of the title (e.g., \code{"Sum of Profit"}). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., \code{"Profit (binned)"}, \code{"Transaction Date (year-month)"}). Otherwise, the title is simply the field name.

\strong{Notes}:
\enumerate{
\item You can customize the default field title format by providing the \href{https://vega.github.io/vega-lite/docs/config.html#top-level-config}{\code{fieldTitle}} property in the \href{https://vega.github.io/vega-lite/docs/config.html}{config} or \href{https://vega.github.io/vega-lite/docs/compile.html#field-title}{\code{fieldTitle} function via the \code{compile} function's options}.
\item If both field definition's \code{title} and axis, header, or legend \code{title} are defined, axis/header/legend title will be used.
}}

\item{value}{(\emph{ValueDefWithCondition<MarkPropFieldOrDatumDef,(string|null)>}) A constant value in visual domain (e.g., \code{"red"} / \code{"#0099ff"} / \href{https://vega.github.io/vega-lite/docs/types.html#gradient}{gradient definition} for color, values between \code{0} to \code{1} for opacity).}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add encoding for url to a vega-lite spec.
}
