% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_FoldTransform}
\alias{vl_FoldTransform}
\title{vl_FoldTransform}
\usage{
vl_FoldTransform(as = NULL, fold = NULL)
}
\arguments{
\item{as}{The output field names for the key and value properties produced by the fold transform.
\strong{Default value:} \code{["key", "value"]} (type = array)}

\item{fold}{An array of data fields indicating the properties to fold. (type = array)}
}
\value{
A modified spec
}
\description{
Create spec for FoldTransform.
}
