% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_joinaggregate}
\alias{vl_joinaggregate}
\title{vl_joinaggregate}
\usage{
vl_joinaggregate(spec, groupby = NULL, joinaggregate = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{groupby}{The data fields for partitioning the data objects into separate groups. If unspecified, all data points will be in a single group. (type = array)}

\item{joinaggregate}{The definition of the fields in the join aggregate, and what calculations to use. (type = array)}
}
\value{
A modified spec
}
\description{
Add JoinAggregateTransform to a vega-lite spec.
}
