% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_lookup}
\alias{vl_lookup}
\title{vl_lookup}
\usage{
vl_lookup(spec, as = NULL, default = NULL, from = NULL,
  lookup = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{as}{(\emph{LookupTransform}) The field or fields for storing the computed formula value.
If \code{from.fields} is specified, the transform will use the same names for \code{as}.
If \code{from.fields} is not specified, \code{as} has to be a string and we put the whole object into the data under the specified name.}

\item{default}{(\emph{LookupTransform}) The default value to use if lookup fails.

\strong{Default value:} \code{null}}

\item{from}{(\emph{LookupTransform}) Secondary data reference.}

\item{lookup}{(\emph{LookupTransform}) Key in primary data source.}
}
\value{
A modified spec
}
\description{
Add LookupTransform to a vega-lite spec.
}
