% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_LookupTransform}
\alias{vl_LookupTransform}
\title{vl_LookupTransform}
\usage{
vl_LookupTransform(as = NULL, default = NULL, from = NULL,
  lookup = NULL)
}
\arguments{
\item{as}{The field or fields for storing the computed formula value.
If \code{from.fields} is specified, the transform will use the same names for \code{as}.
If \code{from.fields} is not specified, \code{as} has to be a string and we put the whole object into the data under the specified name. (type = Varies)}

\item{default}{The default value to use if lookup fails.

\strong{Default value:} \code{null} (type = string)}

\item{from}{Secondary data reference. (type = LookupData)}

\item{lookup}{Key in primary data source. (type = string)}
}
\value{
A modified spec
}
\description{
Create spec for LookupTransform.
}
