% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_AggregateTransform}
\alias{vl_make_AggregateTransform}
\title{vl_make_AggregateTransform}
\usage{
vl_make_AggregateTransform(aggregate = NULL, groupby = NULL)
}
\arguments{
\item{aggregate}{(\emph{AggregateTransform}) Array of objects that define fields to aggregate.}

\item{groupby}{(\emph{AggregateTransform}) The data fields to group by. If not specified, a single group containing all data objects will be used.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Create spec for {obj}
}
