% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_BaseMarkConfig}
\alias{vl_make_BaseMarkConfig}
\title{vl_make_BaseMarkConfig}
\usage{
vl_make_BaseMarkConfig(align = NULL, angle = NULL, baseline = NULL,
  cornerRadius = NULL, cursor = NULL, dir = NULL, dx = NULL,
  dy = NULL, ellipsis = NULL, fill = NULL, fillOpacity = NULL,
  font = NULL, fontSize = NULL, fontStyle = NULL,
  fontWeight = NULL, href = NULL, interpolate = NULL, limit = NULL,
  opacity = NULL, orient = NULL, radius = NULL, shape = NULL,
  size = NULL, stroke = NULL, strokeCap = NULL, strokeDash = NULL,
  strokeDashOffset = NULL, strokeJoin = NULL,
  strokeMiterLimit = NULL, strokeOpacity = NULL, strokeWidth = NULL,
  tension = NULL, text = NULL, theta = NULL, tooltip = NULL,
  x = NULL, x2 = NULL, y = NULL, y2 = NULL)
}
\arguments{
\item{align}{(\emph{BaseMarkConfig}) The horizontal alignment of the text. One of \code{"left"}, \code{"right"}, \code{"center"}.}

\item{angle}{(\emph{BaseMarkConfig}) The rotation angle of the text, in degrees.}

\item{baseline}{(\emph{BaseMarkConfig}) The vertical alignment of the text. One of \code{"top"}, \code{"middle"}, \code{"bottom"}.

\strong{Default value:} \code{"middle"}}

\item{cornerRadius}{(\emph{BaseMarkConfig}) The radius in pixels of rounded rectangle corners.

\strong{Default value:} \code{0}}

\item{cursor}{(\emph{BaseMarkConfig}) The mouse cursor used over the mark. Any valid \href{https://developer.mozilla.org/en-US/docs/Web/CSS/cursor#Values}{CSS cursor type} can be used.}

\item{dir}{(\emph{BaseMarkConfig}) The direction of the text. One of \code{"ltr"} (left-to-right) or \code{"rtl"} (right-to-left). This property determines on which side is truncated in response to the limit parameter.

\strong{Default value:} \code{"ltr"}}

\item{dx}{(\emph{BaseMarkConfig}) The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the \emph{angle} property.}

\item{dy}{(\emph{BaseMarkConfig}) The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the \emph{angle} property.}

\item{ellipsis}{(\emph{BaseMarkConfig}) The ellipsis string for text truncated in response to the limit parameter.

\strong{Default value:} \code{"…"}}

\item{fill}{(\emph{BaseMarkConfig}) Default Fill Color.  This has higher precedence than \code{config.color}

\strong{Default value:} (None)}

\item{fillOpacity}{(\emph{BaseMarkConfig}) The fill opacity (value between [0,1]).

\strong{Default value:} \code{1}}

\item{font}{(\emph{BaseMarkConfig}) The typeface to set the text in (e.g., \code{"Helvetica Neue"}).}

\item{fontSize}{(\emph{BaseMarkConfig}) The font size, in pixels.}

\item{fontStyle}{(\emph{BaseMarkConfig}) The font style (e.g., \code{"italic"}).}

\item{fontWeight}{(\emph{BaseMarkConfig}) The font weight.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{href}{(\emph{BaseMarkConfig}) A URL to load upon mouse click. If defined, the mark acts as a hyperlink.}

\item{interpolate}{(\emph{BaseMarkConfig}) The line interpolation method to use for line and area marks. One of the following:
\itemize{
\item \code{"linear"}: piecewise linear segments, as in a polyline.
\item \code{"linear-closed"}: close the linear segments to form a polygon.
\item \code{"step"}: alternate between horizontal and vertical segments, as in a step function.
\item \code{"step-before"}: alternate between vertical and horizontal segments, as in a step function.
\item \code{"step-after"}: alternate between horizontal and vertical segments, as in a step function.
\item \code{"basis"}: a B-spline, with control point duplication on the ends.
\item \code{"basis-open"}: an open B-spline; may not intersect the start or end.
\item \code{"basis-closed"}: a closed B-spline, as in a loop.
\item \code{"cardinal"}: a Cardinal spline, with control point duplication on the ends.
\item \code{"cardinal-open"}: an open Cardinal spline; may not intersect the start or end, but will intersect other control points.
\item \code{"cardinal-closed"}: a closed Cardinal spline, as in a loop.
\item \code{"bundle"}: equivalent to basis, except the tension parameter is used to straighten the spline.
\item \code{"monotone"}: cubic interpolation that preserves monotonicity in y.
}}

\item{limit}{(\emph{BaseMarkConfig}) The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

\strong{Default value:} \code{0}, indicating no limit}

\item{opacity}{(\emph{BaseMarkConfig}) The overall opacity (value between [0,1]).

\strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick}, \code{circle}, or \code{square} marks or layered \code{bar} charts and \code{1} otherwise.}

\item{orient}{(\emph{BaseMarkConfig}) The orientation of a non-stacked bar, tick, area, and line charts.
The value is either horizontal (default) or vertical.
\itemize{
\item For bar, rule and tick, this determines whether the size of the bar and tick
should be applied to x or y dimension.
\item For area, this property determines the orient property of the Vega output.
\item For line and trail marks, this property determines the sort order of the points in the line
if \code{config.sortLineBy} is not specified.
For stacked charts, this is always determined by the orientation of the stack;
therefore explicitly specified value will be ignored.
}}

\item{radius}{(\emph{BaseMarkConfig}) Polar coordinate radial offset, in pixels, of the text label from the origin determined by the \code{x} and \code{y} properties.}

\item{shape}{(\emph{BaseMarkConfig}) Shape of the point marks. Supported values include:
\itemize{
\item plotting shapes: \code{"circle"}, \code{"square"}, \code{"cross"}, \code{"diamond"}, \code{"triangle-up"}, \code{"triangle-down"}, \code{"triangle-right"}, or \code{"triangle-left"}.
\item the line symbol \code{"stroke"}
\item centered directional shapes \code{"arrow"}, \code{"wedge"}, or \code{"triangle"}
\item a custom \href{https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths}{SVG path string} (For correct sizing, custom shape paths should be defined within a square bounding box with coordinates ranging from -1 to 1 along both the x and y dimensions.)
}

\strong{Default value:} \code{"circle"}}

\item{size}{(\emph{BaseMarkConfig}) The pixel area each the point/circle/square.
For example: in the case of circles, the radius is determined in part by the square root of the size value.

\strong{Default value:} \code{30}}

\item{stroke}{(\emph{BaseMarkConfig}) Default Stroke Color.  This has higher precedence than \code{config.color}

\strong{Default value:} (None)}

\item{strokeCap}{(\emph{BaseMarkConfig}) The stroke cap for line ending style. One of \code{"butt"}, \code{"round"}, or \code{"square"}.

\strong{Default value:} \code{"square"}}

\item{strokeDash}{(\emph{BaseMarkConfig}) An array of alternating stroke, space lengths for creating dashed or dotted lines.}

\item{strokeDashOffset}{(\emph{BaseMarkConfig}) The offset (in pixels) into which to begin drawing with the stroke dash array.}

\item{strokeJoin}{(\emph{BaseMarkConfig}) The stroke line join method. One of \code{"miter"}, \code{"round"} or \code{"bevel"}.

\strong{Default value:} \code{"miter"}}

\item{strokeMiterLimit}{(\emph{BaseMarkConfig}) The miter limit at which to bevel a line join.}

\item{strokeOpacity}{(\emph{BaseMarkConfig}) The stroke opacity (value between [0,1]).

\strong{Default value:} \code{1}}

\item{strokeWidth}{(\emph{BaseMarkConfig}) The stroke width, in pixels.}

\item{tension}{(\emph{BaseMarkConfig}) Depending on the interpolation type, sets the tension parameter (for line and area marks).}

\item{text}{(\emph{BaseMarkConfig}) Placeholder text if the \code{text} channel is not specified}

\item{theta}{(\emph{BaseMarkConfig}) Polar coordinate angle, in radians, of the text label from the origin determined by the \code{x} and \code{y} properties. Values for \code{theta} follow the same convention of \code{arc} mark \code{startAngle} and \code{endAngle} properties: angles are measured in radians, with \code{0} indicating "north".}

\item{tooltip}{(\emph{BaseMarkConfig}) The tooltip text to show upon mouse hover.}

\item{x}{(\emph{BaseMarkConfig}) X coordinates of the marks, or width of horizontal \code{"bar"} and \code{"area"} without \code{x2}.}

\item{x2}{(\emph{BaseMarkConfig}) X2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and  \code{"rule"}.}

\item{y}{(\emph{BaseMarkConfig}) Y coordinates of the marks, or height of vertical \code{"bar"} and \code{"area"} without \code{y2}}

\item{y2}{(\emph{BaseMarkConfig}) Y2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and  \code{"rule"}.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Create spec for {obj}
}
