% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_BaseTitleConfig}
\alias{vl_make_BaseTitleConfig}
\title{vl_make_BaseTitleConfig}
\usage{
vl_make_BaseTitleConfig(align = NULL, anchor = NULL, angle = NULL,
  baseline = NULL, color = NULL, dx = NULL, dy = NULL,
  font = NULL, fontSize = NULL, fontStyle = NULL,
  fontWeight = NULL, frame = NULL, limit = NULL, offset = NULL,
  orient = NULL)
}
\arguments{
\item{align}{(\emph{BaseTitleConfig})}

\item{anchor}{(\emph{BaseTitleConfig}) The anchor position for placing the title. One of \code{"start"}, \code{"middle"}, or \code{"end"}. For example, with an orientation of top these anchor positions map to a left-, center-, or right-aligned title.}

\item{angle}{(\emph{BaseTitleConfig}) Angle in degrees of title text.}

\item{baseline}{(\emph{BaseTitleConfig}) Vertical text baseline for title text. One of \code{"top"}, \code{"middle"}, \code{"bottom"}, or \code{"alphabetic"}.}

\item{color}{(\emph{BaseTitleConfig}) Text color for title text.}

\item{dx}{(\emph{BaseTitleConfig}) Delta offset for title text x-coordinate.}

\item{dy}{(\emph{BaseTitleConfig}) Delta offset for title text y-coordinate.}

\item{font}{(\emph{BaseTitleConfig}) Font name for title text.}

\item{fontSize}{(\emph{BaseTitleConfig}) Font size in pixels for title text.

\strong{Default value:} \code{10}.}

\item{fontStyle}{(\emph{BaseTitleConfig}) Font style for title text.}

\item{fontWeight}{(\emph{BaseTitleConfig}) Font weight for title text.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{frame}{(\emph{BaseTitleConfig}) The reference frame for the anchor position, one of \code{"bounds"} (to anchor relative to the full bounding box) or \code{"group"} (to anchor relative to the group width or height).}

\item{limit}{(\emph{BaseTitleConfig}) The maximum allowed length in pixels of legend labels.}

\item{offset}{(\emph{BaseTitleConfig}) The orthogonal offset in pixels by which to displace the title from its position along the edge of the chart.}

\item{orient}{(\emph{BaseTitleConfig}) Default title orientation (\code{"top"}, \code{"bottom"}, \code{"left"}, or \code{"right"})}
}
\value{
A component of a Vega-Lite spec, corresponding to a BaseTitleConfig.
}
\description{
Create spec for {obj}
}
