% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_BindCheckbox}
\alias{vl_make_BindCheckbox}
\title{vl_make_BindCheckbox}
\usage{
vl_make_BindCheckbox(debounce = NULL, element = NULL, input = NULL,
  name = NULL, type = NULL)
}
\arguments{
\item{debounce}{(\emph{BindCheckbox})}

\item{element}{(\emph{BindCheckbox})}

\item{input}{(\emph{BindCheckbox}) checkbox}

\item{name}{(\emph{BindCheckbox})}

\item{type}{(\emph{BindCheckbox})}
}
\value{
A modified spec
}
\description{
Create spec for BindCheckbox.
}
