% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_BindRadioSelect}
\alias{vl_make_BindRadioSelect}
\title{vl_make_BindRadioSelect}
\usage{
vl_make_BindRadioSelect(debounce = NULL, element = NULL,
  input = NULL, name = NULL, options = NULL, type = NULL)
}
\arguments{
\item{debounce}{(\emph{BindRadioSelect})}

\item{element}{(\emph{BindRadioSelect})}

\item{input}{(\emph{BindRadioSelect}) radio, select}

\item{name}{(\emph{BindRadioSelect})}

\item{options}{(\emph{BindRadioSelect})}

\item{type}{(\emph{BindRadioSelect})}
}
\value{
A modified Vega-Lite Spec
}
\description{
Create spec for {obj}
}
