% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_BindRange}
\alias{vl_make_BindRange}
\title{vl_make_BindRange}
\usage{
vl_make_BindRange(debounce = NULL, element = NULL, input = NULL,
  max = NULL, min = NULL, name = NULL, step = NULL, type = NULL)
}
\arguments{
\item{debounce}{(\emph{BindRange})}

\item{element}{(\emph{BindRange})}

\item{input}{(\emph{BindRange}) range}

\item{max}{(\emph{BindRange})}

\item{min}{(\emph{BindRange})}

\item{name}{(\emph{BindRange})}

\item{step}{(\emph{BindRange})}

\item{type}{(\emph{BindRange})}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BindRange}
}
\value{
A component of a Vega-Lite spec, corresponding to a BindRange.
}
\description{
Create spec for {obj}
}
