% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_BrushConfig}
\alias{vl_make_BrushConfig}
\title{vl_make_BrushConfig}
\usage{
vl_make_BrushConfig(fill = NULL, fillOpacity = NULL, stroke = NULL,
  strokeDash = NULL, strokeDashOffset = NULL, strokeOpacity = NULL,
  strokeWidth = NULL)
}
\arguments{
\item{fill}{(\emph{BrushConfig}) The fill color of the interval mark.

\strong{Default value:} \code{#333333}}

\item{fillOpacity}{(\emph{BrushConfig}) The fill opacity of the interval mark (a value between 0 and 1).

\strong{Default value:} \code{0.125}}

\item{stroke}{(\emph{BrushConfig}) The stroke color of the interval mark.

\strong{Default value:} \code{#ffffff}}

\item{strokeDash}{(\emph{BrushConfig}) An array of alternating stroke and space lengths,
for creating dashed or dotted lines.}

\item{strokeDashOffset}{(\emph{BrushConfig}) The offset (in pixels) with which to begin drawing the stroke dash array.}

\item{strokeOpacity}{(\emph{BrushConfig}) The stroke opacity of the interval mark (a value between 0 and 1).}

\item{strokeWidth}{(\emph{BrushConfig}) The stroke width of the interval mark.}
}
\value{
A component of a Vega-Lite spec, corresponding to a BrushConfig.
}
\description{
Create spec for {obj}
}
