% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_CalculateTransform}
\alias{vl_make_CalculateTransform}
\title{vl_make_CalculateTransform}
\usage{
vl_make_CalculateTransform(as = NULL, calculate = NULL)
}
\arguments{
\item{as}{(\emph{CalculateTransform}) The field for storing the computed formula value.}

\item{calculate}{(\emph{CalculateTransform}) A \href{https://vega.github.io/vega-lite/docs/types.html#expression}{expression} string. Use the variable \code{datum} to refer to the current data object.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: CalculateTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a CalculateTransform.
}
\description{
Create spec for {obj}
}
