% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_FilterTransform}
\alias{vl_make_FilterTransform}
\title{vl_make_FilterTransform}
\usage{
vl_make_FilterTransform(filter = NULL)
}
\arguments{
\item{filter}{(\emph{FilterTransform}) The \code{filter} property must be one of the predicate definitions:
\enumerate{
\item an \href{https://vega.github.io/vega-lite/docs/types.html#expression}{expression} string,
where \code{datum} can be used to refer to the current data object
\item one of the field predicates: \href{https://vega.github.io/vega-lite/docs/filter.html#equal-predicate}{equal},
\href{https://vega.github.io/vega-lite/docs/filter.html#lt-predicate}{lt},
\href{https://vega.github.io/vega-lite/docs/filter.html#lte-predicate}{lte},
\href{https://vega.github.io/vega-lite/docs/filter.html#gt-predicate}{gt},
\href{https://vega.github.io/vega-lite/docs/filter.html#gte-predicate}{gte},
\href{https://vega.github.io/vega-lite/docs/filter.html#range-predicate}{range},
\href{https://vega.github.io/vega-lite/docs/filter.html#one-of-predicate}{oneOf},
or \href{https://vega.github.io/vega-lite/docs/filter.html#valid-predicate}{valid},
\item a \href{https://vega.github.io/vega-lite/docs/filter.html#selection-predicate}{selection predicate}
\item a logical operand that combines (1), (2), or (3).
}}
}
\value{
A component of a Vega-Lite spec, corresponding to a FilterTransform.
}
\description{
Create spec for {obj}
}
