% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_FlattenTransform}
\alias{vl_make_FlattenTransform}
\title{vl_make_FlattenTransform}
\usage{
vl_make_FlattenTransform(as = NULL, flatten = NULL)
}
\arguments{
\item{as}{(\emph{FlattenTransform}) The output field names for extracted array values.

\strong{Default value:} The field name of the corresponding array field}

\item{flatten}{(\emph{FlattenTransform}) An array of one or more data fields containing arrays to flatten.
If multiple fields are specified, their array values should have a parallel structure, ideally with the same length.
If the lengths of parallel arrays do not match,
the longest array will be used with \code{null} values added for missing entries.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: FlattenTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a FlattenTransform.
}
\description{
Create spec for {obj}
}
