% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_FoldTransform}
\alias{vl_make_FoldTransform}
\title{vl_make_FoldTransform}
\usage{
vl_make_FoldTransform(as = NULL, fold = NULL)
}
\arguments{
\item{as}{(\emph{FoldTransform}) The output field names for the key and value properties produced by the fold transform.
\strong{Default value:} \code{["key", "value"]}}

\item{fold}{(\emph{FoldTransform}) An array of data fields indicating the properties to fold.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: FoldTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a FoldTransform.
}
\description{
Create spec for {obj}
}
