% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_LoessTransform}
\alias{vl_make_LoessTransform}
\title{vl_make_LoessTransform}
\usage{
vl_make_LoessTransform(as = NULL, bandwidth = NULL, groupby = NULL,
  loess = NULL, on = NULL)
}
\arguments{
\item{as}{(\emph{LoessTransform}) The output field names for the smoothed points generated by the loess transform.

\strong{Default value:} The field names of the input x and y values.}

\item{bandwidth}{(\emph{LoessTransform}) A bandwidth parameter in the range \code{[0, 1]} that determines the amount of smoothing.

\strong{Default value:} \code{0.3}}

\item{groupby}{(\emph{LoessTransform}) The data fields to group by. If not specified, a single group containing all data objects will be used.}

\item{loess}{(\emph{LoessTransform}) The data field of the dependent variable to smooth.}

\item{on}{(\emph{LoessTransform}) The data field of the independent variable to use a predictor.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: LoessTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a LoessTransform.
}
\description{
Create spec for {obj}
}
