% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_LookupTransform}
\alias{vl_make_LookupTransform}
\title{vl_make_LookupTransform}
\usage{
vl_make_LookupTransform(as = NULL, default = NULL, from = NULL,
  lookup = NULL)
}
\arguments{
\item{as}{(\emph{LookupTransform}) The field or fields for storing the computed formula value.
If \code{from.fields} is specified, the transform will use the same names for \code{as}.
If \code{from.fields} is not specified, \code{as} has to be a string and we put the whole object into the data under the specified name.}

\item{default}{(\emph{LookupTransform}) The default value to use if lookup fails.

\strong{Default value:} \code{null}}

\item{from}{(\emph{LookupTransform}) Secondary data reference.}

\item{lookup}{(\emph{LookupTransform}) Key in primary data source.}
}
\value{
A component of a Vega-Lite spec, corresponding to a LookupTransform.
}
\description{
Create spec for {obj}
}
