% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_QuantileTransform}
\alias{vl_make_QuantileTransform}
\title{vl_make_QuantileTransform}
\usage{
vl_make_QuantileTransform(as = NULL, groupby = NULL, probs = NULL,
  quantile = NULL, step = NULL)
}
\arguments{
\item{as}{(\emph{QuantileTransform}) The output field names for the probability and quantile values.

\strong{Default value:} \code{["prob", "value"]}}

\item{groupby}{(\emph{QuantileTransform}) The data fields to group by. If not specified, a single group containing all data objects will be used.}

\item{probs}{(\emph{QuantileTransform}) An array of probabilities in the range (0, 1) for which to compute quantile values. If not specified, the \emph{step} parameter will be used.}

\item{quantile}{(\emph{QuantileTransform}) The data field for which to perform quantile estimation.}

\item{step}{(\emph{QuantileTransform}) A probability step size (default 0.01) for sampling quantile values. All values from one-half the step size up to 1 (exclusive) will be sampled. This parameter is only used if the \emph{probs} parameter is not provided.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: QuantileTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a QuantileTransform.
}
\description{
Create spec for {obj}
}
