% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_RangeConfig}
\alias{vl_make_RangeConfig}
\title{vl_make_RangeConfig}
\usage{
vl_make_RangeConfig(category = NULL, diverging = NULL,
  heatmap = NULL, ordinal = NULL, ramp = NULL, symbol = NULL)
}
\arguments{
\item{category}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for categorical data.}

\item{diverging}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for diverging quantitative ramps.}

\item{heatmap}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for quantitative heatmaps.}

\item{ordinal}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for rank-ordered data.}

\item{ramp}{(\emph{RangeConfig}) Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for sequential quantitative ramps.}

\item{symbol}{(\emph{RangeConfig}) Array of \href{https://vega.github.io/vega/docs/marks/symbol/}{symbol} names or paths for the default shape palette.}
}
\value{
A component of a Vega-Lite spec, corresponding to a RangeConfig.
}
\description{
Create spec for {obj}
}
