% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_RegressionTransform}
\alias{vl_make_RegressionTransform}
\title{vl_make_RegressionTransform}
\usage{
vl_make_RegressionTransform(as = NULL, extent = NULL, groupby = NULL,
  method = NULL, on = NULL, order = NULL, params = NULL,
  regression = NULL)
}
\arguments{
\item{as}{(\emph{RegressionTransform}) The output field names for the smoothed points generated by the regression transform.

\strong{Default value:} The field names of the input x and y values.}

\item{extent}{(\emph{RegressionTransform}) A [min, max] domain over the independent (x) field for the starting and ending points of the generated trend line.}

\item{groupby}{(\emph{RegressionTransform}) The data fields to group by. If not specified, a single group containing all data objects will be used.}

\item{method}{(\emph{RegressionTransform}) The functional form of the regression model. One of \code{"linear"}, \code{"log"}, \code{"exp"}, \code{"pow"}, \code{"quad"}, or \code{"poly"}.

\strong{Default value:} \code{"linear"}}

\item{on}{(\emph{RegressionTransform}) The data field of the independent variable to use a predictor.}

\item{order}{(\emph{RegressionTransform}) The polynomial order (number of coefficients) for the 'poly' method.

\strong{Default value:} \code{3}}

\item{params}{(\emph{RegressionTransform}) A boolean flag indicating if the transform should return the regression model parameters (one object per group), rather than trend line points.
The resulting objects include a \code{coef} array of fitted coefficient values (starting with the intercept term and then including terms of increasing order)
and an \code{rSquared} value (indicating the total variance explained by the model).

\strong{Default value:} \code{false}}

\item{regression}{(\emph{RegressionTransform}) The data field of the dependent variable to predict.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: RegressionTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a RegressionTransform.
}
\description{
Create spec for {obj}
}
