% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_SampleTransform}
\alias{vl_make_SampleTransform}
\title{vl_make_SampleTransform}
\usage{
vl_make_SampleTransform(sample = NULL)
}
\arguments{
\item{sample}{(\emph{SampleTransform}) The maximum number of data objects to include in the sample.

\strong{Default value:} \code{1000}}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: SampleTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a SampleTransform.
}
\description{
Create spec for {obj}
}
