% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_SchemeConfig}
\alias{vl_make_SchemeConfig}
\title{vl_make_SchemeConfig}
\usage{
vl_make_SchemeConfig(count = NULL, extent = NULL, scheme = NULL)
}
\arguments{
\item{count}{(\emph{SchemeConfig})}

\item{extent}{(\emph{SchemeConfig})}

\item{scheme}{(\emph{SchemeConfig})}
}
\value{
A modified spec
}
\description{
Create spec for SchemeConfig.
}
