% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_StackTransform}
\alias{vl_make_StackTransform}
\title{vl_make_StackTransform}
\usage{
vl_make_StackTransform(as = NULL, groupby = NULL, offset = NULL,
  sort = NULL, stack = NULL)
}
\arguments{
\item{as}{(\emph{StackTransform}) Output field names. This can be either a string or an array of strings with
two elements denoting the name for the fields for stack start and stack end
respectively.
If a single string(eg."val") is provided, the end field will be "val_end".}

\item{groupby}{(\emph{StackTransform}) The data fields to group by.}

\item{offset}{(\emph{StackTransform}) Mode for stacking marks.
\strong{Default value:} \code{"zero"}}

\item{sort}{(\emph{StackTransform}) Field that determines the order of leaves in the stacked charts.}

\item{stack}{(\emph{StackTransform}) The field which is stacked.}
}
\value{
A component of a Vega-Lite spec, corresponding to a StackTransform.
}
\description{
Create spec for {obj}
}
