% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_TimeUnitTransform}
\alias{vl_make_TimeUnitTransform}
\title{vl_make_TimeUnitTransform}
\usage{
vl_make_TimeUnitTransform(as = NULL, field = NULL, timeUnit = NULL)
}
\arguments{
\item{as}{(\emph{TimeUnitTransform}) The output field to write the timeUnit value.}

\item{field}{(\emph{TimeUnitTransform}) The data field to apply time unit.}

\item{timeUnit}{(\emph{TimeUnitTransform}) The timeUnit.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: TimeUnitTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a TimeUnitTransform.
}
\description{
Create spec for {obj}
}
