% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_TitleConfig}
\alias{vl_make_TitleConfig}
\title{vl_make_TitleConfig}
\usage{
vl_make_TitleConfig(align = NULL, anchor = NULL, angle = NULL,
  baseline = NULL, color = NULL, dx = NULL, dy = NULL,
  font = NULL, fontSize = NULL, fontStyle = NULL,
  fontWeight = NULL, frame = NULL, limit = NULL, lineHeight = NULL,
  offset = NULL, orient = NULL, subtitleColor = NULL,
  subtitleFont = NULL, subtitleFontSize = NULL,
  subtitleFontStyle = NULL, subtitleFontWeight = NULL,
  subtitleLineHeight = NULL, subtitlePadding = NULL)
}
\arguments{
\item{align}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Horizontal text alignment for title text. One of \code{"left"}, \code{"center"}, or \code{"right"}.}

\item{anchor}{(\emph{ExcludeMappedValueRef<BaseTitle>}) The anchor position for placing the title and subtitle text. One of \code{"start"}, \code{"middle"}, or \code{"end"}. For example, with an orientation of top these anchor positions map to a left-, center-, or right-aligned title.}

\item{angle}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Angle in degrees of title and subtitle text.}

\item{baseline}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Vertical text baseline for title and subtitle text. One of \code{"top"}, \code{"middle"}, \code{"bottom"}, or \code{"alphabetic"}.}

\item{color}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Text color for title text.}

\item{dx}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Delta offset for title and subtitle text x-coordinate.}

\item{dy}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Delta offset for title and subtitle text y-coordinate.}

\item{font}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font name for title text.}

\item{fontSize}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font size in pixels for title text.}

\item{fontStyle}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font style for title text.}

\item{fontWeight}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font weight for title text.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{frame}{(\emph{ExcludeMappedValueRef<BaseTitle>}) The reference frame for the anchor position, one of \code{"bounds"} (to anchor relative to the full bounding box) or \code{"group"} (to anchor relative to the group width or height).}

\item{limit}{(\emph{ExcludeMappedValueRef<BaseTitle>}) The maximum allowed length in pixels of title and subtitle text.}

\item{lineHeight}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Line height in pixels for multi-line title text.}

\item{offset}{(\emph{ExcludeMappedValueRef<BaseTitle>}) The orthogonal offset in pixels by which to displace the title group from its position along the edge of the chart.}

\item{orient}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Default title orientation (\code{"top"}, \code{"bottom"}, \code{"left"}, or \code{"right"})}

\item{subtitleColor}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Text color for subtitle text.}

\item{subtitleFont}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font name for subtitle text.}

\item{subtitleFontSize}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font size in pixels for subtitle text.}

\item{subtitleFontStyle}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font style for subtitle text.}

\item{subtitleFontWeight}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Font weight for subtitle text.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{subtitleLineHeight}{(\emph{ExcludeMappedValueRef<BaseTitle>}) Line height in pixels for multi-line subtitle text.}

\item{subtitlePadding}{(\emph{ExcludeMappedValueRef<BaseTitle>}) The padding in pixels between title and subtitle text.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: ExcludeMappedValueRef<BaseTitle>}
}
\value{
A component of a Vega-Lite spec, corresponding to a TitleConfig.
}
\description{
Create spec for {obj}
}
