% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_mark_rect}
\alias{vl_mark_rect}
\title{vl_mark_rect}
\usage{
vl_mark_rect(
  spec,
  .object = NULL,
  align = NULL,
  angle = NULL,
  aria = NULL,
  ariaRole = NULL,
  ariaRoleDescription = NULL,
  aspect = NULL,
  bandSize = NULL,
  baseline = NULL,
  binSpacing = NULL,
  blend = NULL,
  clip = NULL,
  color = NULL,
  continuousBandSize = NULL,
  cornerRadius = NULL,
  cornerRadiusBottomLeft = NULL,
  cornerRadiusBottomRight = NULL,
  cornerRadiusEnd = NULL,
  cornerRadiusTopLeft = NULL,
  cornerRadiusTopRight = NULL,
  cursor = NULL,
  description = NULL,
  dir = NULL,
  discreteBandSize = NULL,
  dx = NULL,
  dy = NULL,
  ellipsis = NULL,
  fill = NULL,
  fillOpacity = NULL,
  filled = NULL,
  font = NULL,
  fontSize = NULL,
  fontStyle = NULL,
  fontWeight = NULL,
  height = NULL,
  href = NULL,
  innerRadius = NULL,
  interpolate = NULL,
  invalid = NULL,
  limit = NULL,
  line = NULL,
  lineBreak = NULL,
  lineHeight = NULL,
  opacity = NULL,
  order = NULL,
  orient = NULL,
  outerRadius = NULL,
  padAngle = NULL,
  point = NULL,
  radius = NULL,
  radius2 = NULL,
  radius2Offset = NULL,
  radiusOffset = NULL,
  shape = NULL,
  size = NULL,
  smooth = NULL,
  stroke = NULL,
  strokeCap = NULL,
  strokeDash = NULL,
  strokeDashOffset = NULL,
  strokeJoin = NULL,
  strokeMiterLimit = NULL,
  strokeOffset = NULL,
  strokeOpacity = NULL,
  strokeWidth = NULL,
  style = NULL,
  tension = NULL,
  text = NULL,
  theta = NULL,
  theta2 = NULL,
  theta2Offset = NULL,
  thetaOffset = NULL,
  thickness = NULL,
  timeUnitBand = NULL,
  timeUnitBandPosition = NULL,
  tooltip = NULL,
  url = NULL,
  width = NULL,
  x = NULL,
  x2 = NULL,
  x2Offset = NULL,
  xOffset = NULL,
  y = NULL,
  y2 = NULL,
  y2Offset = NULL,
  yOffset = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: MarkDef}

\item{align}{(\emph{MarkDef}) The horizontal alignment of the text or ranged marks (area, bar, image, rect, rule). One of \code{"left"}, \code{"right"}, \code{"center"}.

\strong{Note:} Expression reference is \emph{not} supported for range marks.}

\item{angle}{(\emph{MarkDef})}

\item{aria}{(\emph{MarkDef})}

\item{ariaRole}{(\emph{MarkDef})}

\item{ariaRoleDescription}{(\emph{MarkDef})}

\item{aspect}{(\emph{MarkDef})}

\item{bandSize}{(\emph{MarkDef}) The width of the ticks.

\strong{Default value:}  3/4 of step (width step for horizontal ticks and height step for vertical ticks).}

\item{baseline}{(\emph{MarkDef}) For text marks, the vertical text baseline. One of \code{"alphabetic"} (default), \code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, \code{"line-bottom"}, or an expression reference that provides one of the valid values. The \code{"line-top"} and \code{"line-bottom"} values operate similarly to \code{"top"} and \code{"bottom"}, but are calculated relative to the \code{lineHeight} rather than \code{fontSize} alone.

For range marks, the vertical alignment of the marks. One of \code{"top"}, \code{"middle"}, \code{"bottom"}.

\strong{Note:} Expression reference is \emph{not} supported for range marks.}

\item{binSpacing}{(\emph{MarkDef}) Offset between bars for binned field. The ideal value for this is either 0 (preferred by statisticians) or 1 (Vega-Lite default, D3 example style).

\strong{Default value:} \code{1}}

\item{blend}{(\emph{MarkDef})}

\item{clip}{(\emph{MarkDef}) Whether a mark be clipped to the enclosing group’s width and height.}

\item{color}{(\emph{MarkDef}) Default color.

\strong{Default value:} <span style="color: #4682b4;">■</span> \code{"#4682b4"}

\strong{Note:} - This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}. - The \code{fill} and \code{stroke} properties have higher precedence than \code{color} and will override \code{color}.}

\item{continuousBandSize}{(\emph{MarkDef}) The default size of the bars on continuous scales.

\strong{Default value:} \code{5}}

\item{cornerRadius}{(\emph{MarkDef})}

\item{cornerRadiusBottomLeft}{(\emph{MarkDef})}

\item{cornerRadiusBottomRight}{(\emph{MarkDef})}

\item{cornerRadiusEnd}{(\emph{MarkDef}) - For vertical bars, top-left and top-right corner radius. - For horizontal bars, top-right and bottom-right corner radius.}

\item{cornerRadiusTopLeft}{(\emph{MarkDef})}

\item{cornerRadiusTopRight}{(\emph{MarkDef})}

\item{cursor}{(\emph{MarkDef})}

\item{description}{(\emph{MarkDef})}

\item{dir}{(\emph{MarkDef})}

\item{discreteBandSize}{(\emph{MarkDef}) The default size of the bars with discrete dimensions. If unspecified, the default size is  \code{step-2}, which provides 2 pixel offset between bars.}

\item{dx}{(\emph{MarkDef})}

\item{dy}{(\emph{MarkDef})}

\item{ellipsis}{(\emph{MarkDef})}

\item{fill}{(\emph{MarkDef}) Default fill color. This property has higher precedence than \code{config.color}. Set to \code{null} to remove fill.

\strong{Default value:} (None)}

\item{fillOpacity}{(\emph{MarkDef})}

\item{filled}{(\emph{MarkDef}) Whether the mark's color should be used as fill color instead of stroke color.

\strong{Default value:} \code{false} for all \code{point}, \code{line}, and \code{rule} marks as well as \code{geoshape} marks for \href{https://vega.github.io/vega-lite/docs/data.html#graticule}{\code{graticule}} data sources; otherwise, \code{true}.

\strong{Note:} This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}.}

\item{font}{(\emph{MarkDef})}

\item{fontSize}{(\emph{MarkDef})}

\item{fontStyle}{(\emph{MarkDef})}

\item{fontWeight}{(\emph{MarkDef})}

\item{height}{(\emph{MarkDef})}

\item{href}{(\emph{MarkDef})}

\item{innerRadius}{(\emph{MarkDef}) The inner radius in pixels of arc marks. \code{innerRadius} is an alias for \code{radius2}.}

\item{interpolate}{(\emph{MarkDef})}

\item{invalid}{(\emph{MarkDef}) Defines how Vega-Lite should handle marks for invalid values (\code{null} and \code{NaN}). - If set to \code{"filter"} (default), all data items with null values will be skipped (for line, trail, and area marks) or filtered (for other marks). - If \code{null}, all data items are included. In this case, invalid values will be interpreted as zeroes.}

\item{limit}{(\emph{MarkDef})}

\item{line}{(\emph{MarkDef}) A flag for overlaying line on top of area marks, or an object defining the properties of the overlayed lines.
\itemize{
\item If this value is an empty object (\code{{}}) or \code{true}, lines with default properties will be used.
\item If this value is \code{false}, no lines would be automatically added to area marks.
}

\strong{Default value:} \code{false}.}

\item{lineBreak}{(\emph{MarkDef})}

\item{lineHeight}{(\emph{MarkDef})}

\item{opacity}{(\emph{MarkDef}) The overall opacity (value between [0,1]).

\strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick}, \code{circle}, or \code{square} marks or layered \code{bar} charts and \code{1} otherwise.}

\item{order}{(\emph{MarkDef}) For line and trail marks, this \code{order} property can be set to \code{null} or \code{false} to make the lines use the original order in the data sources.}

\item{orient}{(\emph{MarkDef}) The orientation of a non-stacked bar, tick, area, and line charts. The value is either horizontal (default) or vertical. - For bar, rule and tick, this determines whether the size of the bar and tick should be applied to x or y dimension. - For area, this property determines the orient property of the Vega output. - For line and trail marks, this property determines the sort order of the points in the line if \code{config.sortLineBy} is not specified. For stacked charts, this is always determined by the orientation of the stack; therefore explicitly specified value will be ignored.}

\item{outerRadius}{(\emph{MarkDef}) The outer radius in pixels of arc marks. \code{outerRadius} is an alias for \code{radius}.}

\item{padAngle}{(\emph{MarkDef})}

\item{point}{(\emph{MarkDef}) A flag for overlaying points on top of line or area marks, or an object defining the properties of the overlayed points.
\itemize{
\item If this property is \code{"transparent"}, transparent points will be used (for enhancing tooltips and selections).
\item If this property is an empty object (\code{{}}) or \code{true}, filled points with default properties will be used.
\item If this property is \code{false}, no points would be automatically added to line or area marks.
}

\strong{Default value:} \code{false}.}

\item{radius}{(\emph{MarkDef}) For arc mark, the primary (outer) radius in pixels.

For text marks, polar coordinate radial offset, in pixels, of the text from the origin determined by the \code{x} and \code{y} properties.}

\item{radius2}{(\emph{MarkDef}) The secondary (inner) radius in pixels of arc marks.}

\item{radius2Offset}{(\emph{MarkDef}) Offset for radius2.}

\item{radiusOffset}{(\emph{MarkDef}) Offset for radius.}

\item{shape}{(\emph{MarkDef})}

\item{size}{(\emph{MarkDef}) Default size for marks. - For \code{point}/\code{circle}/\code{square}, this represents the pixel area of the marks. Note that this value sets the area of the symbol; the side lengths will increase with the square root of this value. - For \code{bar}, this represents the band size of the bar, in pixels. - For \code{text}, this represents the font size, in pixels.

\strong{Default value:} - \code{30} for point, circle, square marks; width/height's \code{step} - \code{2} for bar marks with discrete dimensions; - \code{5} for bar marks with continuous dimensions; - \code{11} for text marks.}

\item{smooth}{(\emph{MarkDef})}

\item{stroke}{(\emph{MarkDef}) Default stroke color. This property has higher precedence than \code{config.color}. Set to \code{null} to remove stroke.

\strong{Default value:} (None)}

\item{strokeCap}{(\emph{MarkDef})}

\item{strokeDash}{(\emph{MarkDef})}

\item{strokeDashOffset}{(\emph{MarkDef})}

\item{strokeJoin}{(\emph{MarkDef})}

\item{strokeMiterLimit}{(\emph{MarkDef})}

\item{strokeOffset}{(\emph{MarkDef})}

\item{strokeOpacity}{(\emph{MarkDef})}

\item{strokeWidth}{(\emph{MarkDef})}

\item{style}{(\emph{MarkDef}) A string or array of strings indicating the name of custom styles to apply to the mark. A style is a named collection of mark property defaults defined within the \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style configuration}. If style is an array, later styles will override earlier styles. Any \href{https://vega.github.io/vega-lite/docs/encoding.html#mark-prop}{mark properties} explicitly defined within the \code{encoding} will override a style default.

\strong{Default value:} The mark's name. For example, a bar mark will have style \code{"bar"} by default. \strong{Note:} Any specified style will augment the default style. For example, a bar mark with \code{"style": "foo"} will receive from \code{config.style.bar} and \code{config.style.foo} (the specified style \code{"foo"} has higher precedence).}

\item{tension}{(\emph{MarkDef})}

\item{text}{(\emph{MarkDef})}

\item{theta}{(\emph{MarkDef}) - For arc marks, the arc length in radians if theta2 is not specified, otherwise the start arc angle. (A value of 0 indicates up or “north”, increasing values proceed clockwise.)
\itemize{
\item For text marks, polar coordinate angle in radians.
}}

\item{theta2}{(\emph{MarkDef}) The end angle of arc marks in radians. A value of 0 indicates up or “north”, increasing values proceed clockwise.}

\item{theta2Offset}{(\emph{MarkDef}) Offset for theta2.}

\item{thetaOffset}{(\emph{MarkDef}) Offset for theta.}

\item{thickness}{(\emph{MarkDef}) Thickness of the tick mark.

\strong{Default value:}  \code{1}}

\item{timeUnitBand}{(\emph{MarkDef}) Default relative band size for a time unit. If set to \code{1}, the bandwidth of the marks will be equal to the time unit band step. If set to \code{0.5}, bandwidth of the marks will be half of the time unit band step.}

\item{timeUnitBandPosition}{(\emph{MarkDef}) Default relative band position for a time unit. If set to \code{0}, the marks will be positioned at the beginning of the time unit band step. If set to \code{0.5}, the marks will be positioned in the middle of the time unit band step.}

\item{tooltip}{(\emph{MarkDef}) The tooltip text string to show upon mouse hover or an object defining which fields should the tooltip be derived from.
\itemize{
\item If \code{tooltip} is \code{true} or \code{{"content": "encoding"}}, then all fields from \code{encoding} will be used. - If \code{tooltip} is \code{{"content": "data"}}, then all fields that appear in the highlighted data point will be used. - If set to \code{null} or \code{false}, then no tooltip will be used.
}

See the \href{https://vega.github.io/vega-lite/docs/tooltip.html}{\code{tooltip}} documentation for a detailed discussion about tooltip  in Vega-Lite.

\strong{Default value:} \code{null}}

\item{url}{(\emph{MarkDef})}

\item{width}{(\emph{MarkDef})}

\item{x}{(\emph{MarkDef}) X coordinates of the marks, or width of horizontal \code{"bar"} and \code{"area"} without specified \code{x2} or \code{width}.

The \code{value} of this channel can be a number or a string \code{"width"} for the width of the plot.}

\item{x2}{(\emph{MarkDef}) X2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and  \code{"rule"}.

The \code{value} of this channel can be a number or a string \code{"width"} for the width of the plot.}

\item{x2Offset}{(\emph{MarkDef}) Offset for x2-position.}

\item{xOffset}{(\emph{MarkDef}) Offset for x-position.}

\item{y}{(\emph{MarkDef}) Y coordinates of the marks, or height of vertical \code{"bar"} and \code{"area"} without specified \code{y2} or \code{height}.

The \code{value} of this channel can be a number or a string \code{"height"} for the height of the plot.}

\item{y2}{(\emph{MarkDef}) Y2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and  \code{"rule"}.

The \code{value} of this channel can be a number or a string \code{"height"} for the height of the plot.}

\item{y2Offset}{(\emph{MarkDef}) Offset for y2-position.}

\item{yOffset}{(\emph{MarkDef}) Offset for y-position.}
}
\value{
A modified Vega-Lite Spec
}
\description{
vl_mark_rect
}
