% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_repeat_row}
\alias{vl_repeat_row}
\title{vl_repeat_row}
\usage{
vl_repeat_row(spec, ...)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{...}{fields to use for repeat (strings)}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add repeat to a vega-lite spec.
}
