% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_repeat_wrap}
\alias{vl_repeat_wrap}
\title{vl_repeat_wrap}
\usage{
vl_repeat_wrap(spec, ..., columns = 2)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{...}{fields to use for repeat (strings)}

\item{columns}{number of columns to add}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add repeat to a vega-lite spec.
}
