% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_resolve_legend_size}
\alias{vl_resolve_legend_size}
\title{vl_resolve_legend_size}
\usage{
vl_resolve_legend_size(spec, how = c("independent", "shared"))
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{how}{how to resolve legend for size, one of independent, shared}
}
\value{
A modified spec
}
\description{
Add resolving axis definitions to a vega-lite spec.
}
