% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_SampleTransform}
\alias{vl_SampleTransform}
\title{vl_SampleTransform}
\usage{
vl_SampleTransform(sample = NULL)
}
\arguments{
\item{sample}{The maximum number of data objects to include in the sample.

\strong{Default value:} \code{1000} (type = number)}
}
\value{
A modified spec
}
\description{
Create spec for SampleTransform.
}
