% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_scale_fillOpacity}
\alias{vl_scale_fillOpacity}
\title{vl_scale_fillOpacity}
\usage{
vl_scale_fillOpacity(spec, base = NULL, bins = NULL, clamp = NULL,
  constant = NULL, domain = NULL, exponent = NULL,
  interpolate = NULL, nice = NULL, padding = NULL,
  paddingInner = NULL, paddingOuter = NULL, range = NULL,
  rangeStep = NULL, round = NULL, scheme = NULL, type = NULL,
  zero = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{base}{The logarithm base of the \code{log} scale (default \code{10}). (type = number)}

\item{bins}{An array of bin boundaries over the scale domain. If provided, axes and legends will use the bin boundaries to inform the choice of tick marks and text labels. (type = array)}

\item{clamp}{If \code{true}, values that exceed the data domain are clamped to either the minimum or maximum range value

\strong{Default value:} derived from the \href{https://vega.github.io/vega-lite/docs/config.html#scale-config}{scale config}'s \code{clamp} (\code{true} by default). (type = boolean)}

\item{constant}{A constant determining the slope of the symlog function around zero. Only used for \code{symlog} scales.

\strong{Default value:} \code{1} (type = number)}

\item{domain}{Customized domain values.

For \emph{quantitative} fields, \code{domain} can take the form of a two-element array with minimum and maximum values.  \href{https://vega.github.io/vega-lite/docs/scale.html#piecewise}{Piecewise scales} can be created by providing a \code{domain} with more than two entries.
If the input field is aggregated, \code{domain} can also be a string value \code{"unaggregated"}, indicating that the domain should include the raw data values prior to the aggregation.

For \emph{temporal} fields, \code{domain} can be a two-element array minimum and maximum values, in the form of either timestamps or the \href{https://vega.github.io/vega-lite/docs/types.html#datetime}{DateTime definition objects}.

For \emph{ordinal} and \emph{nominal} fields, \code{domain} can be an array that lists valid input values.

The \code{selection} property can be used to \href{https://vega.github.io/vega-lite/docs/selection.html#scale-domains}{interactively determine} the scale domain. (type = Varies)}

\item{exponent}{The exponent of the \code{pow} scale. (type = number)}

\item{interpolate}{The interpolation method for range values. By default, a general interpolator for numbers, dates, strings and colors (in HCL space) is used. For color ranges, this property allows interpolation in alternative color spaces. Legal values include \code{rgb}, \code{hsl}, \code{hsl-long}, \code{lab}, \code{hcl}, \code{hcl-long}, \code{cubehelix} and \code{cubehelix-long} ('-long' variants use longer paths in polar coordinate spaces). If object-valued, this property accepts an object with a string-valued \emph{type} property and an optional numeric \emph{gamma} property applicable to rgb and cubehelix interpolators. For more, see the \href{https://github.com/d3/d3-interpolate}{d3-interpolate documentation}.
\itemize{
\item \strong{Default value:} \code{hcl} (type = Varies)
}}

\item{nice}{Extending the domain so that it starts and ends on nice round values. This method typically modifies the scale’s domain, and may only extend the bounds to the nearest round value. Nicing is useful if the domain is computed from data and may be irregular. For example, for a domain of \emph{\link{0.201479…, 0.996679…}}, a nice domain might be \emph{\link{0.2, 1.0}}.

For quantitative scales such as linear, \code{nice} can be either a boolean flag or a number. If \code{nice} is a number, it will represent a desired tick count. This allows greater control over the step size used to extend the bounds, guaranteeing that the returned ticks will exactly cover the domain.

For temporal fields with time and utc scales, the \code{nice} value can be a string indicating the desired time interval. Legal values are \code{"millisecond"}, \code{"second"}, \code{"minute"}, \code{"hour"}, \code{"day"}, \code{"week"}, \code{"month"}, and \code{"year"}. Alternatively, \code{time} and \code{utc} scales can accept an object-valued interval specifier of the form \code{{"interval": "month", "step": 3}}, which includes a desired number of interval steps. Here, the domain would snap to quarter (Jan, Apr, Jul, Oct) boundaries.

\strong{Default value:} \code{true} for unbinned \emph{quantitative} fields; \code{false} otherwise. (type = Varies)}

\item{padding}{For \emph{\href{https://vega.github.io/vega-lite/docs/scale.html#continuous}{continuous}} scales, expands the scale domain to accommodate the specified number of pixels on each of the scale range. The scale range must represent pixels for this parameter to function as intended. Padding adjustment is performed prior to all other adjustments, including the effects of the zero, nice, domainMin, and domainMax properties.

For \emph{\href{https://vega.github.io/vega-lite/docs/scale.html#band}{band}} scales, shortcut for setting \code{paddingInner} and \code{paddingOuter} to the same value.

For \emph{\href{https://vega.github.io/vega-lite/docs/scale.html#point}{point}} scales, alias for \code{paddingOuter}.

\strong{Default value:} For \emph{continuous} scales, derived from the \href{https://vega.github.io/vega-lite/docs/scale.html#config}{scale config}'s \code{continuousPadding}.
For \emph{band and point} scales, see \code{paddingInner} and \code{paddingOuter}. (type = number)}

\item{paddingInner}{The inner padding (spacing) within each band step of band scales, as a fraction of the step size. This value must lie in the range \link{0,1}.

For point scale, this property is invalid as point scales do not have internal band widths (only step sizes between bands).

\strong{Default value:} derived from the \href{https://vega.github.io/vega-lite/docs/scale.html#config}{scale config}'s \code{bandPaddingInner}. (type = number)}

\item{paddingOuter}{The outer padding (spacing) at the ends of the range of band and point scales,
as a fraction of the step size. This value must lie in the range \link{0,1}.

\strong{Default value:} derived from the \href{https://vega.github.io/vega-lite/docs/scale.html#config}{scale config}'s \code{bandPaddingOuter} for band scales and \code{pointPadding} for point scales. (type = number)}

\item{range}{The range of the scale. One of:
\itemize{
\item A string indicating a \href{https://vega.github.io/vega-lite/docs/scale.html#range-config}{pre-defined named scale range} (e.g., example, \code{"symbol"}, or \code{"diverging"}).
\item For \href{https://vega.github.io/vega-lite/docs/scale.html#continuous}{continuous scales}, two-element array indicating  minimum and maximum values, or an array with more than two entries for specifying a \href{https://vega.github.io/vega-lite/docs/scale.html#piecewise}{piecewise scale}.
\item For \href{https://vega.github.io/vega-lite/docs/scale.html#discrete}{discrete} and \href{https://vega.github.io/vega-lite/docs/scale.html#discretizing}{discretizing} scales, an array of desired output values.
}

\strong{Notes:}
\enumerate{
\item For color scales you can also specify a color \href{https://vega.github.io/vega-lite/docs/scale.html#scheme}{scheme} instead of \code{range}.
\item Any directly specified \code{range} for \code{x} and \code{y} channels will be ignored. Range can be customized via the view's corresponding \href{https://vega.github.io/vega-lite/docs/size.html}{size} (\code{width} and \code{height}) or via \href{#range-step}{range steps and paddings properties} for \href{#band}{band} and \href{#point}{point} scales. (type = Varies)
}}

\item{rangeStep}{The distance between the starts of adjacent bands or points in \href{https://vega.github.io/vega-lite/docs/scale.html#band}{band} and \href{https://vega.github.io/vega-lite/docs/scale.html#point}{point} scales.

If \code{rangeStep} is \code{null} or if the view contains the scale's corresponding \href{https://vega.github.io/vega-lite/docs/size.html}{size} (\code{width} for \code{x} scales and \code{height} for \code{y} scales), \code{rangeStep} will be automatically determined to fit the size of the view.

\strong{Default value:}  derived the \href{https://vega.github.io/vega-lite/docs/config.html#scale-config}{scale config}'s \code{textXRangeStep} (\code{90} by default) for x-scales of \code{text} marks and \code{rangeStep} (\code{21} by default) for x-scales of other marks and y-scales.

\strong{Warning}: If \code{rangeStep} is \code{null} and the cardinality of the scale's domain is higher than \code{width} or \code{height}, the rangeStep might become less than one pixel and the mark might not appear correctly. (type = number OR null)}

\item{round}{If \code{true}, rounds numeric output values to integers. This can be helpful for snapping to the pixel grid.

\strong{Default value:} \code{false}. (type = boolean)}

\item{scheme}{A string indicating a color \href{https://vega.github.io/vega-lite/docs/scale.html#scheme}{scheme} name (e.g., \code{"category10"} or \code{"blues"}) or a \href{https://vega.github.io/vega-lite/docs/scale.html#scheme-params}{scheme parameter object}.

Discrete color schemes may be used with \href{https://vega.github.io/vega-lite/docs/scale.html#discrete}{discrete} or \href{https://vega.github.io/vega-lite/docs/scale.html#discretizing}{discretizing} scales. Continuous color schemes are intended for use with color scales.

For the full list of supported schemes, please refer to the \href{https://vega.github.io/vega/docs/schemes/#reference}{Vega Scheme} reference. (type = Varies)}

\item{type}{The type of scale.  Vega-Lite supports the following categories of scale types:
\enumerate{
\item \href{https://vega.github.io/vega-lite/docs/scale.html#continuous}{Continuous Scales} -- mapping continuous domains to continuous output ranges (\href{https://vega.github.io/vega-lite/docs/scale.html#linear}{"linear"}, \href{https://vega.github.io/vega-lite/docs/scale.html#pow}{"pow"}, \href{https://vega.github.io/vega-lite/docs/scale.html#sqrt}{"sqrt"}, \href{https://vega.github.io/vega-lite/docs/scale.html#symlog}{"symlog"}, \href{https://vega.github.io/vega-lite/docs/scale.html#log}{"log"}, \href{https://vega.github.io/vega-lite/docs/scale.html#time}{"time"}, \href{https://vega.github.io/vega-lite/docs/scale.html#utc}{"utc"}.
\item \href{https://vega.github.io/vega-lite/docs/scale.html#discrete}{Discrete Scales} -- mapping discrete domains to discrete (\href{https://vega.github.io/vega-lite/docs/scale.html#ordinal}{"ordinal"}) or continuous (\href{https://vega.github.io/vega-lite/docs/scale.html#band}{"band"} and \href{https://vega.github.io/vega-lite/docs/scale.html#point}{"point"}) output ranges.
\item \href{https://vega.github.io/vega-lite/docs/scale.html#discretizing}{Discretizing Scales} -- mapping continuous domains to discrete output ranges \href{https://vega.github.io/vega-lite/docs/scale.html#bin-ordinal}{"bin-ordinal"}, \href{https://vega.github.io/vega-lite/docs/scale.html#quantile}{"quantile"}, \href{https://vega.github.io/vega-lite/docs/scale.html#quantize}{"quantize"} and \href{https://vega.github.io/vega-lite/docs/scale.html#threshold}{"threshold"}.
}

\strong{Default value:} please see the \href{https://vega.github.io/vega-lite/docs/scale.html#type}{scale type table}. (type = ScaleType)}

\item{zero}{If \code{true}, ensures that a zero baseline value is included in the scale domain.

\strong{Default value:} \code{true} for x and y channels if the quantitative field is not binned and no custom \code{domain} is provided; \code{false} otherwise.

\strong{Note:} Log, time, and utc scales do not support \code{zero}. (type = boolean)}
}
\value{
A modified spec
}
\description{
Add scale to encoding for fillOpacity in a vega-lite spec.
}
\seealso{
\code{\link[=vl_fillOpacity]{vl_fillOpacity()}}
}
