% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_sort_opacity_by_encoding}
\alias{vl_sort_opacity_by_encoding}
\title{vl_sort_opacity_by_encoding}
\usage{
vl_sort_opacity_by_encoding(spec, encoding = NULL, order = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{encoding}{(\emph{SortByEncoding}) The \href{https://vega.github.io/vega-lite/docs/encoding.html#channels}{encoding channel} to sort by (e.g., \code{"x"}, \code{"y"})}

\item{order}{(\emph{SortByEncoding}) The sort order. One of \code{"ascending"} (default), \code{"descending"}, or \code{null} (no not sort).}
}
\value{
A modified spec
}
\description{
Add sort by another encoding to encoding for opacity in a vega-lite spec.
}
\seealso{
\link{vl_sort_opacity} to sort by same field,
\code{\link[=vl_sort_opacity_by_field]{vl_sort_opacity_by_field()}} to sort by another field
}
