% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_sort_y}
\alias{vl_sort_y}
\title{vl_sort_y}
\usage{
vl_sort_y(spec, value)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{value}{either 'ascending' or 'descending' to specify sort order (using
this encoding), a list with a custom ordering, or NA to specify no sorting}
}
\value{
A modified spec
}
\description{
Add sort to encoding for y to a vega-lite spec.
}
\seealso{
\code{\link[=vl_sort_y_by_field]{vl_sort_y_by_field()}} to sort by another field,
\code{\link[=vl_sort_y_by_encoding]{vl_sort_y_by_encoding()}} to sort by another encoding
}
