% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_sort_y_by_field}
\alias{vl_sort_y_by_field}
\title{vl_sort_y_by_field}
\usage{
vl_sort_y_by_field(spec, field = NULL, op = NULL, order = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{field}{The data \href{https://vega.github.io/vega-lite/docs/field.html}{field} to sort by.

\strong{Default value:} If unspecified, defaults to the field specified in the outer data reference. (type = Field)}

\item{op}{An \href{https://vega.github.io/vega-lite/docs/aggregate.html#ops}{aggregate operation} to perform on the field prior to sorting (e.g., \code{"count"}, \code{"mean"} and \code{"median"}).
An aggregation is required when there are multiple values of the sort field for each encoded data field.
The input data objects will be aggregated, grouped by the encoded data field.

For a full list of operations, please see the documentation for \href{https://vega.github.io/vega-lite/docs/aggregate.html#ops}{aggregate}.

\strong{Default value:} \code{"sum"} for stacked plots. Otherwise, \code{"mean"}. (type = AggregateOp)}

\item{order}{The sort order. One of \code{"ascending"} (default), \code{"descending"}, or \code{null} (no not sort). (type = Varies)}
}
\value{
A modified spec
}
\description{
Add sort by another field to encoding for y in a vega-lite spec.
}
\seealso{
\link{vl_sort_y} to sort by same field,
\code{\link[=vl_sort_y_by_encoding]{vl_sort_y_by_encoding()}} to sort by another encoding
}
