% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_stack}
\alias{vl_stack}
\title{vl_stack}
\usage{
vl_stack(
  spec,
  .object = NULL,
  stack = NULL,
  as = NULL,
  groupby = NULL,
  offset = NULL,
  sort = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: StackTransform}

\item{stack}{The field which is stacked.}

\item{as}{Output field names. This can be either a string or an array of strings with two elements denoting the name for the
fields for stack start and stack end respectively. If a single string(e.g., \code{"val"}) is provided, the end field will
be \code{"val_end"}.}

\item{groupby}{The data fields to group by.}

\item{offset}{Mode for stacking marks. One of \code{"zero"} (default), \code{"center"}, or \code{"normalize"}. The \code{"zero"} offset will stack
starting at \code{0}. The \code{"center"} offset will center the stacks. The \code{"normalize"} offset will compute percentage
values for each stack point, with output values in the range \verb{[0,1]}. #' #' \strong{Default value:} \code{"zero"}}

\item{sort}{Field that determines the order of leaves in the stacked charts.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add StackTransform to a vega-lite spec.
}
