% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_stack_y}
\alias{vl_stack_y}
\title{vl_stack_y}
\usage{
vl_stack_y(spec, stack = c("zero", "center", "normalize", NA))
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{stack}{one of 'zero', 'center', 'normalize', NA}
}
\value{
A modified spec
}
\description{
Add stack to encoding for y to a vega-lite spec.
}
\seealso{
\code{\link[=vl_stack]{vl_stack()}} to add a top level stack transform
}
