% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_TimeUnitTransform}
\alias{vl_TimeUnitTransform}
\title{vl_TimeUnitTransform}
\usage{
vl_TimeUnitTransform(as = NULL, field = NULL, timeUnit = NULL)
}
\arguments{
\item{as}{The output field to write the timeUnit value. (type = string)}

\item{field}{The data field to apply time unit. (type = string)}

\item{timeUnit}{The timeUnit. (type = TimeUnit)}
}
\value{
A modified spec
}
\description{
Create spec for TimeUnitTransform.
}
