% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_transform_aggregate}
\alias{vl_transform_aggregate}
\title{vl_transform_aggregate}
\usage{
vl_transform_aggregate(spec, aggregate = NULL, groupby = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{aggregate}{Array of objects that define fields to aggregate. (type = array)}

\item{groupby}{The data fields to group by. If not specified, a single group containing all data objects will be used. (type = array)}
}
\value{
A modified spec
}
\description{
Add AggregateTransform to a vega-lite spec.
}
