% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_transform_calculate}
\alias{vl_transform_calculate}
\title{vl_transform_calculate}
\usage{
vl_transform_calculate(spec, as = NULL, calculate = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{as}{The field for storing the computed formula value. (type = string)}

\item{calculate}{A [expression](https://vega.github.io/vega-lite/docs/types.html#expression) string. Use the variable `datum` to refer to the current data object. (type = string)}
}
\value{
A modified spec
}
\description{
Add CalculateTransform to a vega-lite spec.
}
