% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_transform_filter}
\alias{vl_transform_filter}
\title{vl_transform_filter}
\usage{
vl_transform_filter(spec, filter = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{filter}{The `filter` property must be one of the predicate definitions:

1) an [expression](https://vega.github.io/vega-lite/docs/types.html#expression) string,
where `datum` can be used to refer to the current data object

2) one of the field predicates: [`equal`](https://vega.github.io/vega-lite/docs/filter.html#equal-predicate),
[`lt`](https://vega.github.io/vega-lite/docs/filter.html#lt-predicate),
[`lte`](https://vega.github.io/vega-lite/docs/filter.html#lte-predicate),
[`gt`](https://vega.github.io/vega-lite/docs/filter.html#gt-predicate),
[`gte`](https://vega.github.io/vega-lite/docs/filter.html#gte-predicate),
[`range`](https://vega.github.io/vega-lite/docs/filter.html#range-predicate),
[`oneOf`](https://vega.github.io/vega-lite/docs/filter.html#one-of-predicate),
or [`valid`](https://vega.github.io/vega-lite/docs/filter.html#valid-predicate),

3) a [selection predicate](https://vega.github.io/vega-lite/docs/filter.html#selection-predicate)

4) a logical operand that combines (1), (2), or (3). (type = LogicalOperand<Predicate>)}
}
\value{
A modified spec
}
\description{
Add FilterTransform to a vega-lite spec.
}
