% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_transform_fold}
\alias{vl_transform_fold}
\title{vl_transform_fold}
\usage{
vl_transform_fold(spec, as = NULL, fold = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{as}{The output field names for the key and value properties produced by the fold transform.
__Default value:__ `["key", "value"]` (type = array)}

\item{fold}{An array of data fields indicating the properties to fold. (type = array)}
}
\value{
A modified spec
}
\description{
Add FoldTransform to a vega-lite spec.
}
