% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_transform_impute}
\alias{vl_transform_impute}
\title{vl_transform_impute}
\usage{
vl_transform_impute(spec, frame = NULL, groupby = NULL,
  impute = NULL, key = NULL, keyvals = NULL, method = NULL,
  value = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{frame}{A frame specification as a two-element array used to control the window over which the specified method is applied. The array entries should either be a number indicating the offset from the current data object, or null to indicate unbounded rows preceding or following the current data object.  For example, the value `[-5, 5]` indicates that the window should include five objects preceding and five objects following the current object.

__Default value:__:  `[null, null]` indicating that the window includes all objects. (type = array)}

\item{groupby}{An optional array of fields by which to group the values.
Imputation will then be performed on a per-group basis. (type = array)}

\item{impute}{The data field for which the missing values should be imputed. (type = string)}

\item{key}{A key field that uniquely identifies data objects within a group.
Missing key values (those occurring in the data but not in the current group) will be imputed. (type = string)}

\item{keyvals}{Defines the key values that should be considered for imputation.
An array of key values or an object defining a [number sequence](https://vega.github.io/vega-lite/docs/impute.html#sequence-def).

If provided, this will be used in addition to the key values observed within the input data.  If not provided, the values will be derived from all unique values of the `key` field. For `impute` in `encoding`, the key field is the x-field if the y-field is imputed, or vice versa.

If there is no impute grouping, this property _must_ be specified. (type = Varies)}

\item{method}{The imputation method to use for the field value of imputed data objects.
One of `value`, `mean`, `median`, `max` or `min`.

__Default value:__  `"value"` (type = ImputeMethod)}

\item{value}{The field value to use when the imputation `method` is `"value"`. (type = Varies)}
}
\value{
A modified spec
}
\description{
Add ImputeTransform to a vega-lite spec.
}
