% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_transform_lookup}
\alias{vl_transform_lookup}
\title{vl_transform_lookup}
\usage{
vl_transform_lookup(spec, as = NULL, default = NULL, from = NULL,
  lookup = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{as}{The field or fields for storing the computed formula value.
If `from.fields` is specified, the transform will use the same names for `as`.
If `from.fields` is not specified, `as` has to be a string and we put the whole object into the data under the specified name. (type = Varies)}

\item{default}{The default value to use if lookup fails.

__Default value:__ `null` (type = string)}

\item{from}{Secondary data reference. (type = LookupData)}

\item{lookup}{Key in primary data source. (type = string)}
}
\value{
A modified spec
}
\description{
Add LookupTransform to a vega-lite spec.
}
