% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_transform_sample}
\alias{vl_transform_sample}
\title{vl_transform_sample}
\usage{
vl_transform_sample(spec, sample = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{sample}{The maximum number of data objects to include in the sample.

__Default value:__ `1000` (type = number)}
}
\value{
A modified spec
}
\description{
Add SampleTransform to a vega-lite spec.
}
