% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{vlbuildr-deprecated}
\alias{vlbuildr-deprecated}
\title{Deprecated functions from vlbuildr}
\arguments{
\item{...}{argument to pass to new function}
}
\description{
These functions have been deprecated in favor of a different
naming convention.
}
