% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlbuildr-package.R
\docType{package}
\name{vlbuildr-package}
\alias{vlbuildr}
\alias{vlbuildr-package}
\title{vlbuildr: Build Vega-Lite Specs in R}
\description{
Build vegalite specs in R. Inspired by vegalite package as well as vega-lite-api and altair. Dependent on vegawidget for actually rendering specs!
}
\section{Types of functions}{

This package has three types of functions:
\itemize{
\item \code{vl_chart} initializes a new vega-lite specification which can be build upon.
\item Other \code{vl_} functions which build components to add to the spec and add them
at the correct place.
\item Functions grouped in \code{vl} list and accessible via \verb{vl$}.
These functions make a sub-component of Vega-Lite spec. These can help build components
that might be arguments to a \code{vl_} function, although generally the goal is for the \code{vl_} functions
to be granular enough that vcomplex inputs are not needed.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://vegawidget.github.io/vlbuildr/index.html}
}

}
\author{
\strong{Maintainer}: Alicia Schep \email{aschep@gmail.com}

Authors:
\itemize{
  \item Vega-Lite Developers (The Vega-Lite schema is used to directly populate much of the documentation of this package.)
}

}
\keyword{internal}
