% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{add_data}
\alias{add_data}
\alias{vl_add_data}
\title{add_data}
\usage{
vl_add_data(spec, format = NULL, name = NULL, url = NULL,
  values = NULL, sequence = NULL, sphere = NULL, graticule = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{format}{An object that specifies the format for parsing the data. (type = DataFormat)}

\item{name}{Provide a placeholder name and bind data at runtime. (type = string)}

\item{url}{An URL from which to load the data set. Use the `format.type` property
to ensure the loaded data is correctly parsed. (type = string)}

\item{values}{The full data set, included inline. This can be an array of objects or primitive values, an object, or a string.
Arrays of primitive values are ingested as objects with a `data` property. Strings are parsed according to the specified format type. (type = InlineDataset)}

\item{sequence}{Generate a sequence of numbers. (type = SequenceParams)}

\item{sphere}{Generate sphere GeoJSON data for the full globe. (type = Varies)}

\item{graticule}{Generate graticule GeoJSON data for geographic reference lines. (type = Varies)}
}
\value{
A modified spec
}
\description{
Add data to a vega-lite spec.
}
