% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{aggregate_encoding}
\alias{aggregate_encoding}
\alias{vl_aggregate_angle}
\alias{vl_aggregate_color}
\alias{vl_aggregate_description}
\alias{vl_aggregate_detail}
\alias{vl_aggregate_fill}
\alias{vl_aggregate_fillOpacity}
\alias{vl_aggregate_href}
\alias{vl_aggregate_key}
\alias{vl_aggregate_latitude}
\alias{vl_aggregate_latitude2}
\alias{vl_aggregate_longitude}
\alias{vl_aggregate_longitude2}
\alias{vl_aggregate_opacity}
\alias{vl_aggregate_order}
\alias{vl_aggregate_radius}
\alias{vl_aggregate_radius2}
\alias{vl_aggregate_shape}
\alias{vl_aggregate_size}
\alias{vl_aggregate_stroke}
\alias{vl_aggregate_strokeDash}
\alias{vl_aggregate_strokeOpacity}
\alias{vl_aggregate_strokeWidth}
\alias{vl_aggregate_text}
\alias{vl_aggregate_theta}
\alias{vl_aggregate_theta2}
\alias{vl_aggregate_tooltip}
\alias{vl_aggregate_url}
\alias{vl_aggregate_x}
\alias{vl_aggregate_x2}
\alias{vl_aggregate_xError}
\alias{vl_aggregate_xError2}
\alias{vl_aggregate_xOffset}
\alias{vl_aggregate_y}
\alias{vl_aggregate_y2}
\alias{vl_aggregate_yError}
\alias{vl_aggregate_yError2}
\alias{vl_aggregate_yOffset}
\alias{vl_aggregate_facet}
\alias{vl_aggregate_row}
\alias{vl_aggregate_column}
\title{Add aggregate transform to encoding}
\usage{
vl_aggregate_angle(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_color(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_description(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_detail(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_fill(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_fillOpacity(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_href(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_key(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_latitude(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_latitude2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_longitude(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_longitude2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_opacity(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_order(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_radius(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_radius2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_shape(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_size(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_stroke(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_strokeDash(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_strokeOpacity(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_strokeWidth(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_text(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_theta(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_theta2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_tooltip(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_url(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_x(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_x2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_xError(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_xError2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_xOffset(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_y(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_y2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_yError(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_yError2(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_yOffset(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_facet(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_row(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)

vl_aggregate_column(
  spec,
  aggregate = c("argmax", "argmin", "average", "count", "distinct", "max", "mean",
    "median", "min", "missing", "product", "q1", "q3", "ci0", "ci1", "stderr", "stdev",
    "stdevp", "sum", "valid", "values", "variance", "variancep", NA)
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{aggregate}{One of 'argmax', 'argmin', 'average', 'count', 'distinct', 'max', 'mean', 'median', 'min', 'missing', 'product', 'q1', 'q3', 'ci0', 'ci1', 'stderr', 'stdev', 'stdevp', 'sum', 'valid', 'values', 'variance', 'variancep', NA}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add aggregate parameters to an encoding
}
