% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{sort_encoding}
\alias{sort_encoding}
\alias{vl_sort_color}
\alias{vl_sort_fill}
\alias{vl_sort_fillOpacity}
\alias{vl_sort_opacity}
\alias{vl_sort_order}
\alias{vl_sort_shape}
\alias{vl_sort_size}
\alias{vl_sort_stroke}
\alias{vl_sort_strokeOpacity}
\alias{vl_sort_strokeWidth}
\alias{vl_sort_x}
\alias{vl_sort_y}
\title{Add sort transform to encoding}
\usage{
vl_sort_color(spec, value)

vl_sort_fill(spec, value)

vl_sort_fillOpacity(spec, value)

vl_sort_opacity(spec, value)

vl_sort_order(spec, value)

vl_sort_shape(spec, value)

vl_sort_size(spec, value)

vl_sort_stroke(spec, value)

vl_sort_strokeOpacity(spec, value)

vl_sort_strokeWidth(spec, value)

vl_sort_x(spec, value)

vl_sort_y(spec, value)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{value}{either 'ascending' or 'descending' to specify sort order (using
this encoding), a list with a custom ordering, or NA to specify no sorting}
}
\value{
A modified spec
}
\description{
Add sort parameters to an encoding
}
\seealso{
\code{\link[=sort_encoding_by_field]{sort_encoding_by_field()}} and \code{\link[=sort_encoding_by_encoding]{sort_encoding_by_encoding()}}
}
