% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{stack_encoding}
\alias{stack_encoding}
\alias{vl_stack_x}
\alias{vl_stack_y}
\title{Add stack transform to encoding}
\usage{
vl_stack_x(spec, stack = c("zero", "center", "normalize", NA))

vl_stack_y(spec, stack = c("zero", "center", "normalize", NA))
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{stack}{one of 'zero', 'center', 'normalize', NA}
}
\value{
A modified spec
}
\description{
Add stack parameters to an encoding
}
\seealso{
\code{\link[=vl_stack]{vl_stack()}} to add a top level stack transform
}
