% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$Axis}
\alias{vl$Axis}
\title{Axis}
\arguments{
\item{bandPosition}{(\emph{Axis}) An interpolation fraction indicating where, for \code{band} scales, axis ticks should be positioned. A value of \code{0} places ticks at the left edge of their bands. A value of \code{0.5} places ticks in the middle of their bands.

\strong{Default value:} \code{0.5}}

\item{domain}{(\emph{Axis}) A boolean flag indicating if the domain (the axis baseline) should be included as part of the axis.

\strong{Default value:} \code{true}}

\item{domainColor}{(\emph{Axis}) Color of axis domain line.

\strong{Default value:} \code{"gray"}.}

\item{domainDash}{(\emph{Axis}) An array of alternating [stroke, space] lengths for dashed domain lines.}

\item{domainDashOffset}{(\emph{Axis}) The pixel offset at which to start drawing with the domain dash array.}

\item{domainOpacity}{(\emph{Axis}) Opacity of the axis domain line.}

\item{domainWidth}{(\emph{Axis}) Stroke width of axis domain line

\strong{Default value:} \code{1}}

\item{format}{(\emph{Axis}) The text formatting pattern for labels of guides (axes, legends, headers) and text marks.
\itemize{
\item If the format type is \code{"number"} (e.g., for quantitative fields), this is D3's \href{https://github.com/d3/d3-format#locale_format}{number format pattern}.
\item If the format type is \code{"time"} (e.g., for temporal fields), this is D3's \href{https://github.com/d3/d3-time-format#locale_format}{time format pattern}.
}

See the \href{https://vega.github.io/vega-lite/docs/format.html}{format documentation} for more examples.

\strong{Default value:}  Derived from \href{https://vega.github.io/vega-lite/docs/config.html#format}{numberFormat} config for number format and from \href{https://vega.github.io/vega-lite/docs/config.html#format}{timeFormat} config for time format.}

\item{formatType}{(\emph{Axis}) The format type for labels (\code{"number"} or \code{"time"}).

\strong{Default value:}
\itemize{
\item \code{"time"} for temporal fields and ordinal and nomimal fields with \code{timeUnit}.
\item \code{"number"} for quantitative fields as well as ordinal and nomimal fields without \code{timeUnit}.
}}

\item{grid}{(\emph{Axis}) A boolean flag indicating if grid lines should be included as part of the axis

\strong{Default value:} \code{true} for \href{https://vega.github.io/vega-lite/docs/scale.html#continuous}{continuous scales} that are not binned; otherwise, \code{false}.}

\item{gridColor}{(\emph{Axis})}

\item{gridDash}{(\emph{Axis})}

\item{gridDashOffset}{(\emph{Axis})}

\item{gridOpacity}{(\emph{Axis})}

\item{gridWidth}{(\emph{Axis})}

\item{labelAlign}{(\emph{Axis})}

\item{labelAngle}{(\emph{Axis}) The rotation angle of the axis labels.

\strong{Default value:} \code{-90} for nominal and ordinal fields; \code{0} otherwise.}

\item{labelBaseline}{(\emph{Axis})}

\item{labelBound}{(\emph{Axis}) Indicates if labels should be hidden if they exceed the axis range. If \code{false} (the default) no bounds overlap analysis is performed. If \code{true}, labels will be hidden if they exceed the axis range by more than 1 pixel. If this property is a number, it specifies the pixel tolerance: the maximum amount by which a label bounding box may exceed the axis range.

\strong{Default value:} \code{false}.}

\item{labelColor}{(\emph{Axis})}

\item{labelExpr}{(\emph{Axis}) \href{https://vega.github.io/vega/docs/expressions/}{Vega expression} for customizing labels.

\strong{Note:} The label text and value can be assessed via the \code{label} and \code{value} properties of the axis's backing \code{datum} object.}

\item{labelFlush}{(\emph{Axis}) Indicates if the first and last axis labels should be aligned flush with the scale range. Flush alignment for a horizontal axis will left-align the first label and right-align the last label. For vertical axes, bottom and top text baselines are applied instead. If this property is a number, it also indicates the number of pixels by which to offset the first and last labels; for example, a value of 2 will flush-align the first and last labels and also push them 2 pixels outward from the center of the axis. The additional adjustment can sometimes help the labels better visually group with corresponding axis ticks.

\strong{Default value:} \code{true} for axis of a continuous x-scale. Otherwise, \code{false}.}

\item{labelFlushOffset}{(\emph{Axis}) Indicates the number of pixels by which to offset flush-adjusted labels. For example, a value of \code{2} will push flush-adjusted labels 2 pixels outward from the center of the axis. Offsets can help the labels better visually group with corresponding axis ticks.

\strong{Default value:} \code{0}.}

\item{labelFont}{(\emph{Axis})}

\item{labelFontSize}{(\emph{Axis})}

\item{labelFontStyle}{(\emph{Axis})}

\item{labelFontWeight}{(\emph{Axis})}

\item{labelLimit}{(\emph{Axis}) Maximum allowed pixel width of axis tick labels.

\strong{Default value:} \code{180}}

\item{labelOpacity}{(\emph{Axis})}

\item{labelOverlap}{(\emph{Axis}) The strategy to use for resolving overlap of axis labels. If \code{false} (the default), no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a strategy of removing every other label is used (this works well for standard linear axes). If set to \code{"greedy"}, a linear scan of the labels is performed, removing any labels that overlaps with the last visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{true} for non-nominal fields with non-log scales; \code{"greedy"} for log scales; otherwise \code{false}.}

\item{labelPadding}{(\emph{Axis}) The padding, in pixels, between axis and text labels.

\strong{Default value:} \code{2}}

\item{labelSeparation}{(\emph{Axis}) The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default \code{0}). This property is ignored if \emph{labelOverlap} resolution is not enabled.}

\item{labels}{(\emph{Axis}) A boolean flag indicating if labels should be included as part of the axis.

\strong{Default value:} \code{true}.}

\item{maxExtent}{(\emph{Axis}) The maximum extent in pixels that axis ticks and labels should use. This determines a maximum offset value for axis titles.

\strong{Default value:} \code{undefined}.}

\item{minExtent}{(\emph{Axis}) The minimum extent in pixels that axis ticks and labels should use. This determines a minimum offset value for axis titles.

\strong{Default value:} \code{30} for y-axis; \code{undefined} for x-axis.}

\item{offset}{(\emph{Axis}) The offset, in pixels, by which to displace the axis from the edge of the enclosing group or data rectangle.

\strong{Default value:} derived from the \href{https://vega.github.io/vega-lite/docs/config.html#facet-scale-config}{axis config}'s \code{offset} (\code{0} by default)}

\item{orient}{(\emph{Axis}) The orientation of the axis. One of \code{"top"}, \code{"bottom"}, \code{"left"} or \code{"right"}. The orientation can be used to further specialize the axis type (e.g., a y-axis oriented towards the right edge of the chart).

\strong{Default value:} \code{"bottom"} for x-axes and \code{"left"} for y-axes.}

\item{position}{(\emph{Axis}) The anchor position of the axis in pixels. For x-axes with top or bottom orientation, this sets the axis group x coordinate. For y-axes with left or right orientation, this sets the axis group y coordinate.

\strong{Default value}: \code{0}}

\item{tickBand}{(\emph{Axis}) For band scales, indicates if ticks and grid lines should be placed at the center of a band (default) or at the band extents to indicate intervals.}

\item{tickColor}{(\emph{Axis})}

\item{tickCount}{(\emph{Axis}) A desired number of ticks, for axes visualizing quantitative scales. The resulting number may be different so that values are "nice" (multiples of 2, 5, 10) and lie within the underlying scale's range.

\strong{Default value}: Determine using a formula \code{ceil(width/40)} for x and \code{ceil(height/40)} for y.}

\item{tickDash}{(\emph{Axis})}

\item{tickDashOffset}{(\emph{Axis})}

\item{tickExtra}{(\emph{Axis}) Boolean flag indicating if an extra axis tick should be added for the initial position of the axis. This flag is useful for styling axes for \code{band} scales such that ticks are placed on band boundaries rather in the middle of a band. Use in conjunction with \code{"bandPosition": 1} and an axis \code{"padding"} value of \code{0}.}

\item{tickMinStep}{(\emph{Axis}) The minimum desired step between axis ticks, in terms of scale domain values. For example, a value of \code{1} indicates that ticks should not be less than 1 unit apart. If \code{tickMinStep} is specified, the \code{tickCount} value will be adjusted, if necessary, to enforce the minimum step value.

\strong{Default value}: \code{undefined}}

\item{tickOffset}{(\emph{Axis}) Position offset in pixels to apply to ticks, labels, and gridlines.}

\item{tickOpacity}{(\emph{Axis})}

\item{tickRound}{(\emph{Axis}) Boolean flag indicating if pixel position values should be rounded to the nearest integer.

\strong{Default value:} \code{true}}

\item{tickSize}{(\emph{Axis}) The size in pixels of axis ticks.

\strong{Default value:} \code{5}}

\item{tickWidth}{(\emph{Axis})}

\item{ticks}{(\emph{Axis}) Boolean value that determines whether the axis should include ticks.

\strong{Default value:} \code{true}}

\item{title}{(\emph{Axis}) A title for the field. If \code{null}, the title will be removed.

\strong{Default value:}  derived from the field's name and transformation function (\code{aggregate}, \code{bin} and \code{timeUnit}). If the field has an aggregate function, the function is displayed as part of the title (e.g., \code{"Sum of Profit"}). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., \code{"Profit (binned)"}, \code{"Transaction Date (year-month)"}). Otherwise, the title is simply the field name.

\strong{Notes}:
\enumerate{
\item You can customize the default field title format by providing the \href{https://vega.github.io/vega-lite/docs/config.html#top-level-config}{fieldTitle} property in the \href{https://vega.github.io/vega-lite/docs/config.html}{config} or \href{https://vega.github.io/vega-lite/docs/compile.html#field-title}{fieldTitle function via the compile function's options}.
\item If both field definition's \code{title} and axis, header, or legend \code{title} are defined, axis/header/legend title will be used.
}}

\item{titleAlign}{(\emph{Axis}) Horizontal text alignment of axis titles.}

\item{titleAnchor}{(\emph{Axis}) Text anchor position for placing axis titles.}

\item{titleAngle}{(\emph{Axis}) Angle in degrees of axis titles.}

\item{titleBaseline}{(\emph{Axis}) Vertical text baseline for axis titles.}

\item{titleColor}{(\emph{Axis}) Color of the title, can be in hex color code or regular color name.}

\item{titleFont}{(\emph{Axis}) Font of the title. (e.g., \code{"Helvetica Neue"}).}

\item{titleFontSize}{(\emph{Axis}) Font size of the title.}

\item{titleFontStyle}{(\emph{Axis}) Font style of the title.}

\item{titleFontWeight}{(\emph{Axis}) Font weight of the title.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{titleLimit}{(\emph{Axis}) Maximum allowed pixel width of axis titles.}

\item{titleLineHeight}{(\emph{Axis}) Line height in pixels for multi-line title text.}

\item{titleOpacity}{(\emph{Axis}) Opacity of the axis title.}

\item{titlePadding}{(\emph{Axis}) The padding, in pixels, between title and axis.}

\item{titleX}{(\emph{Axis}) X-coordinate of the axis title relative to the axis group.}

\item{titleY}{(\emph{Axis}) Y-coordinate of the axis title relative to the axis group.}

\item{translate}{(\emph{Axis}) Translation offset in pixels applied to the axis group mark x and y. If specified, overrides the default behavior of a 0.5 offset to pixel-align stroked lines.}

\item{values}{(\emph{Axis}) Explicitly set the visible axis tick values.}

\item{zindex}{(\emph{Axis}) A non-negative integer indicating the z-index of the axis.
If zindex is 0, axes should be drawn behind all chart elements.
To put them in front, set \code{zindex} to \code{1} or more.

\strong{Default value:} \code{0} (behind the marks).}
}
\value{
A component of a Vega-Lite spec, corresponding to Axis definition.
}
\description{
Create object for Axis
}
