% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BindCheckbox}
\alias{vl$BindCheckbox}
\title{BindCheckbox}
\arguments{
\item{debounce}{(\emph{BindCheckbox})}

\item{element}{(\emph{BindCheckbox})}

\item{input}{(\emph{BindCheckbox})}

\item{name}{(\emph{BindCheckbox})}

\item{type}{(\emph{BindCheckbox})}
}
\value{
A component of a Vega-Lite spec, corresponding to BindCheckbox definition.
}
\description{
Create object for BindCheckbox
}
