% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BindCheckbox}
\alias{vl$BindCheckbox}
\title{BindCheckbox}
\arguments{
\item{debounce}{number}

\item{element}{Element}

\item{input}{string}

\item{name}{string}

\item{type}{string}
}
\value{
A component of a Vega-Lite spec, corresponding to BindCheckbox definition.
}
\description{
Create object for BindCheckbox
}
