% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BindRadioSelect}
\alias{vl$BindRadioSelect}
\title{BindRadioSelect}
\arguments{
\item{debounce}{(\emph{BindRadioSelect})}

\item{element}{(\emph{BindRadioSelect})}

\item{input}{(\emph{BindRadioSelect}) radio, select}

\item{labels}{(\emph{BindRadioSelect})}

\item{name}{(\emph{BindRadioSelect})}

\item{options}{(\emph{BindRadioSelect})}

\item{type}{(\emph{BindRadioSelect})}
}
\value{
A component of a Vega-Lite spec, corresponding to BindRadioSelect definition.
}
\description{
Create object for BindRadioSelect
}
