% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BindRange}
\alias{vl$BindRange}
\title{BindRange}
\arguments{
\item{debounce}{number}

\item{element}{Element}

\item{input}{string}

\item{max}{number}

\item{min}{number}

\item{name}{string}

\item{step}{number}

\item{type}{string}
}
\value{
A component of a Vega-Lite spec, corresponding to BindRange definition.
}
\description{
Create object for BindRange
}
