% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$BindRange}
\alias{vl$BindRange}
\title{BindRange}
\arguments{
\item{debounce}{(\emph{BindRange})}

\item{element}{(\emph{BindRange})}

\item{input}{(\emph{BindRange})}

\item{max}{(\emph{BindRange})}

\item{min}{(\emph{BindRange})}

\item{name}{(\emph{BindRange})}

\item{step}{(\emph{BindRange})}

\item{type}{(\emph{BindRange})}
}
\value{
A component of a Vega-Lite spec, corresponding to BindRange definition.
}
\description{
Create object for BindRange
}
