% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$ErrorBandConfig}
\alias{vl$ErrorBandConfig}
\title{ErrorBandConfig}
\arguments{
\item{band}{(\emph{ErrorBandConfig})}

\item{borders}{(\emph{ErrorBandConfig})}

\item{extent}{(\emph{ErrorBandConfig}) The extent of the band. Available options include: - \code{"ci"}: Extend the band to the confidence interval of the mean. - \code{"stderr"}: The size of band are set to the value of standard error, extending from the mean. - \code{"stdev"}: The size of band are set to the value of standard deviation, extending from the mean. - \code{"iqr"}: Extend the band to the q1 and q3.

\strong{Default value:} \code{"stderr"}.}

\item{interpolate}{(\emph{ErrorBandConfig}) The line interpolation method for the error band. One of the following: - \code{"linear"}: piecewise linear segments, as in a polyline. - \code{"linear-closed"}: close the linear segments to form a polygon. - \code{"step"}: a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes at the midpoint of each pair of adjacent x-values. - \code{"step-before"}: a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes before the x-value. - \code{"step-after"}: a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes after the x-value. - \code{"basis"}: a B-spline, with control point duplication on the ends. - \code{"basis-open"}: an open B-spline; may not intersect the start or end. - \code{"basis-closed"}: a closed B-spline, as in a loop. - \code{"cardinal"}: a Cardinal spline, with control point duplication on the ends. - \code{"cardinal-open"}: an open Cardinal spline; may not intersect the start or end, but will intersect other control points. - \code{"cardinal-closed"}: a closed Cardinal spline, as in a loop. - \code{"bundle"}: equivalent to basis, except the tension parameter is used to straighten the spline. - \code{"monotone"}: cubic interpolation that preserves monotonicity in y.}

\item{tension}{(\emph{ErrorBandConfig}) The tension parameter for the interpolation type of the error band.}
}
\value{
A component of a Vega-Lite spec, corresponding to ErrorBandConfig definition.
}
\description{
Create object for ErrorBandConfig
}
