% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$ErrorBarConfig}
\alias{vl$ErrorBarConfig}
\title{ErrorBarConfig}
\arguments{
\item{extent}{The extent of the rule. Available options include: - \code{"ci"}: Extend the rule
to the confidence interval of the mean. - \code{"stderr"}: The size of rule are set
to the value of standard error, extending from the mean. - \code{"stdev"}: The size
of rule are set to the value of standard deviation, extending from the mean. -
\code{"iqr"}: Extend the rule to the q1 and q3.

\strong{Default value:} \code{"stderr"}.}

\item{rule}{boolean

Or: MarkConfig<ExprOrSignalRef>}

\item{size}{Size of the ticks of an error bar}

\item{thickness}{Thickness of the ticks and the bar of an error bar}

\item{ticks}{boolean

Or: MarkConfig<ExprOrSignalRef>}
}
\value{
A component of a Vega-Lite spec, corresponding to ErrorBarConfig definition.
}
\description{
Create object for ErrorBarConfig
}
