% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$LegendConfig}
\alias{vl$LegendConfig}
\title{LegendConfig}
\arguments{
\item{aria}{}

\item{clipHeight}{}

\item{columnPadding}{}

\item{columns}{}

\item{cornerRadius}{}

\item{description}{}

\item{direction}{The direction of the legend, one of \code{"vertical"} or \code{"horizontal"}.

\strong{Default value:} - For top-/bottom-\code{orient}ed legends,
\code{"horizontal"} - For left-/right-\code{orient}ed legends, \code{"vertical"} - For
top/bottom-left/right-\code{orient}ed legends, \code{"horizontal"} for gradient legends
and \code{"vertical"} for symbol legends.}

\item{disable}{Disable legend by default}

\item{fillColor}{}

\item{gradientDirection}{}

\item{gradientHorizontalMaxLength}{Max legend length for a horizontal gradient when \code{config.legend.gradientLength}
is undefined.

\strong{Default value:} \code{200}}

\item{gradientHorizontalMinLength}{Min legend length for a horizontal gradient when \code{config.legend.gradientLength}
is undefined.

\strong{Default value:} \code{100}}

\item{gradientLabelLimit}{}

\item{gradientLabelOffset}{}

\item{gradientLength}{}

\item{gradientOpacity}{}

\item{gradientStrokeColor}{}

\item{gradientStrokeWidth}{}

\item{gradientThickness}{}

\item{gradientVerticalMaxLength}{Max legend length for a vertical gradient when \code{config.legend.gradientLength} is
undefined.

\strong{Default value:} \code{200}}

\item{gradientVerticalMinLength}{Min legend length for a vertical gradient when \code{config.legend.gradientLength} is
undefined.

\strong{Default value:} \code{100}}

\item{gridAlign}{}

\item{labelAlign}{}

\item{labelBaseline}{}

\item{labelColor}{}

\item{labelFont}{}

\item{labelFontSize}{}

\item{labelFontStyle}{}

\item{labelFontWeight}{}

\item{labelLimit}{}

\item{labelOffset}{}

\item{labelOpacity}{}

\item{labelOverlap}{The strategy to use for resolving overlap of labels in gradient legends. If
\code{false}, no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a
strategy of removing every other label is used. If set to \code{"greedy"}, a linear
scan of the labels is performed, removing any label that overlaps with the last
visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{"greedy"} for \verb{log scales otherwise }true`.}

\item{labelPadding}{}

\item{labelSeparation}{}

\item{layout}{}

\item{legendX}{}

\item{legendY}{}

\item{offset}{}

\item{orient}{The orientation of the legend, which determines how the legend is positioned
within the scene. One of \code{"left"}, \code{"right"}, \code{"top"}, \code{"bottom"}, \code{"top-left"},
\code{"top-right"}, \code{"bottom-left"}, \code{"bottom-right"}, \code{"none"}.

\strong{Default value:} \code{"right"}}

\item{padding}{}

\item{rowPadding}{}

\item{strokeColor}{}

\item{strokeDash}{}

\item{strokeWidth}{}

\item{symbolBaseFillColor}{}

\item{symbolBaseStrokeColor}{}

\item{symbolDash}{}

\item{symbolDashOffset}{}

\item{symbolDirection}{}

\item{symbolFillColor}{}

\item{symbolLimit}{}

\item{symbolOffset}{}

\item{symbolOpacity}{}

\item{symbolSize}{}

\item{symbolStrokeColor}{}

\item{symbolStrokeWidth}{}

\item{symbolType}{}

\item{tickCount}{}

\item{title}{Set to null to disable title for the axis, legend, or header.}

\item{titleAlign}{}

\item{titleAnchor}{}

\item{titleBaseline}{}

\item{titleColor}{}

\item{titleFont}{}

\item{titleFontSize}{}

\item{titleFontStyle}{}

\item{titleFontWeight}{}

\item{titleLimit}{}

\item{titleLineHeight}{}

\item{titleOpacity}{}

\item{titleOrient}{}

\item{titlePadding}{}

\item{unselectedOpacity}{The opacity of unselected legend entries.

\strong{Default value:} 0.35.}

\item{zindex}{}
}
\value{
A component of a Vega-Lite spec, corresponding to LegendConfig definition.
}
\description{
Create object for LegendConfig
}
