% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$RangeConfig}
\alias{vl$RangeConfig}
\title{RangeConfig}
\arguments{
\item{category}{Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for
categorical data.}

\item{diverging}{Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for diverging
quantitative ramps.}

\item{heatmap}{Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for
quantitative heatmaps.}

\item{ordinal}{Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for
rank-ordered data.}

\item{ramp}{Default \href{https://vega.github.io/vega/docs/schemes/}{color scheme} for sequential
quantitative ramps.}

\item{symbol}{Array of \href{https://vega.github.io/vega/docs/marks/symbol/}{symbol} names or paths
for the default shape palette.}

\item{...}{Additional elements for constructing object.}
}
\value{
A component of a Vega-Lite spec, corresponding to RangeConfig definition.
}
\description{
Create object for RangeConfig
}
