% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$TickConfig}
\alias{vl$TickConfig}
\title{TickConfig}
\arguments{
\item{align}{The horizontal alignment of the text or ranged marks (area, bar, image, rect,
rule). One of \code{"left"}, \code{"right"}, \code{"center"}.

\strong{Note:} Expression reference is \emph{not} supported for range marks.}

\item{angle}{}

\item{aria}{}

\item{ariaRole}{}

\item{ariaRoleDescription}{}

\item{aspect}{}

\item{bandSize}{The width of the ticks.

\strong{Default value:} 3/4 of step (width step for horizontal ticks and height step
for vertical ticks).}

\item{baseline}{For text marks, the vertical text baseline. One of \code{"alphabetic"} (default),
\code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, \code{"line-bottom"}, or an
expression reference that provides one of the valid values. The \code{"line-top"}
and \code{"line-bottom"} values operate similarly to \code{"top"} and \code{"bottom"}, but are
calculated relative to the \code{lineHeight} rather than \code{fontSize} alone.

For range marks, the vertical alignment of the marks. One of \code{"top"},
\code{"middle"}, \code{"bottom"}.

\strong{Note:} Expression reference is \emph{not} supported for range marks.}

\item{blend}{}

\item{color}{Default color.

\strong{Default value:} <span style="color: #4682b4;">■</span> \code{"#4682b4"}

\strong{Note:} - This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}. - The
\code{fill} and \code{stroke} properties have higher precedence than \code{color} and will
override \code{color}.}

\item{cornerRadius}{}

\item{cornerRadiusBottomLeft}{}

\item{cornerRadiusBottomRight}{}

\item{cornerRadiusTopLeft}{}

\item{cornerRadiusTopRight}{}

\item{cursor}{}

\item{description}{}

\item{dir}{}

\item{dx}{}

\item{dy}{}

\item{ellipsis}{}

\item{endAngle}{}

\item{fill}{Default fill color. This property has higher precedence than \code{config.color}. Set
to \code{null} to remove fill.

\strong{Default value:} (None)}

\item{fillOpacity}{}

\item{filled}{Whether the mark's color should be used as fill color instead of stroke color.

\strong{Default value:} \code{false} for all \code{point}, \code{line},
and \code{rule} marks as well as \code{geoshape} marks for
\href{https://vega.github.io/vega-lite/docs/data.html#graticule}{\code{graticule}} data
sources; otherwise, \code{true}.

\strong{Note:} This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}.}

\item{font}{}

\item{fontSize}{}

\item{fontStyle}{}

\item{fontWeight}{}

\item{height}{}

\item{href}{}

\item{innerRadius}{The inner radius in pixels of arc marks. \code{innerRadius} is an alias for
\code{radius2}.}

\item{interpolate}{}

\item{invalid}{Defines how Vega-Lite should handle marks for invalid values (\code{null} and
\code{NaN}). - If set to \code{"filter"} (default), all data items with null values will
be skipped (for line, trail, and area marks) or filtered (for other marks). -
If \code{null}, all data items are included. In this case, invalid values will be
interpreted as zeroes.}

\item{limit}{}

\item{lineBreak}{}

\item{lineHeight}{}

\item{opacity}{The overall opacity (value between [0,1]).

\strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick}, \code{circle},
or \code{square} marks or layered \code{bar} charts and \code{1} otherwise.}

\item{order}{For line and trail marks, this \code{order} property can be set to \code{null} or \code{false}
to make the lines use the original order in the data sources.}

\item{orient}{The orientation of a non-stacked bar, tick, area, and line charts. The value
is either horizontal (default) or vertical. - For bar, rule and tick, this
determines whether the size of the bar and tick should be applied to x or y
dimension. - For area, this property determines the orient property of the Vega
output. - For line and trail marks, this property determines the sort order
of the points in the line if \code{config.sortLineBy} is not specified. For stacked
charts, this is always determined by the orientation of the stack; therefore
explicitly specified value will be ignored.}

\item{outerRadius}{The outer radius in pixels of arc marks. \code{outerRadius} is an alias for \code{radius}.}

\item{padAngle}{}

\item{radius}{For arc mark, the primary (outer) radius in pixels.

For text marks, polar coordinate radial offset, in pixels, of the text from the
origin determined by the \code{x} and \code{y} properties.}

\item{radius2}{The secondary (inner) radius in pixels of arc marks.}

\item{shape}{}

\item{size}{Default size for marks. - For \code{point}/\code{circle}/\code{square}, this represents the
pixel area of the marks. Note that this value sets the area of the symbol; the
side lengths will increase with the square root of this value. - For \code{bar}, this
represents the band size of the bar, in pixels. - For \code{text}, this represents
the font size, in pixels.

\strong{Default value:} - \code{30} for point, circle, square marks; width/height's
\code{step} - \code{2} for bar marks with discrete dimensions; - \code{5} for bar marks with
continuous dimensions; - \code{11} for text marks.}

\item{smooth}{}

\item{startAngle}{}

\item{stroke}{Default stroke color. This property has higher precedence than \code{config.color}.
Set to \code{null} to remove stroke.

\strong{Default value:} (None)}

\item{strokeCap}{}

\item{strokeDash}{}

\item{strokeDashOffset}{}

\item{strokeJoin}{}

\item{strokeMiterLimit}{}

\item{strokeOffset}{}

\item{strokeOpacity}{}

\item{strokeWidth}{}

\item{tension}{}

\item{text}{}

\item{theta}{\itemize{
\item For arc marks, the arc length in radians if theta2 is not specified, otherwise
the start arc angle. (A value of 0 indicates up or “north”, increasing values
proceed clockwise.)
\item For text marks, polar coordinate angle in radians.
}}

\item{theta2}{The end angle of arc marks in radians. A value of 0 indicates up or “north”,
increasing values proceed clockwise.}

\item{thickness}{Thickness of the tick mark.

\strong{Default value:} \code{1}}

\item{timeUnitBand}{Default relative band size for a time unit. If set to \code{1}, the bandwidth of the
marks will be equal to the time unit band step. If set to \code{0.5}, bandwidth of
the marks will be half of the time unit band step.}

\item{timeUnitBandPosition}{Default relative band position for a time unit. If set to \code{0}, the marks will
be positioned at the beginning of the time unit band step. If set to \code{0.5}, the
marks will be positioned in the middle of the time unit band step.}

\item{tooltip}{The tooltip text string to show upon mouse hover or an object defining which
fields should the tooltip be derived from.
\itemize{
\item If \code{tooltip} is \code{true} or \code{{"content": "encoding"}}, then all fields from
\code{encoding} will be used. - If \code{tooltip} is \code{{"content": "data"}}, then all
fields that appear in the highlighted data point will be used. - If set to
\code{null} or \code{false}, then no tooltip will be used.
}

See the \href{https://vega.github.io/vega-lite/docs/tooltip.html}{\code{tooltip}}
documentation for a detailed discussion about tooltip in Vega-Lite.

\strong{Default value:} \code{null}}

\item{url}{}

\item{width}{}

\item{x}{X coordinates of the marks, or width of horizontal \code{"bar"} and \code{"area"} without
specified \code{x2} or \code{width}.

The \code{value} of this channel can be a number or a string \code{"width"} for the width
of the plot.}

\item{x2}{X2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and \code{"rule"}.

The \code{value} of this channel can be a number or a string \code{"width"} for the width
of the plot.}

\item{y}{Y coordinates of the marks, or height of vertical \code{"bar"} and \code{"area"} without
specified \code{y2} or \code{height}.

The \code{value} of this channel can be a number or a string \code{"height"} for the
height of the plot.}

\item{y2}{Y2 coordinates for ranged \code{"area"}, \code{"bar"}, \code{"rect"}, and \code{"rule"}.

The \code{value} of this channel can be a number or a string \code{"height"} for the
height of the plot.}
}
\value{
A component of a Vega-Lite spec, corresponding to TickConfig definition.
}
\description{
Create object for TickConfig
}
