% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$TitleConfig}
\alias{vl$TitleConfig}
\title{TitleConfig}
\arguments{
\item{align}{Horizontal text alignment for title text. One of \code{"left"}, \code{"center"}, or
\code{"right"}.}

\item{anchor}{}

\item{angle}{}

\item{aria}{}

\item{baseline}{Vertical text baseline for title and subtitle text. One of \code{"alphabetic"}
(default), \code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, or \code{"line-bottom"}.
The \code{"line-top"} and \code{"line-bottom"} values operate similarly to \code{"top"}
and \code{"bottom"}, but are calculated relative to the \emph{lineHeight} rather than
\emph{fontSize} alone.}

\item{color}{}

\item{dx}{}

\item{dy}{}

\item{font}{}

\item{fontSize}{}

\item{fontStyle}{}

\item{fontWeight}{}

\item{frame}{}

\item{limit}{}

\item{lineHeight}{}

\item{offset}{}

\item{orient}{}

\item{subtitleColor}{}

\item{subtitleFont}{}

\item{subtitleFontSize}{}

\item{subtitleFontStyle}{}

\item{subtitleFontWeight}{}

\item{subtitleLineHeight}{}

\item{subtitlePadding}{}

\item{zindex}{}
}
\value{
A component of a Vega-Lite spec, corresponding to TitleConfig definition.
}
\description{
Create object for TitleConfig
}
