% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl$ViewConfig}
\alias{vl$ViewConfig}
\title{ViewConfig}
\arguments{
\item{clip}{(\emph{ViewConfig}) Whether the view should be clipped.}

\item{continuousHeight}{(\emph{ViewConfig}) The default height when the plot has a continuous y-field.

\strong{Default value:} \code{200}}

\item{continuousWidth}{(\emph{ViewConfig}) The default width when the plot has a continuous x-field.

\strong{Default value:} \code{200}}

\item{cornerRadius}{(\emph{ViewConfig}) The radius in pixels of rounded rectangle corners.

\strong{Default value:} \code{0}}

\item{discreteHeight}{(\emph{ViewConfig}) The default height when the plot has either a discrete y-field or no y-field.

\strong{Default value:} a step size based on \code{config.view.step}.}

\item{discreteWidth}{(\emph{ViewConfig}) The default width when the plot has either a discrete x-field or no x-field.

\strong{Default value:} a step size based on \code{config.view.step}.}

\item{fill}{(\emph{ViewConfig}) The fill color.

\strong{Default value:} \code{undefined}}

\item{fillOpacity}{(\emph{ViewConfig}) The fill opacity (value between [0,1]).

\strong{Default value:} \code{1}}

\item{height}{(\emph{ViewConfig}) Default height}

\item{opacity}{(\emph{ViewConfig}) The overall opacity (value between [0,1]).

\strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick}, \code{circle}, or \code{square} marks or layered \code{bar} charts and \code{1} otherwise.}

\item{step}{(\emph{ViewConfig}) Default step size for x-/y- discrete fields.}

\item{stroke}{(\emph{ViewConfig}) The stroke color.

\strong{Default value:} \code{"#ddd"}}

\item{strokeCap}{(\emph{ViewConfig}) The stroke cap for line ending style. One of \code{"butt"}, \code{"round"}, or \code{"square"}.

\strong{Default value:} \code{"square"}}

\item{strokeDash}{(\emph{ViewConfig}) An array of alternating stroke, space lengths for creating dashed or dotted lines.}

\item{strokeDashOffset}{(\emph{ViewConfig}) The offset (in pixels) into which to begin drawing with the stroke dash array.}

\item{strokeJoin}{(\emph{ViewConfig}) The stroke line join method. One of \code{"miter"}, \code{"round"} or \code{"bevel"}.

\strong{Default value:} \code{"miter"}}

\item{strokeMiterLimit}{(\emph{ViewConfig}) The miter limit at which to bevel a line join.}

\item{strokeOpacity}{(\emph{ViewConfig}) The stroke opacity (value between [0,1]).

\strong{Default value:} \code{1}}

\item{strokeWidth}{(\emph{ViewConfig}) The stroke width, in pixels.}

\item{width}{(\emph{ViewConfig}) Default width}
}
\value{
A component of a Vega-Lite spec, corresponding to ViewConfig definition.
}
\description{
Create object for ViewConfig
}
