% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl.R
\docType{data}
\name{vl}
\alias{vl}
\title{vl}
\format{An object of class \code{vl} (inherits from \code{environment}) of length 81.}
\usage{
vl
}
\description{
Holds functions for making objects based on Vega-Lite spec
}
\keyword{datasets}
