% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_boxplot_config}
\alias{vl_add_boxplot_config}
\title{vl_add_boxplot_config}
\usage{
vl_add_boxplot_config(spec, box = NULL, extent = NULL, median = NULL,
  outliers = NULL, rule = NULL, size = NULL, ticks = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{box}{(\emph{BoxPlotConfig})}

\item{extent}{(\emph{BoxPlotConfig}) The extent of the whiskers. Available options include:
\itemize{
\item \code{"min-max"}: min and max are the lower and upper whiskers respectively.
\item A number representing multiple of the interquartile range.  This number will be multiplied by the IQR to determine whisker boundary, which spans from the smallest data to the largest data within the range [Q1 - k * IQR, Q3 + k * IQR] where \emph{Q1} and \emph{Q3} are the first and third quartiles while \emph{IQR} is the interquartile range (\emph{Q3-Q1}).
}

\strong{Default value:} \code{1.5}.}

\item{median}{(\emph{BoxPlotConfig})}

\item{outliers}{(\emph{BoxPlotConfig})}

\item{rule}{(\emph{BoxPlotConfig})}

\item{size}{(\emph{BoxPlotConfig}) Size of the box and median tick of a box plot}

\item{ticks}{(\emph{BoxPlotConfig})}
}
\value{
A modified spec
}
\description{
Add boxplot config (BoxPlotConfig) to a vega-lite spec.
}
