% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_data}
\alias{vl_add_data}
\title{vl_add_data}
\usage{
vl_add_data(spec, values = NULL, format = NULL, name = NULL,
  url = NULL, sequence = NULL, sphere = NULL, graticule = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{values}{(\emph{InlineData}) The full data set, included inline. This can be an array of objects or primitive values, an object, or a string.
Arrays of primitive values are ingested as objects with a \code{data} property. Strings are parsed according to the specified format type.}

\item{format}{(\emph{UrlData, InlineData, NamedData}) An object that specifies the format for parsing the data.}

\item{name}{(\emph{UrlData, InlineData, NamedData, SequenceGenerator, SphereGenerator, GraticuleGenerator}) Provide a placeholder name and bind data at runtime.}

\item{url}{(\emph{UrlData}) An URL from which to load the data set. Use the \code{format.type} property
to ensure the loaded data is correctly parsed.}

\item{sequence}{(\emph{SequenceGenerator}) Generate a sequence of numbers.}

\item{sphere}{(\emph{SphereGenerator}) Generate sphere GeoJSON data for the full globe.}

\item{graticule}{(\emph{GraticuleGenerator}) Generate graticule GeoJSON data for geographic reference lines.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add data to a vega-lite spec
}
