% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_interval_selection}
\alias{vl_add_interval_selection}
\title{vl_add_interval_selection}
\usage{
vl_add_interval_selection(
  spec,
  selection_name,
  .object = NULL,
  bind = NULL,
  clear = NULL,
  empty = NULL,
  encodings = NULL,
  fields = NULL,
  init = NULL,
  mark = NULL,
  on = NULL,
  resolve = NULL,
  translate = NULL,
  type = NULL,
  zoom = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{selection_name}{Name for selection}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: IntervalSelection}

\item{bind}{(\emph{IntervalSelection}) Establishes a two-way binding between the interval selection and the scales
used within the same view. This allows a user to interactively pan and
zoom the view.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/bind.html}{\code{bind}} documentation.}

\item{clear}{(\emph{IntervalSelection}) Clears the selection, emptying it of all values. Can be a
\href{https://vega.github.io/vega/docs/event-streams/}{Event Stream} or \code{false} to disable.

\strong{Default value:} \code{dblclick}.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/clear.html}{\code{clear}} documentation.}

\item{empty}{(\emph{IntervalSelection}) By default, \code{all} data values are considered to lie within an empty selection.
When set to \code{none}, empty selections contain no data values.}

\item{encodings}{(\emph{IntervalSelection}) An array of encoding channels. The corresponding data field values
must match for a data tuple to fall within the selection.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/project.html}{\code{encodings}} documentation.}

\item{fields}{(\emph{IntervalSelection}) An array of field names whose values must match for a data tuple to
fall within the selection.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/project.html}{\code{fields}} documentation.}

\item{init}{(\emph{IntervalSelection}) Initialize the selection with a mapping between \href{https://vega.github.io/vega-lite/docs/project.html}{projected channels or field names} and arrays of
initial values.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/init.html}{\code{init}} documentation.}

\item{mark}{(\emph{IntervalSelection}) An interval selection also adds a rectangle mark to depict the
extents of the interval. The \code{mark} property can be used to customize the
appearance of the mark.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/selection-mark.html}{\code{mark}} documentation.}

\item{on}{(\emph{IntervalSelection}) A \href{https://vega.github.io/vega/docs/event-streams/}{Vega event stream} (object or selector) that triggers the selection.
For interval selections, the event stream must specify a \href{https://vega.github.io/vega/docs/event-streams/#between-filters}{start and end}.}

\item{resolve}{(\emph{IntervalSelection}) With layered and multi-view displays, a strategy that determines how
selections' data queries are resolved when applied in a filter transform,
conditional encoding rule, or scale domain.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/selection-resolve.html}{\code{resolve}} documentation.}

\item{translate}{(\emph{IntervalSelection}) When truthy, allows a user to interactively move an interval selection
back-and-forth. Can be \code{true}, \code{false} (to disable panning), or a
\href{https://vega.github.io/vega/docs/event-streams/}{Vega event stream definition}
which must include a start and end event to trigger continuous panning.

\strong{Default value:} \code{true}, which corresponds to
\verb{[mousedown, window:mouseup] > window:mousemove!} which corresponds to
clicks and dragging within an interval selection to reposition it.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/translate.html}{\code{translate}} documentation.}

\item{type}{(\emph{IntervalSelection}) Determines the default event processing and data query for the selection. Vega-Lite currently supports three selection types:
\itemize{
\item \code{"single"} -- to select a single discrete data value on \code{click}.
\item \code{"multi"} -- to select multiple discrete data value; the first value is selected on \code{click} and additional values toggled on shift-\code{click}.
\item \code{"interval"} -- to select a continuous range of data values on \code{drag}.
}}

\item{zoom}{(\emph{IntervalSelection}) When truthy, allows a user to interactively resize an interval selection.
Can be \code{true}, \code{false} (to disable zooming), or a \href{https://vega.github.io/vega/docs/event-streams/}{Vega event stream definition}. Currently,
only \code{wheel} events are supported.

\strong{Default value:} \code{true}, which corresponds to \verb{wheel!}.

\strong{See also:} \href{https://vega.github.io/vega-lite/docs/zoom.html}{\code{zoom}} documentation.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add IntervalSelection to a vega-lite spec.
}
