% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_range_config}
\alias{vl_add_range_config}
\title{vl_add_range_config}
\usage{
vl_add_range_config(spec, category = NULL, diverging = NULL,
  heatmap = NULL, ordinal = NULL, ramp = NULL, symbol = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{category}{(\emph{RangeConfig}) Default range for \emph{nominal} (categorical) fields.}

\item{diverging}{(\emph{RangeConfig}) Default range for diverging \emph{quantitative} fields.}

\item{heatmap}{(\emph{RangeConfig}) Default range for \emph{quantitative} heatmaps.}

\item{ordinal}{(\emph{RangeConfig}) Default range for \emph{ordinal} fields.}

\item{ramp}{(\emph{RangeConfig}) Default range for \emph{quantitative} and \emph{temporal} fields.}

\item{symbol}{(\emph{RangeConfig}) Default range palette for the \code{shape} channel.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add range config (RangeConfig) to a vega-lite spec.
}
