% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_add_scale_config}
\alias{vl_add_scale_config}
\title{vl_add_scale_config}
\usage{
vl_add_scale_config(spec, bandPaddingInner = NULL,
  bandPaddingOuter = NULL, barBandPaddingInner = NULL, clamp = NULL,
  continuousPadding = NULL, maxBandSize = NULL, maxFontSize = NULL,
  maxOpacity = NULL, maxSize = NULL, maxStrokeWidth = NULL,
  minBandSize = NULL, minFontSize = NULL, minOpacity = NULL,
  minSize = NULL, minStrokeWidth = NULL, pointPadding = NULL,
  quantileCount = NULL, quantizeCount = NULL,
  rectBandPaddingInner = NULL, round = NULL,
  useUnaggregatedDomain = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{bandPaddingInner}{(\emph{ScaleConfig}) Default inner padding for \code{x} and \code{y} band-ordinal scales.

\strong{Default value:}
\itemize{
\item \code{barBandPaddingInner} for bar marks (\code{0.1} by default)
\item \code{rectBandPaddingInner} for rect and other marks (\code{0} by default)
}}

\item{bandPaddingOuter}{(\emph{ScaleConfig}) Default outer padding for \code{x} and \code{y} band-ordinal scales.

\strong{Default value:} \code{paddingInner/2} (which makes \emph{width/height = number of unique values * step})}

\item{barBandPaddingInner}{(\emph{ScaleConfig}) Default inner padding for \code{x} and \code{y} band-ordinal scales of \code{"bar"} marks.

\strong{Default value:} \code{0.1}}

\item{clamp}{(\emph{ScaleConfig}) If true, values that exceed the data domain are clamped to either the minimum or maximum range value}

\item{continuousPadding}{(\emph{ScaleConfig}) Default padding for continuous scales.

\strong{Default:} \code{5} for continuous x-scale of a vertical bar and continuous y-scale of a horizontal bar.; \code{0} otherwise.}

\item{maxBandSize}{(\emph{ScaleConfig}) The default max value for mapping quantitative fields to bar's size/bandSize.

If undefined (default), we will use the axis's size (width or height) - 1.}

\item{maxFontSize}{(\emph{ScaleConfig}) The default max value for mapping quantitative fields to text's size/fontSize.

\strong{Default value:} \code{40}}

\item{maxOpacity}{(\emph{ScaleConfig}) Default max opacity for mapping a field to opacity.

\strong{Default value:} \code{0.8}}

\item{maxSize}{(\emph{ScaleConfig}) Default max value for point size scale.}

\item{maxStrokeWidth}{(\emph{ScaleConfig}) Default max strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks.

\strong{Default value:} \code{4}}

\item{minBandSize}{(\emph{ScaleConfig}) The default min value for mapping quantitative fields to bar and tick's size/bandSize scale with zero=false.

\strong{Default value:} \code{2}}

\item{minFontSize}{(\emph{ScaleConfig}) The default min value for mapping quantitative fields to tick's size/fontSize scale with zero=false

\strong{Default value:} \code{8}}

\item{minOpacity}{(\emph{ScaleConfig}) Default minimum opacity for mapping a field to opacity.

\strong{Default value:} \code{0.3}}

\item{minSize}{(\emph{ScaleConfig}) Default minimum value for point size scale with zero=false.

\strong{Default value:} \code{9}}

\item{minStrokeWidth}{(\emph{ScaleConfig}) Default minimum strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks with zero=false.

\strong{Default value:} \code{1}}

\item{pointPadding}{(\emph{ScaleConfig}) Default outer padding for \code{x} and \code{y} point-ordinal scales.

\strong{Default value:} \code{0.5} (which makes \emph{width/height = number of unique values * step})}

\item{quantileCount}{(\emph{ScaleConfig}) Default range cardinality for \href{https://vega.github.io/vega-lite/docs/scale.html#quantile}{quantile} scale.

\strong{Default value:} \code{4}}

\item{quantizeCount}{(\emph{ScaleConfig}) Default range cardinality for \href{https://vega.github.io/vega-lite/docs/scale.html#quantize}{quantize} scale.

\strong{Default value:} \code{4}}

\item{rectBandPaddingInner}{(\emph{ScaleConfig}) Default inner padding for \code{x} and \code{y} band-ordinal scales of \code{"rect"} marks.

\strong{Default value:} \code{0}}

\item{round}{(\emph{ScaleConfig}) If true, rounds numeric output values to integers.
This can be helpful for snapping to the pixel grid.
(Only available for \code{x}, \code{y}, and \code{size} scales.)}

\item{useUnaggregatedDomain}{(\emph{ScaleConfig}) Use the source data range before aggregation as scale domain instead of aggregated data for aggregate axis.

This is equivalent to setting \code{domain} to \code{"unaggregate"} for aggregated \emph{quantitative} fields by default.

This property only works with aggregate functions that produce values within the raw data domain (\code{"mean"}, \code{"average"}, \code{"median"}, \code{"q1"}, \code{"q3"}, \code{"min"}, \code{"max"}). For other aggregations that produce values outside of the raw data domain (e.g. \code{"count"}, \code{"sum"}), this property is ignored.

\strong{Default value:} \code{false}}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add scale config (ScaleConfig) to a vega-lite spec.
}
