% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_aggregate}
\alias{vl_aggregate}
\title{vl_aggregate}
\usage{
vl_aggregate(spec = NULL, aggregate = NULL, groupby = NULL)
}
\arguments{
\item{spec}{A vega-lite spec}

\item{aggregate}{(\emph{AggregateTransform}) Array of objects that define fields to aggregate.}

\item{groupby}{(\emph{AggregateTransform}) The data fields to group by. If not specified, a single group containing all data objects will be used.}
}
\value{
A modified spec
}
\description{
Add AggregateTransform to a vega-lite spec.
}
