% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_aggregate}
\alias{vl_aggregate}
\title{vl_aggregate}
\usage{
vl_aggregate(spec, .object = NULL, aggregate = NULL, groupby = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: AggregateTransform}

\item{aggregate}{(\emph{AggregateTransform}) Array of objects that define fields to aggregate.}

\item{groupby}{(\emph{AggregateTransform}) The data fields to group by. If not specified, a single group containing all data objects will be used.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add AggregateTransform to a vega-lite spec.
}
