% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bin}
\alias{vl_bin}
\title{vl_bin}
\usage{
vl_bin(spec, .object = NULL, as = NULL, bin = NULL, field = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BinTransform}

\item{as}{(\emph{BinTransform}) The output fields at which to write the start and end bin values.}

\item{bin}{(\emph{BinTransform}) An object indicating bin properties, or simply \code{true} for using default bin parameters.}

\item{field}{(\emph{BinTransform}) The data field to bin.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add BinTransform to a vega-lite spec.
}
