% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bin}
\alias{vl_bin}
\title{vl_bin}
\usage{
vl_bin(spec, bin = NULL, as = NULL, field = NULL, .object = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{bin}{An object indicating bin properties, or simply \code{true} for using default bin
parameters.}

\item{as}{The output fields at which to write the start and end bin values. This can be
either a string or an array of strings with two elements denoting the name for
the fields for bin start and bin end respectively. If a single string (e.g.,
\code{"val"}) is provided, the end field will be \code{"val_end"}.}

\item{field}{The data field to bin.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than \code{spec}. Objects can be of type: BinTransform}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add BinTransform to a vega-lite spec.
}
