% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_checkbox_input}
\alias{vl_bind_checkbox_input}
\title{vl_bind_checkbox_input}
\usage{
vl_bind_checkbox_input(
  spec,
  parameter_name,
  projection_name = NULL,
  .object = NULL,
  debounce = NULL,
  element = NULL,
  name = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{parameter_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BindCheckbox}

\item{debounce}{If defined, delays event handling until the specified milliseconds have elapsed
since the last event was fired.}

\item{element}{An optional CSS selector string indicating the parent element to which the input
element should be added. By default, all input elements are added within the
parent container of the Vega view.}

\item{name}{By default, the signal name is used to label input elements. This \code{name}
property can be used instead to specify a custom label for the bound signal.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add checkbox binding to a vega-lite spec.
}
