% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_radio_input}
\alias{vl_bind_radio_input}
\title{vl_bind_radio_input}
\usage{
vl_bind_radio_input(
  spec,
  selection_name,
  projection_name = NULL,
  .object = NULL,
  debounce = NULL,
  element = NULL,
  labels = NULL,
  name = NULL,
  options = NULL,
  type = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{selection_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BindRadioSelect}

\item{debounce}{(\emph{BindRadioSelect})}

\item{element}{(\emph{BindRadioSelect})}

\item{labels}{(\emph{BindRadioSelect})}

\item{name}{(\emph{BindRadioSelect})}

\item{options}{(\emph{BindRadioSelect})}

\item{type}{(\emph{BindRadioSelect})}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add radio binding to a vega-lite spec.
}
