% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_radio_input}
\alias{vl_bind_radio_input}
\title{vl_bind_radio_input}
\usage{
vl_bind_radio_input(spec, selection_name, projection_name = NULL,
  debounce = NULL, element = NULL, name = NULL, options = NULL,
  type = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{selection_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{debounce}{(type = number)}

\item{element}{(type = Element)}

\item{name}{(type = string)}

\item{options}{(type = array)}

\item{type}{(type = string)}
}
\value{
A modified spec
}
\description{
Add binding to a radio input to a vega-lite spec.
}
