% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_range_input}
\alias{vl_bind_range_input}
\title{vl_bind_range_input}
\usage{
vl_bind_range_input(spec, selection_name, projection_name = NULL,
  debounce = NULL, element = NULL, max = NULL, min = NULL,
  name = NULL, step = NULL, type = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{selection_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{debounce}{(type = number)}

\item{element}{(type = Element)}

\item{max}{(type = number)}

\item{min}{(type = number)}

\item{name}{(type = string)}

\item{step}{(type = number)}

\item{type}{(type = string)}
}
\value{
A modified spec
}
\description{
Add binding to a range input to a vega-lite spec.
}
