% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_range_input}
\alias{vl_bind_range_input}
\title{vl_bind_range_input}
\usage{
vl_bind_range_input(
  spec,
  selection_name,
  projection_name = NULL,
  .object = NULL,
  debounce = NULL,
  element = NULL,
  max = NULL,
  min = NULL,
  name = NULL,
  step = NULL,
  type = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{selection_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BindRange}

\item{debounce}{number}

\item{element}{Element}

\item{max}{number}

\item{min}{number}

\item{name}{string}

\item{step}{number}

\item{type}{string}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add range binding to a vega-lite spec.
}
