% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_range_input}
\alias{vl_bind_range_input}
\title{vl_bind_range_input}
\usage{
vl_bind_range_input(
  spec,
  parameter_name,
  projection_name = NULL,
  debounce = NULL,
  element = NULL,
  max = NULL,
  min = NULL,
  name = NULL,
  step = NULL,
  .object = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{parameter_name}{Name of parameter to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{debounce}{If defined, delays event handling until the specified milliseconds have elapsed
since the last event was fired.}

\item{element}{An optional CSS selector string indicating the parent element to which the input
element should be added. By default, all input elements are added within the
parent container of the Vega view.}

\item{max}{Sets the maximum slider value. Defaults to the larger of the signal value and
\code{100}.}

\item{min}{Sets the minimum slider value. Defaults to the smaller of the signal value and
\code{0}.}

\item{name}{By default, the signal name is used to label input elements. This \code{name}
property can be used instead to specify a custom label for the bound signal.}

\item{step}{Sets the minimum slider increment. If undefined, the step size will be
automatically determined based on the \code{min} and \code{max} values.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than \code{spec}. Objects can be of type: BindRange}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add range binding to a vega-lite spec.
}
