% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_select_input}
\alias{vl_bind_select_input}
\title{vl_bind_select_input}
\usage{
vl_bind_select_input(
  spec,
  selection_name,
  projection_name = NULL,
  .object = NULL,
  debounce = NULL,
  element = NULL,
  labels = NULL,
  name = NULL,
  options = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{selection_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BindRadioSelect}

\item{debounce}{If defined, delays event handling until the specified milliseconds have elapsed
since the last event was fired.}

\item{element}{An optional CSS selector string indicating the parent element to which the input
element should be added. By default, all input elements are added within the
parent container of the Vega view.}

\item{labels}{An array of label strings to represent the \code{options} values. If unspecified, the
\code{options} value will be coerced to a string and used as the label.}

\item{name}{By default, the signal name is used to label input elements. This \code{name}
property can be used instead to specify a custom label for the bound signal.}

\item{options}{An array of options to select from.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add select binding to a vega-lite spec.
}
