% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_bind_select_input}
\alias{vl_bind_select_input}
\title{vl_bind_select_input}
\usage{
vl_bind_select_input(spec, selection_name, projection_name = NULL,
  debounce = NULL, element = NULL, labels = NULL, name = NULL,
  options = NULL, type = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{selection_name}{Name of selection to add binding to}

\item{projection_name}{Name of projection (field or encoding) within selection}

\item{debounce}{(\emph{BindRadioSelect})}

\item{element}{(\emph{BindRadioSelect})}

\item{labels}{(\emph{BindRadioSelect})}

\item{name}{(\emph{BindRadioSelect})}

\item{options}{(\emph{BindRadioSelect})}

\item{type}{(\emph{BindRadioSelect})}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add select binding to a vega-lite spec.
}
