% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_calculate}
\alias{vl_calculate}
\title{vl_calculate}
\usage{
vl_calculate(spec, as = NULL, calculate = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{as}{(\emph{CalculateTransform}) The field for storing the computed formula value.}

\item{calculate}{(\emph{CalculateTransform}) A \href{https://vega.github.io/vega-lite/docs/types.html#expression}{expression} string. Use the variable \code{datum} to refer to the current data object.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add CalculateTransform to a vega-lite spec.
}
