% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_calculate}
\alias{vl_calculate}
\title{vl_calculate}
\usage{
vl_calculate(spec, calculate = NULL, as = NULL, .object = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{calculate}{A \href{https://vega.github.io/vega-lite/docs/types.html#expression}{expression}
string. Use the variable \code{datum} to refer to the current data object.}

\item{as}{The field for storing the computed formula value.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than \code{spec}. Objects can be of type: CalculateTransform}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add CalculateTransform to a vega-lite spec.
}
