% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_legend}
\alias{vl_config_legend}
\title{vl_config_legend}
\usage{
vl_config_legend(
  spec,
  .object = NULL,
  clipHeight = NULL,
  columnPadding = NULL,
  columns = NULL,
  cornerRadius = NULL,
  fillColor = NULL,
  gradientDirection = NULL,
  gradientHorizontalMaxLength = NULL,
  gradientHorizontalMinLength = NULL,
  gradientLabelLimit = NULL,
  gradientLabelOffset = NULL,
  gradientLength = NULL,
  gradientOpacity = NULL,
  gradientStrokeColor = NULL,
  gradientStrokeWidth = NULL,
  gradientThickness = NULL,
  gradientVerticalMaxLength = NULL,
  gradientVerticalMinLength = NULL,
  gridAlign = NULL,
  labelAlign = NULL,
  labelBaseline = NULL,
  labelColor = NULL,
  labelFont = NULL,
  labelFontSize = NULL,
  labelFontStyle = NULL,
  labelFontWeight = NULL,
  labelLimit = NULL,
  labelOffset = NULL,
  labelOpacity = NULL,
  labelOverlap = NULL,
  labelPadding = NULL,
  labelSeparation = NULL,
  layout = NULL,
  legendX = NULL,
  legendY = NULL,
  offset = NULL,
  orient = NULL,
  padding = NULL,
  rowPadding = NULL,
  strokeColor = NULL,
  strokeDash = NULL,
  strokeWidth = NULL,
  symbolBaseFillColor = NULL,
  symbolBaseStrokeColor = NULL,
  symbolDash = NULL,
  symbolDashOffset = NULL,
  symbolDirection = NULL,
  symbolFillColor = NULL,
  symbolLimit = NULL,
  symbolOffset = NULL,
  symbolOpacity = NULL,
  symbolSize = NULL,
  symbolStrokeColor = NULL,
  symbolStrokeWidth = NULL,
  symbolType = NULL,
  tickCount = NULL,
  title = NULL,
  titleAlign = NULL,
  titleAnchor = NULL,
  titleBaseline = NULL,
  titleColor = NULL,
  titleFont = NULL,
  titleFontSize = NULL,
  titleFontStyle = NULL,
  titleFontWeight = NULL,
  titleLimit = NULL,
  titleLineHeight = NULL,
  titleOpacity = NULL,
  titleOrient = NULL,
  titlePadding = NULL,
  unselectedOpacity = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: LegendConfig}

\item{clipHeight}{(\emph{LegendConfig}) The height in pixels to clip symbol legend entries and limit their size.}

\item{columnPadding}{(\emph{LegendConfig}) The horizontal padding in pixels between symbol legend entries.

\strong{Default value:} \code{10}.}

\item{columns}{(\emph{LegendConfig}) The number of columns in which to arrange symbol legend entries. A value of \code{0} or lower indicates a single row with one column per entry.}

\item{cornerRadius}{(\emph{LegendConfig}) Corner radius for the full legend.}

\item{fillColor}{(\emph{LegendConfig}) Background fill color for the full legend.}

\item{gradientDirection}{(\emph{LegendConfig}) The default direction (\code{"horizontal"} or \code{"vertical"}) for gradient legends.

\strong{Default value:} \code{"vertical"}.}

\item{gradientHorizontalMaxLength}{(\emph{LegendConfig}) Max legend length for a horizontal gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{200}}

\item{gradientHorizontalMinLength}{(\emph{LegendConfig}) Min legend length for a horizontal gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{100}}

\item{gradientLabelLimit}{(\emph{LegendConfig}) The maximum allowed length in pixels of color ramp gradient labels.}

\item{gradientLabelOffset}{(\emph{LegendConfig}) Vertical offset in pixels for color ramp gradient labels.

\strong{Default value:} \code{2}.}

\item{gradientLength}{(\emph{LegendConfig}) The length in pixels of the primary axis of a color gradient. This value corresponds to the height of a vertical gradient or the width of a horizontal gradient.

\strong{Default value:} \code{200}.}

\item{gradientOpacity}{(\emph{LegendConfig}) Opacity of the color gradient.}

\item{gradientStrokeColor}{(\emph{LegendConfig}) The color of the gradient stroke, can be in hex color code or regular color name.

\strong{Default value:} \code{"lightGray"}.}

\item{gradientStrokeWidth}{(\emph{LegendConfig}) The width of the gradient stroke, in pixels.

\strong{Default value:} \code{0}.}

\item{gradientThickness}{(\emph{LegendConfig}) The thickness in pixels of the color gradient. This value corresponds to the width of a vertical gradient or the height of a horizontal gradient.

\strong{Default value:} \code{16}.}

\item{gradientVerticalMaxLength}{(\emph{LegendConfig}) Max legend length for a vertical gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{200}}

\item{gradientVerticalMinLength}{(\emph{LegendConfig}) Min legend length for a vertical gradient when \code{config.legend.gradientLength} is undefined.

\strong{Default value:} \code{100}}

\item{gridAlign}{(\emph{LegendConfig}) The alignment to apply to symbol legends rows and columns. The supported string values are \code{"all"}, \code{"each"} (the default), and \code{none}. For more information, see the \href{https://vega.github.io/vega/docs/layout}{grid layout documentation}.

\strong{Default value:} \code{"each"}.}

\item{labelAlign}{(\emph{LegendConfig}) The alignment of the legend label, can be left, center, or right.}

\item{labelBaseline}{(\emph{LegendConfig}) The position of the baseline of legend label, can be \code{"top"}, \code{"middle"}, \code{"bottom"}, or \code{"alphabetic"}.

\strong{Default value:} \code{"middle"}.}

\item{labelColor}{(\emph{LegendConfig}) The color of the legend label, can be in hex color code or regular color name.}

\item{labelFont}{(\emph{LegendConfig}) The font of the legend label.}

\item{labelFontSize}{(\emph{LegendConfig}) The font size of legend label.

\strong{Default value:} \code{10}.}

\item{labelFontStyle}{(\emph{LegendConfig}) The font style of legend label.}

\item{labelFontWeight}{(\emph{LegendConfig}) The font weight of legend label.}

\item{labelLimit}{(\emph{LegendConfig}) Maximum allowed pixel width of legend tick labels.

\strong{Default value:} \code{160}.}

\item{labelOffset}{(\emph{LegendConfig}) The offset of the legend label.}

\item{labelOpacity}{(\emph{LegendConfig}) Opacity of labels.}

\item{labelOverlap}{(\emph{LegendConfig}) The strategy to use for resolving overlap of labels in gradient legends. If \code{false}, no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a strategy of removing every other label is used. If set to \code{"greedy"}, a linear scan of the labels is performed, removing any label that overlaps with the last visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{"greedy"} for \verb{log scales otherwise }true`.}

\item{labelPadding}{(\emph{LegendConfig}) Padding in pixels between the legend and legend labels.}

\item{labelSeparation}{(\emph{LegendConfig}) The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default \code{0}). This property is ignored if \emph{labelOverlap} resolution is not enabled.}

\item{layout}{(\emph{LegendConfig}) Legend orient group layout parameters.}

\item{legendX}{(\emph{LegendConfig}) Custom x-position for legend with orient "none".}

\item{legendY}{(\emph{LegendConfig}) Custom y-position for legend with orient "none".}

\item{offset}{(\emph{LegendConfig}) The offset in pixels by which to displace the legend from the data rectangle and axes.

\strong{Default value:} \code{18}.}

\item{orient}{(\emph{LegendConfig}) The orientation of the legend, which determines how the legend is positioned within the scene. One of "left", "right", "top-left", "top-right", "bottom-left", "bottom-right", "none".

\strong{Default value:} \code{"right"}}

\item{padding}{(\emph{LegendConfig}) The padding between the border and content of the legend group.

\strong{Default value:} \code{0}.}

\item{rowPadding}{(\emph{LegendConfig}) The vertical padding in pixels between symbol legend entries.

\strong{Default value:} \code{2}.}

\item{strokeColor}{(\emph{LegendConfig}) Border stroke color for the full legend.}

\item{strokeDash}{(\emph{LegendConfig}) Border stroke dash pattern for the full legend.}

\item{strokeWidth}{(\emph{LegendConfig}) Border stroke width for the full legend.}

\item{symbolBaseFillColor}{(\emph{LegendConfig}) Default fill color for legend symbols. Only applied if there is no \code{"fill"} scale color encoding for the legend.

\strong{Default value:} \code{"transparent"}.}

\item{symbolBaseStrokeColor}{(\emph{LegendConfig}) Default stroke color for legend symbols. Only applied if there is no \code{"fill"} scale color encoding for the legend.

\strong{Default value:} \code{"gray"}.}

\item{symbolDash}{(\emph{LegendConfig}) An array of alternating [stroke, space] lengths for dashed symbol strokes.}

\item{symbolDashOffset}{(\emph{LegendConfig}) The pixel offset at which to start drawing with the symbol stroke dash array.}

\item{symbolDirection}{(\emph{LegendConfig}) The default direction (\code{"horizontal"} or \code{"vertical"}) for symbol legends.

\strong{Default value:} \code{"vertical"}.}

\item{symbolFillColor}{(\emph{LegendConfig}) The color of the legend symbol,}

\item{symbolLimit}{(\emph{LegendConfig}) The maximum number of allowed entries for a symbol legend. Additional entries will be dropped.}

\item{symbolOffset}{(\emph{LegendConfig}) Horizontal pixel offset for legend symbols.

\strong{Default value:} \code{0}.}

\item{symbolOpacity}{(\emph{LegendConfig}) Opacity of the legend symbols.}

\item{symbolSize}{(\emph{LegendConfig}) The size of the legend symbol, in pixels.

\strong{Default value:} \code{100}.}

\item{symbolStrokeColor}{(\emph{LegendConfig}) Stroke color for legend symbols.}

\item{symbolStrokeWidth}{(\emph{LegendConfig}) The width of the symbol's stroke.

\strong{Default value:} \code{1.5}.}

\item{symbolType}{(\emph{LegendConfig}) The symbol shape. One of the plotting shapes \code{circle} (default), \code{square}, \code{cross}, \code{diamond}, \code{triangle-up}, \code{triangle-down}, \code{triangle-right}, or \code{triangle-left}, the line symbol \code{stroke}, or one of the centered directional shapes \code{arrow}, \code{wedge}, or \code{triangle}. Alternatively, a custom \href{https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths}{SVG path string} can be provided. For correct sizing, custom shape paths should be defined within a square bounding box with coordinates ranging from -1 to 1 along both the x and y dimensions.

\strong{Default value:} \code{"circle"}.}

\item{tickCount}{(\emph{LegendConfig}) The desired number of tick values for quantitative legends.}

\item{title}{(\emph{LegendConfig}) Set to null to disable title for the axis, legend, or header.}

\item{titleAlign}{(\emph{LegendConfig}) Horizontal text alignment for legend titles.

\strong{Default value:} \code{"left"}.}

\item{titleAnchor}{(\emph{LegendConfig}) Text anchor position for placing legend titles.}

\item{titleBaseline}{(\emph{LegendConfig}) Vertical text baseline for legend titles.

\strong{Default value:} \code{"top"}.}

\item{titleColor}{(\emph{LegendConfig}) The color of the legend title, can be in hex color code or regular color name.}

\item{titleFont}{(\emph{LegendConfig}) The font of the legend title.}

\item{titleFontSize}{(\emph{LegendConfig}) The font size of the legend title.}

\item{titleFontStyle}{(\emph{LegendConfig}) The font style of the legend title.}

\item{titleFontWeight}{(\emph{LegendConfig}) The font weight of the legend title.
This can be either a string (e.g \code{"bold"}, \code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} = \code{400} and \code{"bold"} = \code{700}).}

\item{titleLimit}{(\emph{LegendConfig}) Maximum allowed pixel width of legend titles.

\strong{Default value:} \code{180}.}

\item{titleLineHeight}{(\emph{LegendConfig}) Line height in pixels for multi-line title text.}

\item{titleOpacity}{(\emph{LegendConfig}) Opacity of the legend title.}

\item{titleOrient}{(\emph{LegendConfig}) Orientation of the legend title.}

\item{titlePadding}{(\emph{LegendConfig}) The padding, in pixels, between title and legend.

\strong{Default value:} \code{5}.}

\item{unselectedOpacity}{(\emph{LegendConfig}) The opacity of unselected legend entries.

\strong{Default value:} 0.35.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add legend config (LegendConfig) to a vega-lite spec.
}
