% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_legend}
\alias{vl_config_legend}
\title{vl_config_legend}
\usage{
vl_config_legend(
  spec,
  .object = NULL,
  aria = NULL,
  clipHeight = NULL,
  columnPadding = NULL,
  columns = NULL,
  cornerRadius = NULL,
  description = NULL,
  direction = NULL,
  disable = NULL,
  fillColor = NULL,
  gradientDirection = NULL,
  gradientHorizontalMaxLength = NULL,
  gradientHorizontalMinLength = NULL,
  gradientLabelLimit = NULL,
  gradientLabelOffset = NULL,
  gradientLength = NULL,
  gradientOpacity = NULL,
  gradientStrokeColor = NULL,
  gradientStrokeWidth = NULL,
  gradientThickness = NULL,
  gradientVerticalMaxLength = NULL,
  gradientVerticalMinLength = NULL,
  gridAlign = NULL,
  labelAlign = NULL,
  labelBaseline = NULL,
  labelColor = NULL,
  labelFont = NULL,
  labelFontSize = NULL,
  labelFontStyle = NULL,
  labelFontWeight = NULL,
  labelLimit = NULL,
  labelOffset = NULL,
  labelOpacity = NULL,
  labelOverlap = NULL,
  labelPadding = NULL,
  labelSeparation = NULL,
  layout = NULL,
  legendX = NULL,
  legendY = NULL,
  offset = NULL,
  orient = NULL,
  padding = NULL,
  rowPadding = NULL,
  strokeColor = NULL,
  strokeDash = NULL,
  strokeWidth = NULL,
  symbolBaseFillColor = NULL,
  symbolBaseStrokeColor = NULL,
  symbolDash = NULL,
  symbolDashOffset = NULL,
  symbolDirection = NULL,
  symbolFillColor = NULL,
  symbolLimit = NULL,
  symbolOffset = NULL,
  symbolOpacity = NULL,
  symbolSize = NULL,
  symbolStrokeColor = NULL,
  symbolStrokeWidth = NULL,
  symbolType = NULL,
  tickCount = NULL,
  title = NULL,
  titleAlign = NULL,
  titleAnchor = NULL,
  titleBaseline = NULL,
  titleColor = NULL,
  titleFont = NULL,
  titleFontSize = NULL,
  titleFontStyle = NULL,
  titleFontWeight = NULL,
  titleLimit = NULL,
  titleLineHeight = NULL,
  titleOpacity = NULL,
  titleOrient = NULL,
  titlePadding = NULL,
  unselectedOpacity = NULL,
  zindex = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: LegendConfig}

\item{aria}{}

\item{clipHeight}{}

\item{columnPadding}{}

\item{columns}{}

\item{cornerRadius}{}

\item{description}{}

\item{direction}{The direction of the legend, one of \code{"vertical"} or \code{"horizontal"}.

\strong{Default value:} - For top-/bottom-\code{orient}ed legends,
\code{"horizontal"} - For left-/right-\code{orient}ed legends, \code{"vertical"} - For
top/bottom-left/right-\code{orient}ed legends, \code{"horizontal"} for gradient legends
and \code{"vertical"} for symbol legends.}

\item{disable}{Disable legend by default}

\item{fillColor}{}

\item{gradientDirection}{}

\item{gradientHorizontalMaxLength}{Max legend length for a horizontal gradient when \code{config.legend.gradientLength}
is undefined.

\strong{Default value:} \code{200}}

\item{gradientHorizontalMinLength}{Min legend length for a horizontal gradient when \code{config.legend.gradientLength}
is undefined.

\strong{Default value:} \code{100}}

\item{gradientLabelLimit}{}

\item{gradientLabelOffset}{}

\item{gradientLength}{}

\item{gradientOpacity}{}

\item{gradientStrokeColor}{}

\item{gradientStrokeWidth}{}

\item{gradientThickness}{}

\item{gradientVerticalMaxLength}{Max legend length for a vertical gradient when \code{config.legend.gradientLength} is
undefined.

\strong{Default value:} \code{200}}

\item{gradientVerticalMinLength}{Min legend length for a vertical gradient when \code{config.legend.gradientLength} is
undefined.

\strong{Default value:} \code{100}}

\item{gridAlign}{}

\item{labelAlign}{}

\item{labelBaseline}{}

\item{labelColor}{}

\item{labelFont}{}

\item{labelFontSize}{}

\item{labelFontStyle}{}

\item{labelFontWeight}{}

\item{labelLimit}{}

\item{labelOffset}{}

\item{labelOpacity}{}

\item{labelOverlap}{The strategy to use for resolving overlap of labels in gradient legends. If
\code{false}, no overlap reduction is attempted. If set to \code{true} or \code{"parity"}, a
strategy of removing every other label is used. If set to \code{"greedy"}, a linear
scan of the labels is performed, removing any label that overlaps with the last
visible label (this often works better for log-scaled axes).

\strong{Default value:} \code{"greedy"} for \verb{log scales otherwise }true`.}

\item{labelPadding}{}

\item{labelSeparation}{}

\item{layout}{}

\item{legendX}{}

\item{legendY}{}

\item{offset}{}

\item{orient}{The orientation of the legend, which determines how the legend is positioned
within the scene. One of \code{"left"}, \code{"right"}, \code{"top"}, \code{"bottom"}, \code{"top-left"},
\code{"top-right"}, \code{"bottom-left"}, \code{"bottom-right"}, \code{"none"}.

\strong{Default value:} \code{"right"}}

\item{padding}{}

\item{rowPadding}{}

\item{strokeColor}{}

\item{strokeDash}{}

\item{strokeWidth}{}

\item{symbolBaseFillColor}{}

\item{symbolBaseStrokeColor}{}

\item{symbolDash}{}

\item{symbolDashOffset}{}

\item{symbolDirection}{}

\item{symbolFillColor}{}

\item{symbolLimit}{}

\item{symbolOffset}{}

\item{symbolOpacity}{}

\item{symbolSize}{}

\item{symbolStrokeColor}{}

\item{symbolStrokeWidth}{}

\item{symbolType}{}

\item{tickCount}{}

\item{title}{Set to null to disable title for the axis, legend, or header.}

\item{titleAlign}{}

\item{titleAnchor}{}

\item{titleBaseline}{}

\item{titleColor}{}

\item{titleFont}{}

\item{titleFontSize}{}

\item{titleFontStyle}{}

\item{titleFontWeight}{}

\item{titleLimit}{}

\item{titleLineHeight}{}

\item{titleOpacity}{}

\item{titleOrient}{}

\item{titlePadding}{}

\item{unselectedOpacity}{The opacity of unselected legend entries.

\strong{Default value:} 0.35.}

\item{zindex}{}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add legend config (LegendConfig) to a vega-lite spec.
}
