% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_selection}
\alias{vl_config_selection}
\title{vl_config_selection}
\usage{
vl_config_selection(spec, interval = NULL, point = NULL, .object = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{interval}{The default definition for an
\href{https://vega.github.io/vega-lite/docs/parameter.html#select}{\code{interval}}
selection. All properties and transformations for an interval selection
definition (except \code{type}) may be specified here.

For instance, setting \code{interval} to \code{{"translate": false}} disables the ability
to move interval selections by default.}

\item{point}{The default definition for a
\href{https://vega.github.io/vega-lite/docs/parameter.html#select}{\code{point}}
selection. All properties and transformations for a point selection definition
(except \code{type}) may be specified here.

For instance, setting \code{point} to \code{{"on": "dblclick"}} populates point selections
on double-click by default.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than \code{spec}. Objects can be of type: SelectionConfig}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add selection config (SelectionConfig) to a vega-lite spec.
}
