% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_selection}
\alias{vl_config_selection}
\title{vl_config_selection}
\usage{
vl_config_selection(
  spec,
  .object = NULL,
  interval = NULL,
  multi = NULL,
  single = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: SelectionConfig}

\item{interval}{The default definition for an
\href{https://vega.github.io/vega-lite/docs/selection.html#type}{\code{interval}}
selection. All properties and transformations for an interval selection
definition (except \code{type}) may be specified here.

For instance, setting \code{interval} to \code{{"translate": false}} disables the ability
to move interval selections by default.}

\item{multi}{The default definition for a
\href{https://vega.github.io/vega-lite/docs/selection.html#type}{\code{multi}} selection.
All properties and transformations for a multi selection definition (except
\code{type}) may be specified here.

For instance, setting \code{multi} to \code{{"toggle": "event.altKey"}} adds additional
values to multi selections when clicking with the alt-key pressed by default.}

\item{single}{The default definition for a
\href{https://vega.github.io/vega-lite/docs/selection.html#type}{\code{single}} selection.
All properties and transformations for a single selection definition (except
\code{type}) may be specified here.

For instance, setting \code{single} to \code{{"on": "dblclick"}} populates single
selections on double-click by default.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add selection config (SelectionConfig) to a vega-lite spec.
}
