% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_title}
\alias{vl_config_title}
\title{vl_config_title}
\usage{
vl_config_title(
  spec,
  align = NULL,
  anchor = NULL,
  angle = NULL,
  aria = NULL,
  baseline = NULL,
  color = NULL,
  dx = NULL,
  dy = NULL,
  font = NULL,
  fontSize = NULL,
  fontStyle = NULL,
  fontWeight = NULL,
  frame = NULL,
  limit = NULL,
  lineHeight = NULL,
  offset = NULL,
  orient = NULL,
  subtitleColor = NULL,
  subtitleFont = NULL,
  subtitleFontSize = NULL,
  subtitleFontStyle = NULL,
  subtitleFontWeight = NULL,
  subtitleLineHeight = NULL,
  subtitlePadding = NULL,
  zindex = NULL,
  .object = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{align}{Horizontal text alignment for title text. One of \code{"left"}, \code{"center"}, or
\code{"right"}.}

\item{anchor}{The anchor position for placing the title and subtitle text. One of \code{"start"},
\code{"middle"}, or \code{"end"}. For example, with an orientation of top these anchor
positions map to a left-, center-, or right-aligned title.

Or: ExprRef}

\item{angle}{Angle in degrees of title and subtitle text.

Or: ExprRef}

\item{aria}{A boolean flag indicating if \href{https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA}{ARIA attributes}
should be included (SVG output only). If \code{false}, the "aria-hidden" attribute
will be set on the output SVG group, removing the title from the ARIA
accessibility tree.

\strong{Default value:} \code{true}

Or: ExprRef}

\item{baseline}{Vertical text baseline for title and subtitle text. One of \code{"alphabetic"}
(default), \code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, or \code{"line-bottom"}.
The \code{"line-top"} and \code{"line-bottom"} values operate similarly to \code{"top"}
and \code{"bottom"}, but are calculated relative to the \emph{lineHeight} rather than
\emph{fontSize} alone.}

\item{color}{Text color for title text.

Or: ExprRef}

\item{dx}{Delta offset for title and subtitle text x-coordinate.

Or: ExprRef}

\item{dy}{Delta offset for title and subtitle text y-coordinate.

Or: ExprRef}

\item{font}{Font name for title text.

Or: ExprRef}

\item{fontSize}{Font size in pixels for title text.

Or: ExprRef}

\item{fontStyle}{Font style for title text.

Or: ExprRef}

\item{fontWeight}{Font weight for title text. This can be either a string (e.g \code{"bold"},
\code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} =
\code{400} and \code{"bold"} = \code{700}).

Or: ExprRef}

\item{frame}{The reference frame for the anchor position, one of \code{"bounds"} (to anchor
relative to the full bounding box) or \code{"group"} (to anchor relative to the group
width or height).

Or: ExprRef}

\item{limit}{The maximum allowed length in pixels of title and subtitle text.

Or: ExprRef}

\item{lineHeight}{Line height in pixels for multi-line title text or title text with \code{"line-top"}
or \code{"line-bottom"} baseline.

Or: ExprRef}

\item{offset}{The orthogonal offset in pixels by which to displace the title group from its
position along the edge of the chart.

Or: ExprRef}

\item{orient}{Default title orientation (\code{"top"}, \code{"bottom"}, \code{"left"}, or \code{"right"})

Or: ExprRef}

\item{subtitleColor}{Text color for subtitle text.

Or: ExprRef}

\item{subtitleFont}{Font name for subtitle text.

Or: ExprRef}

\item{subtitleFontSize}{Font size in pixels for subtitle text.

Or: ExprRef}

\item{subtitleFontStyle}{Font style for subtitle text.

Or: ExprRef}

\item{subtitleFontWeight}{Font weight for subtitle text. This can be either a string (e.g \code{"bold"},
\code{"normal"}) or a number (\code{100}, \code{200}, \code{300}, ..., \code{900} where \code{"normal"} =
\code{400} and \code{"bold"} = \code{700}).

Or: ExprRef}

\item{subtitleLineHeight}{Line height in pixels for multi-line subtitle text.

Or: ExprRef}

\item{subtitlePadding}{The padding in pixels between title and subtitle text.

Or: ExprRef}

\item{zindex}{The integer z-index indicating the layering of the title group relative to other
axis, mark, and legend groups.

\strong{Default value:} \code{0}.

Or: ExprRef}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BaseTitleNoValueRefs}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add title config (TitleConfig) to a vega-lite spec.
}
