% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_title}
\alias{vl_config_title}
\title{vl_config_title}
\usage{
vl_config_title(
  spec,
  .object = NULL,
  align = NULL,
  anchor = NULL,
  angle = NULL,
  aria = NULL,
  baseline = NULL,
  color = NULL,
  dx = NULL,
  dy = NULL,
  font = NULL,
  fontSize = NULL,
  fontStyle = NULL,
  fontWeight = NULL,
  frame = NULL,
  limit = NULL,
  lineHeight = NULL,
  offset = NULL,
  orient = NULL,
  subtitleColor = NULL,
  subtitleFont = NULL,
  subtitleFontSize = NULL,
  subtitleFontStyle = NULL,
  subtitleFontWeight = NULL,
  subtitleLineHeight = NULL,
  subtitlePadding = NULL,
  zindex = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BaseTitleNoValueRefs}

\item{align}{Horizontal text alignment for title text. One of \code{"left"}, \code{"center"}, or
\code{"right"}.}

\item{anchor}{}

\item{angle}{}

\item{aria}{}

\item{baseline}{Vertical text baseline for title and subtitle text. One of \code{"alphabetic"}
(default), \code{"top"}, \code{"middle"}, \code{"bottom"}, \code{"line-top"}, or \code{"line-bottom"}.
The \code{"line-top"} and \code{"line-bottom"} values operate similarly to \code{"top"}
and \code{"bottom"}, but are calculated relative to the \emph{lineHeight} rather than
\emph{fontSize} alone.}

\item{color}{}

\item{dx}{}

\item{dy}{}

\item{font}{}

\item{fontSize}{}

\item{fontStyle}{}

\item{fontWeight}{}

\item{frame}{}

\item{limit}{}

\item{lineHeight}{}

\item{offset}{}

\item{orient}{}

\item{subtitleColor}{}

\item{subtitleFont}{}

\item{subtitleFontSize}{}

\item{subtitleFontStyle}{}

\item{subtitleFontWeight}{}

\item{subtitleLineHeight}{}

\item{subtitlePadding}{}

\item{zindex}{}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add title config (TitleConfig) to a vega-lite spec.
}
