% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_config_view}
\alias{vl_config_view}
\title{vl_config_view}
\usage{
vl_config_view(
  spec,
  .object = NULL,
  clip = NULL,
  continuousHeight = NULL,
  continuousWidth = NULL,
  cornerRadius = NULL,
  cursor = NULL,
  discreteHeight = NULL,
  discreteWidth = NULL,
  fill = NULL,
  fillOpacity = NULL,
  height = NULL,
  opacity = NULL,
  step = NULL,
  stroke = NULL,
  strokeCap = NULL,
  strokeDash = NULL,
  strokeDashOffset = NULL,
  strokeJoin = NULL,
  strokeMiterLimit = NULL,
  strokeOpacity = NULL,
  strokeWidth = NULL,
  width = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: ViewConfig}

\item{clip}{Whether the view should be clipped.}

\item{continuousHeight}{The default height when the plot has a continuous y-field for x or latitude, or
has arc marks.

\strong{Default value:} \code{200}}

\item{continuousWidth}{The default width when the plot has a continuous field for x or longitude, or
has arc marks.

\strong{Default value:} \code{200}}

\item{cornerRadius}{}

\item{cursor}{The mouse cursor used over the view. Any valid \href{https://developer.mozilla.org/en-US/docs/Web/CSS/cursor#Values}{CSS cursor type} can be
used.}

\item{discreteHeight}{The default height when the plot has non arc marks and either a discrete y-field
or no y-field. The height can be either a number indicating a fixed height or an
object in the form of \code{{step: number}} defining the height per discrete step.

\strong{Default value:} a step size based on \code{config.view.step}.}

\item{discreteWidth}{The default width when the plot has non-arc marks and either a discrete x-field
or no x-field. The width can be either a number indicating a fixed width or an
object in the form of \code{{step: number}} defining the width per discrete step.

\strong{Default value:} a step size based on \code{config.view.step}.}

\item{fill}{The fill color.

\strong{Default value:} \code{undefined}}

\item{fillOpacity}{}

\item{height}{Default height

\strong{Deprecated:} Since Vega-Lite 4.0. Please use continuousHeight and
discreteHeight instead.}

\item{opacity}{The overall opacity (value between [0,1]).

\strong{Default value:} \code{0.7} for non-aggregate plots with \code{point}, \code{tick}, \code{circle},
or \code{square} marks or layered \code{bar} charts and \code{1} otherwise.}

\item{step}{Default step size for x-/y- discrete fields.}

\item{stroke}{The stroke color.

\strong{Default value:} \code{"#ddd"}}

\item{strokeCap}{}

\item{strokeDash}{}

\item{strokeDashOffset}{}

\item{strokeJoin}{}

\item{strokeMiterLimit}{}

\item{strokeOpacity}{}

\item{strokeWidth}{}

\item{width}{Default width

\strong{Deprecated:} Since Vega-Lite 4.0. Please use continuousWidth and
discreteWidth instead.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add view config (ViewConfig) to a vega-lite spec.
}
