% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_joinaggregate}
\alias{vl_joinaggregate}
\title{vl_joinaggregate}
\usage{
vl_joinaggregate(spec, joinaggregate = NULL, groupby = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{joinaggregate}{The definition of the fields in the join aggregate, and what calculations to
use.}

\item{groupby}{The data fields for partitioning the data objects into separate groups. If
unspecified, all data points will be in a single group.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add JoinAggregateTransform to a vega-lite spec.
}
