% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_loess}
\alias{vl_loess}
\title{vl_loess}
\usage{
vl_loess(
  spec,
  .object = NULL,
  loess = NULL,
  on = NULL,
  as = NULL,
  bandwidth = NULL,
  groupby = NULL
)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: LoessTransform}

\item{loess}{The data field of the dependent variable to smooth.}

\item{on}{The data field of the independent variable to use a predictor.}

\item{as}{The output field names for the smoothed points generated by the loess transform.

\strong{Default value:} The field names of the input x and y values.}

\item{bandwidth}{A bandwidth parameter in the range \verb{[0, 1]} that determines the amount of
smoothing.

\strong{Default value:} \code{0.3}}

\item{groupby}{The data fields to group by. If not specified, a single group containing all
data objects will be used.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add LoessTransform to a vega-lite spec.
}
