% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_BindRadioSelect}
\alias{vl_make_BindRadioSelect}
\title{vl_make_BindRadioSelect}
\usage{
vl_make_BindRadioSelect(debounce = NULL, element = NULL,
  input = NULL, labels = NULL, name = NULL, options = NULL,
  type = NULL)
}
\arguments{
\item{debounce}{(\emph{BindRadioSelect})}

\item{element}{(\emph{BindRadioSelect})}

\item{input}{(\emph{BindRadioSelect}) radio, select}

\item{labels}{(\emph{BindRadioSelect})}

\item{name}{(\emph{BindRadioSelect})}

\item{options}{(\emph{BindRadioSelect})}

\item{type}{(\emph{BindRadioSelect})}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BindRadioSelect}
}
\value{
A component of a Vega-Lite spec, corresponding to a BindRadioSelect.
}
\description{
Create spec for {obj}
}
