% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_JoinAggregateTransform}
\alias{vl_make_JoinAggregateTransform}
\title{vl_make_JoinAggregateTransform}
\usage{
vl_make_JoinAggregateTransform(groupby = NULL, joinaggregate = NULL)
}
\arguments{
\item{groupby}{(\emph{JoinAggregateTransform}) The data fields for partitioning the data objects into separate groups. If unspecified, all data points will be in a single group.}

\item{joinaggregate}{(\emph{JoinAggregateTransform}) The definition of the fields in the join aggregate, and what calculations to use.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: JoinAggregateTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a JoinAggregateTransform.
}
\description{
Create spec for {obj}
}
