% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_LookupTransform}
\alias{vl_make_LookupTransform}
\title{vl_make_LookupTransform}
\usage{
vl_make_LookupTransform(as = NULL, default = NULL, from = NULL,
  lookup = NULL)
}
\arguments{
\item{as}{(\emph{LookupTransform}) The output fields on which to store the looked up data values.

For data lookups, this property may be left blank if \code{from.fields}
has been specified (those field names will be used); if \code{from.fields}
has not been specified, \code{as} must be a string.

For selection lookups, this property is optional: if unspecified,
looked up values will be stored under a property named for the selection;
and if specified, it must correspond to \code{from.fields}.}

\item{default}{(\emph{LookupTransform}) The default value to use if lookup fails.

\strong{Default value:} \code{null}}

\item{from}{(\emph{LookupTransform}) Data source or selection for secondary data reference.}

\item{lookup}{(\emph{LookupTransform}) Key in primary data source.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: LookupTransform}
}
\value{
A component of a Vega-Lite spec, corresponding to a LookupTransform.
}
\description{
Create spec for {obj}
}
