% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_SampleTransform}
\alias{vl_make_SampleTransform}
\title{vl_make_SampleTransform}
\usage{
vl_make_SampleTransform(sample = NULL)
}
\arguments{
\item{sample}{(\emph{SampleTransform}) The maximum number of data objects to include in the sample.

\strong{Default value:} \code{1000}}
}
\value{
A modified spec
}
\description{
Create spec for SampleTransform.
}
