% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_SelectionConfig}
\alias{vl_make_SelectionConfig}
\title{vl_make_SelectionConfig}
\usage{
vl_make_SelectionConfig(interval = NULL, multi = NULL, single = NULL)
}
\arguments{
\item{interval}{(\emph{SelectionConfig}) The default definition for an \href{https://vega.github.io/vega-lite/docs/selection.html#type}{interval} selection. All properties and transformations
for an interval selection definition (except \code{type}) may be specified here.

For instance, setting \code{interval} to \code{{"translate": false}} disables the ability to move
interval selections by default.}

\item{multi}{(\emph{SelectionConfig}) The default definition for a \href{https://vega.github.io/vega-lite/docs/selection.html#type}{multi} selection. All properties and transformations
for a multi selection definition (except \code{type}) may be specified here.

For instance, setting \code{multi} to \code{{"toggle": "event.altKey"}} adds additional values to
multi selections when clicking with the alt-key pressed by default.}

\item{single}{(\emph{SelectionConfig}) The default definition for a \href{https://vega.github.io/vega-lite/docs/selection.html#type}{single} selection. All properties and transformations
for a single selection definition (except \code{type}) may be specified here.

For instance, setting \code{single} to \code{{"on": "dblclick"}} populates single selections on double-click by default.}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: SelectionConfig}
}
\value{
A component of a Vega-Lite spec, corresponding to a SelectionConfig.
}
\description{
Create spec for {obj}
}
