% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_make_TimeUnitTransform}
\alias{vl_make_TimeUnitTransform}
\title{vl_make_TimeUnitTransform}
\usage{
vl_make_TimeUnitTransform(as = NULL, field = NULL, timeUnit = NULL)
}
\arguments{
\item{as}{(\emph{TimeUnitTransform}) The output field to write the timeUnit value.}

\item{field}{(\emph{TimeUnitTransform}) The data field to apply time unit.}

\item{timeUnit}{(\emph{TimeUnitTransform}) The timeUnit.}
}
\value{
A modified spec
}
\description{
Create spec for TimeUnitTransform.
}
