% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_mark_boxplot}
\alias{vl_mark_boxplot}
\title{vl_mark_boxplot}
\usage{
vl_mark_boxplot(spec, box = NULL, clip = NULL, color = NULL,
  extent = NULL, median = NULL, opacity = NULL, orient = NULL,
  outliers = NULL, rule = NULL, size = NULL, ticks = NULL,
  type = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{box}{(\emph{BoxPlotDef})}

\item{clip}{(\emph{BoxPlotDef}) Whether a composite mark be clipped to the enclosing group’s width and height.}

\item{color}{(\emph{BoxPlotDef}) Default color.  Note that \code{fill} and \code{stroke} have higher precedence than \code{color} and will override \code{color}.

\strong{Default value:} <span style="color: #4682b4;">■</span> \code{"#4682b4"}

\strong{Note:} This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}.}

\item{extent}{(\emph{BoxPlotDef}) The extent of the whiskers. Available options include:
\itemize{
\item \code{"min-max"}: min and max are the lower and upper whiskers respectively.
\item A number representing multiple of the interquartile range.  This number will be multiplied by the IQR to determine whisker boundary, which spans from the smallest data to the largest data within the range \emph{[Q1 - k * IQR, Q3 + k * IQR]} where \emph{Q1} and \emph{Q3} are the first and third quartiles while \emph{IQR} is the interquartile range (\emph{Q3-Q1}).
}

\strong{Default value:} \code{1.5}.}

\item{median}{(\emph{BoxPlotDef})}

\item{opacity}{(\emph{BoxPlotDef}) The opacity (value between [0,1]) of the mark.}

\item{orient}{(\emph{BoxPlotDef}) Orientation of the box plot.  This is normally automatically determined based on types of fields on x and y channels. However, an explicit \code{orient} be specified when the orientation is ambiguous.

\strong{Default value:} \code{"vertical"}.}

\item{outliers}{(\emph{BoxPlotDef})}

\item{rule}{(\emph{BoxPlotDef})}

\item{size}{(\emph{BoxPlotDef}) Size of the box and median tick of a box plot}

\item{ticks}{(\emph{BoxPlotDef})}
}
\value{
A modified Vega-Lite Spec
}
\description{
vl_mark_boxplot
}
