% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_mark_errorband}
\alias{vl_mark_errorband}
\title{vl_mark_errorband}
\usage{
vl_mark_errorband(spec, band = NULL, borders = NULL, clip = NULL,
  color = NULL, extent = NULL, interpolate = NULL, opacity = NULL,
  orient = NULL, tension = NULL, type = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{band}{(\emph{ErrorBandDef})}

\item{borders}{(\emph{ErrorBandDef})}

\item{clip}{(\emph{ErrorBandDef}) Whether a composite mark be clipped to the enclosing group’s width and height.}

\item{color}{(\emph{ErrorBandDef}) Default color.  Note that \code{fill} and \code{stroke} have higher precedence than \code{color} and will override \code{color}.

\strong{Default value:} <span style="color: #4682b4;">■</span> \code{"#4682b4"}

\strong{Note:} This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}.}

\item{extent}{(\emph{ErrorBandDef}) The extent of the band. Available options include:
\itemize{
\item \code{"ci"}: Extend the band to the confidence interval of the mean.
\item \code{"stderr"}: The size of band are set to the value of standard error, extending from the mean.
\item \code{"stdev"}: The size of band are set to the value of standard deviation, extending from the mean.
\item \code{"iqr"}: Extend the band to the q1 and q3.
}

\strong{Default value:} \code{"stderr"}.}

\item{interpolate}{(\emph{ErrorBandDef}) The line interpolation method for the error band. One of the following:
\itemize{
\item \code{"linear"}: piecewise linear segments, as in a polyline.
\item \code{"linear-closed"}: close the linear segments to form a polygon.
\item \code{"step"}: a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes at the midpoint of each pair of adjacent x-values.
\item \code{"step-before"}: a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes before the x-value.
\item \code{"step-after"}: a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes after the x-value.
\item \code{"basis"}: a B-spline, with control point duplication on the ends.
\item \code{"basis-open"}: an open B-spline; may not intersect the start or end.
\item \code{"basis-closed"}: a closed B-spline, as in a loop.
\item \code{"cardinal"}: a Cardinal spline, with control point duplication on the ends.
\item \code{"cardinal-open"}: an open Cardinal spline; may not intersect the start or end, but will intersect other control points.
\item \code{"cardinal-closed"}: a closed Cardinal spline, as in a loop.
\item \code{"bundle"}: equivalent to basis, except the tension parameter is used to straighten the spline.
\item \code{"monotone"}: cubic interpolation that preserves monotonicity in y.
}}

\item{opacity}{(\emph{ErrorBandDef}) The opacity (value between [0,1]) of the mark.}

\item{orient}{(\emph{ErrorBandDef}) Orientation of the error band. This is normally automatically determined, but can be specified when the orientation is ambiguous and cannot be automatically determined.}

\item{tension}{(\emph{ErrorBandDef}) The tension parameter for the interpolation type of the error band.}
}
\value{
A modified Vega-Lite Spec
}
\description{
vl_mark_errorband
}
