% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_mark_errorbar}
\alias{vl_mark_errorbar}
\title{vl_mark_errorbar}
\usage{
vl_mark_errorbar(spec, .object = NULL, clip = NULL, color = NULL,
  extent = NULL, opacity = NULL, orient = NULL, rule = NULL,
  ticks = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: ErrorBarDef}

\item{clip}{(\emph{ErrorBarDef}) Whether a composite mark be clipped to the enclosing group’s width and height.}

\item{color}{(\emph{ErrorBarDef}) Default color.

\strong{Default value:} <span style="color: #4682b4;">■</span> \code{"#4682b4"}

\strong{Note:}
\itemize{
\item This property cannot be used in a \href{https://vega.github.io/vega-lite/docs/mark.html#style-config}{style config}.
\item The \code{fill} and \code{stroke} properties have higher precedence than \code{color} and will override \code{color}.
}}

\item{extent}{(\emph{ErrorBarDef}) The extent of the rule. Available options include:
\itemize{
\item \code{"ci"}: Extend the rule to the confidence interval of the mean.
\item \code{"stderr"}: The size of rule are set to the value of standard error, extending from the mean.
\item \code{"stdev"}: The size of rule are set to the value of standard deviation, extending from the mean.
\item \code{"iqr"}: Extend the rule to the q1 and q3.
}

\strong{Default value:} \code{"stderr"}.}

\item{opacity}{(\emph{ErrorBarDef}) The opacity (value between [0,1]) of the mark.}

\item{orient}{(\emph{ErrorBarDef}) Orientation of the error bar. This is normally automatically determined, but can be specified when the orientation is ambiguous and cannot be automatically determined.}

\item{rule}{(\emph{ErrorBarDef})}

\item{ticks}{(\emph{ErrorBarDef})}
}
\value{
A modified Vega-Lite Spec
}
\description{
vl_mark_errorbar
}
