% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_repeat_row}
\alias{vl_repeat_row}
\title{vl_repeat_row}
\usage{
vl_repeat_row(spec, ...)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{...}{fields to use for repeat (strings)}
}
\value{
A modified spec
}
\description{
Add repeating by row to a vega-lite spec.
}
\seealso{
\code{\link[=vl_repeat_col]{vl_repeat_col()}}, \code{\link[=vl_repeat_wrap]{vl_repeat_wrap()}}
}
