% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_repeat_wrap}
\alias{vl_repeat_wrap}
\title{vl_repeat_wrap}
\usage{
vl_repeat_wrap(spec, ..., columns = 2)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{...}{fields to use for repeat (strings)}

\item{columns}{number of columns to use}
}
\value{
A modified spec
}
\description{
Add repeating by wrap to a vega-lite spec.
}
\seealso{
\code{\link[=vl_repeat_col]{vl_repeat_col()}}, \code{\link[=vl_repeat_row]{vl_repeat_row()}}
}
