% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_sample}
\alias{vl_sample}
\title{vl_sample}
\usage{
vl_sample(spec, sample = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{sample}{(\emph{SampleTransform}) The maximum number of data objects to include in the sample.

\strong{Default value:} \code{1000}}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add SampleTransform to a vega-lite spec.
}
