% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_stack}
\alias{vl_stack}
\title{vl_stack}
\usage{
vl_stack(spec, as = NULL, groupby = NULL, offset = NULL,
  sort = NULL, stack = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{as}{Output field names. This can be either a string or an array of strings with
two elements denoting the name for the fields for stack start and stack end
respectively.
If a single string(eg."val") is provided, the end field will be "val_end". (type = Varies)}

\item{groupby}{The data fields to group by. (type = array)}

\item{offset}{Mode for stacking marks.
\strong{Default value:} \code{"zero"} (type = string)}

\item{sort}{Field that determines the order of leaves in the stacked charts. (type = array)}

\item{stack}{The field which is stacked. (type = string)}
}
\value{
A modified spec
}
\description{
Add StackTransform to a vega-lite spec.
}
