% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_timeunit}
\alias{vl_timeunit}
\title{vl_timeunit}
\usage{
vl_timeunit(spec, .object = NULL, as = NULL, field = NULL,
  timeUnit = NULL)
}
\arguments{
\item{spec}{An input vega-lite spec}

\item{.object}{Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: TimeUnitTransform}

\item{as}{(\emph{TimeUnitTransform}) The output field to write the timeUnit value.}

\item{field}{(\emph{TimeUnitTransform}) The data field to apply time unit.}

\item{timeUnit}{(\emph{TimeUnitTransform}) The timeUnit.}
}
\value{
A modified Vega-Lite Spec
}
\description{
Add TimeUnitTransform to a vega-lite spec.
}
