% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_autogen_api.R
\name{vl_timeunit}
\alias{vl_timeunit}
\title{vl_timeunit}
\usage{
vl_timeunit(spec, as = NULL, field = NULL, timeUnit = NULL)
}
\arguments{
\item{spec}{A vega-lite spec.}

\item{as}{The output field to write the timeUnit value. (type = string)}

\item{field}{The data field to apply time unit. (type = string)}

\item{timeUnit}{The timeUnit. (type = TimeUnit)}
}
\value{
A modified spec
}
\description{
Add TimeUnitTransform to a vega-lite spec.
}
