% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/zzz_autogen_api.R
\name{vlbuildr-deprecated}
\alias{vlbuildr-deprecated}
\alias{vl_make_Color}
\alias{vl_make_Detail}
\alias{vl_make_Fill}
\alias{vl_make_FillOpacity}
\alias{vl_make_Href}
\alias{vl_make_Key}
\alias{vl_make_Latitude}
\alias{vl_make_Latitude2}
\alias{vl_make_Longitude}
\alias{vl_make_Longitude2}
\alias{vl_make_Opacity}
\alias{vl_make_Order}
\alias{vl_make_Shape}
\alias{vl_make_Size}
\alias{vl_make_Stroke}
\alias{vl_make_StrokeOpacity}
\alias{vl_make_StrokeWidth}
\alias{vl_make_Text}
\alias{vl_make_Tooltip}
\alias{vl_make_Url}
\alias{vl_make_X}
\alias{vl_make_X2}
\alias{vl_make_XError}
\alias{vl_make_XError2}
\alias{vl_make_Y}
\alias{vl_make_Y2}
\alias{vl_make_YError}
\alias{vl_make_YError2}
\alias{vl_make_AggregateTransform}
\alias{vl_make_BinTransform}
\alias{vl_make_CalculateTransform}
\alias{vl_make_DensityTransform}
\alias{vl_make_FilterTransform}
\alias{vl_make_FlattenTransform}
\alias{vl_make_FoldTransform}
\alias{vl_make_ImputeTransform}
\alias{vl_make_JoinAggregateTransform}
\alias{vl_make_LoessTransform}
\alias{vl_make_LookupTransform}
\alias{vl_make_QuantileTransform}
\alias{vl_make_RegressionTransform}
\alias{vl_make_TimeUnitTransform}
\alias{vl_make_SampleTransform}
\alias{vl_make_StackTransform}
\alias{vl_make_WindowTransform}
\alias{vl_make_PivotTransform}
\alias{vl_make_SingleSelection}
\alias{vl_make_MultiSelection}
\alias{vl_make_IntervalSelection}
\alias{vl_add_area_config}
\alias{vl_add_axis_config}
\alias{vl_add_axisBand_config}
\alias{vl_add_axisBottom_config}
\alias{vl_add_axisLeft_config}
\alias{vl_add_axisRight_config}
\alias{vl_add_axisTop_config}
\alias{vl_add_axisX_config}
\alias{vl_add_axisY_config}
\alias{vl_add_bar_config}
\alias{vl_add_boxplot_config}
\alias{vl_add_circle_config}
\alias{vl_add_concat_config}
\alias{vl_add_errorband_config}
\alias{vl_add_errorbar_config}
\alias{vl_add_facet_config}
\alias{vl_add_geoshape_config}
\alias{vl_add_header_config}
\alias{vl_add_headerColumn_config}
\alias{vl_add_headerFacet_config}
\alias{vl_add_headerRow_config}
\alias{vl_add_image_config}
\alias{vl_add_legend_config}
\alias{vl_add_line_config}
\alias{vl_add_mark_config}
\alias{vl_add_point_config}
\alias{vl_add_projection_config}
\alias{vl_add_range_config}
\alias{vl_add_rect_config}
\alias{vl_add_repeat_config}
\alias{vl_add_rule_config}
\alias{vl_add_scale_config}
\alias{vl_add_selection_config}
\alias{vl_add_square_config}
\alias{vl_add_text_config}
\alias{vl_add_tick_config}
\alias{vl_add_title_config}
\alias{vl_add_trail_config}
\alias{vl_add_view_config}
\alias{vl_add_config}
\alias{vl_make_AreaConfig}
\alias{vl_make_AxisConfig}
\alias{vl_make_BaseMarkConfig}
\alias{vl_make_BoxPlotConfig}
\alias{vl_make_BrushConfig}
\alias{vl_make_CompositionConfig}
\alias{vl_make_Config}
\alias{vl_make_ErrorBandConfig}
\alias{vl_make_ErrorBarConfig}
\alias{vl_make_HeaderConfig}
\alias{vl_make_IntervalSelectionConfig}
\alias{vl_make_LegendConfig}
\alias{vl_make_LineConfig}
\alias{vl_make_MarkConfig}
\alias{vl_make_MultiSelectionConfig}
\alias{vl_make_ProjectionConfig}
\alias{vl_make_RangeConfig}
\alias{vl_make_RectConfig}
\alias{vl_make_ScaleConfig}
\alias{vl_make_SelectionConfig}
\alias{vl_make_SingleSelectionConfig}
\alias{vl_make_TickConfig}
\alias{vl_make_TitleConfig}
\alias{vl_make_ViewConfig}
\alias{vl_make_BinParams}
\alias{vl_make_Axis}
\alias{vl_make_Scale}
\alias{vl_make_Legend}
\alias{vl_make_BindCheckbox}
\alias{vl_make_BindRange}
\alias{vl_make_BindRadioSelect}
\title{Deprecated functions from vlbuildr}
\usage{
vl_make_Color(...)

vl_make_Detail(...)

vl_make_Fill(...)

vl_make_FillOpacity(...)

vl_make_Href(...)

vl_make_Key(...)

vl_make_Latitude(...)

vl_make_Latitude2(...)

vl_make_Longitude(...)

vl_make_Longitude2(...)

vl_make_Opacity(...)

vl_make_Order(...)

vl_make_Shape(...)

vl_make_Size(...)

vl_make_Stroke(...)

vl_make_StrokeOpacity(...)

vl_make_StrokeWidth(...)

vl_make_Text(...)

vl_make_Tooltip(...)

vl_make_Url(...)

vl_make_X(...)

vl_make_X2(...)

vl_make_XError(...)

vl_make_XError2(...)

vl_make_Y(...)

vl_make_Y2(...)

vl_make_YError(...)

vl_make_YError2(...)

vl_make_AggregateTransform(...)

vl_make_BinTransform(...)

vl_make_CalculateTransform(...)

vl_make_DensityTransform(...)

vl_make_FilterTransform(...)

vl_make_FlattenTransform(...)

vl_make_FoldTransform(...)

vl_make_ImputeTransform(...)

vl_make_JoinAggregateTransform(...)

vl_make_LoessTransform(...)

vl_make_LookupTransform(...)

vl_make_QuantileTransform(...)

vl_make_RegressionTransform(...)

vl_make_TimeUnitTransform(...)

vl_make_SampleTransform(...)

vl_make_StackTransform(...)

vl_make_WindowTransform(...)

vl_make_PivotTransform(...)

vl_make_SingleSelection(...)

vl_make_MultiSelection(...)

vl_make_IntervalSelection(...)

vl_add_area_config(...)

vl_add_axis_config(...)

vl_add_axisBand_config(...)

vl_add_axisBottom_config(...)

vl_add_axisLeft_config(...)

vl_add_axisRight_config(...)

vl_add_axisTop_config(...)

vl_add_axisX_config(...)

vl_add_axisY_config(...)

vl_add_bar_config(...)

vl_add_boxplot_config(...)

vl_add_circle_config(...)

vl_add_concat_config(...)

vl_add_errorband_config(...)

vl_add_errorbar_config(...)

vl_add_facet_config(...)

vl_add_geoshape_config(...)

vl_add_header_config(...)

vl_add_headerColumn_config(...)

vl_add_headerFacet_config(...)

vl_add_headerRow_config(...)

vl_add_image_config(...)

vl_add_legend_config(...)

vl_add_line_config(...)

vl_add_mark_config(...)

vl_add_point_config(...)

vl_add_projection_config(...)

vl_add_range_config(...)

vl_add_rect_config(...)

vl_add_repeat_config(...)

vl_add_rule_config(...)

vl_add_scale_config(...)

vl_add_selection_config(...)

vl_add_square_config(...)

vl_add_text_config(...)

vl_add_tick_config(...)

vl_add_title_config(...)

vl_add_trail_config(...)

vl_add_view_config(...)

vl_add_config(...)

vl_make_AreaConfig(...)

vl_make_AxisConfig(...)

vl_make_BaseMarkConfig(...)

vl_make_BoxPlotConfig(...)

vl_make_BrushConfig(...)

vl_make_CompositionConfig(...)

vl_make_Config(...)

vl_make_ErrorBandConfig(...)

vl_make_ErrorBarConfig(...)

vl_make_HeaderConfig(...)

vl_make_IntervalSelectionConfig(...)

vl_make_LegendConfig(...)

vl_make_LineConfig(...)

vl_make_MarkConfig(...)

vl_make_MultiSelectionConfig(...)

vl_make_ProjectionConfig(...)

vl_make_RangeConfig(...)

vl_make_RectConfig(...)

vl_make_ScaleConfig(...)

vl_make_SelectionConfig(...)

vl_make_SingleSelectionConfig(...)

vl_make_TickConfig(...)

vl_make_TitleConfig(...)

vl_make_ViewConfig(...)

vl_make_BinParams(...)

vl_make_Axis(...)

vl_make_Scale(...)

vl_make_Legend(...)

vl_make_BindCheckbox(...)

vl_make_BindRange(...)

vl_make_BindRadioSelect(...)
}
\arguments{
\item{...}{argument to pass to new function}
}
\description{
These functions have been deprecated in favor of a different
naming convention.
}
