% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlbuildr-package.R
\docType{package}
\name{vlbuildr-package}
\alias{vlbuildr}
\alias{vlbuildr-package}
\title{vlbuildr: Build vega-lite specs in R}
\description{
Build vegalite specs in R. Inspired by vegalite package as well as 
  vega-lite-api and altair. Dependent on vegawidget for actually rendering specs!
}
\author{
\strong{Maintainer}: Aicia Schep \email{aschep@gmail.com}

}
\keyword{internal}
